Computes d1 from Black-Scholes-Merton model.

d1 formula implementation, from Black-Scholes-Merton model.

Usage

bsmd1(spot, strike, time, rate, yield, sigma)

Arguments

spot

current stock price.

strike

the strike price.

time

time to option expiration in years.

rate

the risk-free interest rate.

yield

the dividends that are expected to be paid.

sigma

volatility of the stock price.

Value

Vector with d1.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

Examples

bsmd1(c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
#> [1] 1.4036052 1.2015781 0.9953852 0.7848511
bsmd1(50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
#>    [1] 270.7212813 135.3610157  90.2410938  67.6812578  54.1454563  45.1216719
#>    [7]  38.6761830  33.8421289  30.0823646  27.0746031  24.6137528  22.5630859
#>   [13]  20.8279447  19.3407165  18.0518188  16.9240645  15.9290165  15.0445573
#>   [19]  14.2532253  13.5410516  12.8967277  12.3110014  11.7762296  11.2860430
#>   [25]  10.8350913  10.4188474  10.0334549   9.6756083   9.3424580   9.0315344
#>   [31]   8.7406865   8.4680322   8.2119176   7.9708833   7.7436366   7.5290286
#>   [37]   7.3260346   7.1337377   6.9513149   6.7780258   6.6132020   6.4562388
#>   [43]   6.3065879   6.1637507   6.0272729   5.8967398   5.7717719   5.6520215
#>   [49]   5.5371690   5.4269206   5.3210055   5.2191737   5.1211940   5.0268524
#>   [55]   4.9359506   4.8483041   4.7637418   4.6821040   4.6032421   4.5270172
#>   [61]   4.4532997   4.3819682   4.3129092   4.2460161   4.1811889   4.1183338
#>   [67]   4.0573624   3.9981916   3.9407432   3.8849433   3.8307223   3.7780143
#>   [73]   3.7267573   3.6768923   3.6283638   3.5811188   3.5351075   3.4902825
#>   [79]   3.4465985   3.4040129   3.3624850   3.3219760   3.2824492   3.2438694
#>   [85]   3.2062033   3.1694190   3.1334860   3.0983754   3.0640593   3.0305115
#>   [91]   2.9977064   2.9656199   2.9342288   2.9035110   2.8734451   2.8440107
#>   [97]   2.8151885   2.7869595   2.7593059   2.7322103   2.7056563   2.6796278
#>  [103]   2.6541095   2.6290868   2.6045455   2.5804720   2.5568531   2.5336762
#>  [109]   2.5109292   2.4886003   2.4666782   2.4451521   2.4240113   2.4032459
#>  [115]   2.3828459   2.3628020   2.3431050   2.3237460   2.3047166   2.2860086
#>  [121]   2.2676139   2.2495248   2.2317340   2.2142341   2.1970183   2.1800796
#>  [127]   2.1634117   2.1470081   2.1308626   2.1149695   2.0993228   2.0839169
#>  [133]   2.0687465   2.0538062   2.0390910   2.0245958   2.0103159   1.9962466
#>  [139]   1.9823833   1.9687217   1.9552573   1.9419861   1.9289041   1.9160072
#>  [145]   1.9032916   1.8907536   1.8783897   1.8661962   1.8541697   1.8423069
#>  [151]   1.8306045   1.8190594   1.8076685   1.7964288   1.7853373   1.7743912
#>  [157]   1.7635878   1.7529242   1.7423980   1.7320064   1.7217471   1.7116175
#>  [163]   1.7016152   1.6917380   1.6819835   1.6723496   1.6628340   1.6534347
#>  [169]   1.6441496   1.6349767   1.6259139   1.6169595   1.6081115   1.5993680
#>  [175]   1.5907273   1.5821877   1.5737474   1.5654047   1.5571580   1.5490057
#>  [181]   1.5409463   1.5329782   1.5250999   1.5173100   1.5096069   1.5019894
#>  [187]   1.4944560   1.4870055   1.4796364   1.4723475   1.4651376   1.4580054
#>  [193]   1.4509496   1.4439692   1.4370630   1.4302298   1.4234684   1.4167779
#>  [199]   1.4101572   1.4036052   1.3971208   1.3907031   1.3843511   1.3780639
#>  [205]   1.3718404   1.3656798   1.3595811   1.3535434   1.3475659   1.3416478
#>  [211]   1.3357881   1.3299860   1.3242408   1.3185515   1.3129176   1.3073381
#>  [217]   1.3018124   1.2963396   1.2909191   1.2855501   1.2802320   1.2749641
#>  [223]   1.2697457   1.2645760   1.2594546   1.2543807   1.2493537   1.2443729
#>  [229]   1.2394379   1.2345480   1.2297025   1.2249010   1.2201428   1.2154275
#>  [235]   1.2107544   1.2061230   1.2015328   1.1969833   1.1924740   1.1880043
#>  [241]   1.1835738   1.1791819   1.1748283   1.1705124   1.1662338   1.1619920
#>  [247]   1.1577866   1.1536171   1.1494831   1.1453841   1.1413198   1.1372898
#>  [253]   1.1332936   1.1293308   1.1254011   1.1215040   1.1176392   1.1138063
#>  [259]   1.1100049   1.1062347   1.1024953   1.0987864   1.0951075   1.0914585
#>  [265]   1.0878388   1.0842482   1.0806864   1.0771531   1.0736479   1.0701705
#>  [271]   1.0667206   1.0632979   1.0599021   1.0565330   1.0531901   1.0498733
#>  [277]   1.0465822   1.0433167   1.0400763   1.0368608   1.0336700   1.0305037
#>  [283]   1.0273614   1.0242431   1.0211484   1.0180770   1.0150289   1.0120036
#>  [289]   1.0090010   1.0060208   1.0030628   1.0001268   0.9972126   0.9943198
#>  [295]   0.9914484   0.9885981   0.9857686   0.9829598   0.9801715   0.9774034
#>  [301]   0.9746554   0.9719273   0.9692188   0.9665297   0.9638599   0.9612093
#>  [307]   0.9585775   0.9559644   0.9533698   0.9507936   0.9482356   0.9456956
#>  [313]   0.9431734   0.9406689   0.9381818   0.9357121   0.9332596   0.9308240
#>  [319]   0.9284053   0.9260032   0.9236177   0.9212485   0.9188956   0.9165587
#>  [325]   0.9142378   0.9119326   0.9096431   0.9073690   0.9051103   0.9028668
#>  [331]   0.9006383   0.8984248   0.8962261   0.8940420   0.8918725   0.8897174
#>  [337]   0.8875765   0.8854498   0.8833371   0.8812383   0.8791533   0.8770820
#>  [343]   0.8750241   0.8729797   0.8709486   0.8689307   0.8669259   0.8649340
#>  [349]   0.8629550   0.8609887   0.8590350   0.8570938   0.8551651   0.8532487
#>  [355]   0.8513445   0.8494523   0.8475722   0.8457040   0.8438476   0.8420029
#>  [361]   0.8401698   0.8383482   0.8365380   0.8347391   0.8329515   0.8311749
#>  [367]   0.8294095   0.8276550   0.8259113   0.8241785   0.8224563   0.8207447
#>  [373]   0.8190436   0.8173530   0.8156728   0.8140027   0.8123429   0.8106932
#>  [379]   0.8090535   0.8074238   0.8058039   0.8041938   0.8025934   0.8010027
#>  [385]   0.7994215   0.7978498   0.7962875   0.7947346   0.7931910   0.7916565
#>  [391]   0.7901312   0.7886149   0.7871076   0.7856092   0.7841197   0.7826390
#>  [397]   0.7811670   0.7797036   0.7782488   0.7768026   0.7753648   0.7739354
#>  [403]   0.7725143   0.7711016   0.7696970   0.7683006   0.7669122   0.7655319
#>  [409]   0.7641596   0.7627952   0.7614386   0.7600899   0.7587489   0.7574155
#>  [415]   0.7560898   0.7547717   0.7534611   0.7521580   0.7508622   0.7495739
#>  [421]   0.7482928   0.7470190   0.7457524   0.7444930   0.7432407   0.7419954
#>  [427]   0.7407571   0.7395258   0.7383014   0.7370838   0.7358730   0.7346691
#>  [433]   0.7334718   0.7322812   0.7310972   0.7299198   0.7287489   0.7275845
#>  [439]   0.7264266   0.7252751   0.7241299   0.7229910   0.7218584   0.7207321
#>  [445]   0.7196119   0.7184978   0.7173899   0.7162880   0.7151922   0.7141023
#>  [451]   0.7130184   0.7119403   0.7108682   0.7098018   0.7087413   0.7076865
#>  [457]   0.7066374   0.7055940   0.7045562   0.7035240   0.7024974   0.7014763
#>  [463]   0.7004607   0.6994505   0.6984458   0.6974464   0.6964524   0.6954637
#>  [469]   0.6944803   0.6935022   0.6925293   0.6915615   0.6905989   0.6896414
#>  [475]   0.6886890   0.6877417   0.6867993   0.6858620   0.6849296   0.6840021
#>  [481]   0.6830796   0.6821619   0.6812490   0.6803410   0.6794377   0.6785392
#>  [487]   0.6776453   0.6767562   0.6758717   0.6749919   0.6741167   0.6732460
#>  [493]   0.6723799   0.6715183   0.6706612   0.6698085   0.6689603   0.6681165
#>  [499]   0.6672771   0.6664421   0.6656113   0.6647849   0.6639628   0.6631449
#>  [505]   0.6623313   0.6615218   0.6607165   0.6599154   0.6591184   0.6583256
#>  [511]   0.6575368   0.6567520   0.6559713   0.6551946   0.6544219   0.6536532
#>  [517]   0.6528884   0.6521275   0.6513705   0.6506174   0.6498681   0.6491227
#>  [523]   0.6483810   0.6476432   0.6469091   0.6461788   0.6454521   0.6447292
#>  [529]   0.6440100   0.6432944   0.6425825   0.6418741   0.6411694   0.6404682
#>  [535]   0.6397706   0.6390766   0.6383860   0.6376989   0.6370154   0.6363352
#>  [541]   0.6356586   0.6349853   0.6343154   0.6336490   0.6329858   0.6323261
#>  [547]   0.6316696   0.6310165   0.6303666   0.6297201   0.6290767   0.6284367
#>  [553]   0.6277998   0.6271661   0.6265356   0.6259083   0.6252842   0.6246631
#>  [559]   0.6240452   0.6234304   0.6228187   0.6222100   0.6216044   0.6210018
#>  [565]   0.6204023   0.6198057   0.6192121   0.6186215   0.6180339   0.6174492
#>  [571]   0.6168674   0.6162885   0.6157125   0.6151394   0.6145692   0.6140018
#>  [577]   0.6134372   0.6128755   0.6123165   0.6117604   0.6112070   0.6106564
#>  [583]   0.6101085   0.6095634   0.6090210   0.6084813   0.6079443   0.6074099
#>  [589]   0.6068782   0.6063492   0.6058228   0.6052990   0.6047779   0.6042593
#>  [595]   0.6037433   0.6032299   0.6027191   0.6022108   0.6017050   0.6012017
#>  [601]   0.6007010   0.6002027   0.5997069   0.5992136   0.5987228   0.5982344
#>  [607]   0.5977484   0.5972649   0.5967837   0.5963050   0.5958286   0.5953546
#>  [613]   0.5948830   0.5944137   0.5939468   0.5934822   0.5930199   0.5925599
#>  [619]   0.5921022   0.5916468   0.5911937   0.5907428   0.5902942   0.5898478
#>  [625]   0.5894037   0.5889617   0.5885220   0.5880844   0.5876491   0.5872159
#>  [631]   0.5867849   0.5863561   0.5859294   0.5855048   0.5850823   0.5846620
#>  [637]   0.5842438   0.5838276   0.5834136   0.5830016   0.5825917   0.5821838
#>  [643]   0.5817780   0.5813743   0.5809725   0.5805728   0.5801751   0.5797794
#>  [649]   0.5793856   0.5789939   0.5786041   0.5782163   0.5778304   0.5774465
#>  [655]   0.5770646   0.5766845   0.5763064   0.5759301   0.5755558   0.5751834
#>  [661]   0.5748128   0.5744442   0.5740773   0.5737124   0.5733493   0.5729880
#>  [667]   0.5726286   0.5722710   0.5719152   0.5715612   0.5712091   0.5708587
#>  [673]   0.5705101   0.5701633   0.5698182   0.5694749   0.5691334   0.5687936
#>  [679]   0.5684555   0.5681192   0.5677846   0.5674517   0.5671205   0.5667910
#>  [685]   0.5664632   0.5661371   0.5658126   0.5654899   0.5651688   0.5648493
#>  [691]   0.5645315   0.5642154   0.5639008   0.5635879   0.5632767   0.5629670
#>  [697]   0.5626589   0.5623525   0.5620476   0.5617443   0.5614426   0.5611425
#>  [703]   0.5608439   0.5605469   0.5602515   0.5599576   0.5596652   0.5593743
#>  [709]   0.5590850   0.5587972   0.5585109   0.5582262   0.5579429   0.5576611
#>  [715]   0.5573808   0.5571020   0.5568247   0.5565488   0.5562744   0.5560014
#>  [721]   0.5557299   0.5554599   0.5551913   0.5549241   0.5546583   0.5543940
#>  [727]   0.5541311   0.5538695   0.5536094   0.5533507   0.5530934   0.5528375
#>  [733]   0.5525829   0.5523297   0.5520779   0.5518275   0.5515784   0.5513307
#>  [739]   0.5510843   0.5508392   0.5505955   0.5503531   0.5501121   0.5498724
#>  [745]   0.5496339   0.5493968   0.5491610   0.5489265   0.5486933   0.5484614
#>  [751]   0.5482307   0.5480014   0.5477733   0.5475465   0.5473209   0.5470966
#>  [757]   0.5468736   0.5466518   0.5464312   0.5462119   0.5459938   0.5457769
#>  [763]   0.5455613   0.5453469   0.5451337   0.5449217   0.5447109   0.5445013
#>  [769]   0.5442930   0.5440858   0.5438797   0.5436749   0.5434713   0.5432688
#>  [775]   0.5430675   0.5428673   0.5426683   0.5424705   0.5422738   0.5420782
#>  [781]   0.5418838   0.5416906   0.5414984   0.5413074   0.5411176   0.5409288
#>  [787]   0.5407411   0.5405546   0.5403692   0.5401848   0.5400016   0.5398195
#>  [793]   0.5396384   0.5394585   0.5392796   0.5391018   0.5389251   0.5387494
#>  [799]   0.5385748   0.5384013   0.5382288   0.5380574   0.5378870   0.5377177
#>  [805]   0.5375494   0.5373822   0.5372160   0.5370508   0.5368866   0.5367235
#>  [811]   0.5365614   0.5364003   0.5362402   0.5360811   0.5359230   0.5357660
#>  [817]   0.5356099   0.5354548   0.5353007   0.5351476   0.5349955   0.5348443
#>  [823]   0.5346941   0.5345449   0.5343967   0.5342494   0.5341031   0.5339578
#>  [829]   0.5338134   0.5336699   0.5335274   0.5333859   0.5332452   0.5331056
#>  [835]   0.5329668   0.5328290   0.5326921   0.5325561   0.5324211   0.5322869
#>  [841]   0.5321537   0.5320214   0.5318900   0.5317595   0.5316299   0.5315012
#>  [847]   0.5313734   0.5312465   0.5311205   0.5309953   0.5308711   0.5307477
#>  [853]   0.5306252   0.5305035   0.5303828   0.5302629   0.5301439   0.5300257
#>  [859]   0.5299084   0.5297919   0.5296763   0.5295615   0.5294476   0.5293345
#>  [865]   0.5292223   0.5291109   0.5290003   0.5288906   0.5287817   0.5286736
#>  [871]   0.5285663   0.5284599   0.5283543   0.5282495   0.5281455   0.5280423
#>  [877]   0.5279399   0.5278383   0.5277375   0.5276375   0.5275383   0.5274399
#>  [883]   0.5273423   0.5272455   0.5271495   0.5270542   0.5269597   0.5268660
#>  [889]   0.5267731   0.5266809   0.5265895   0.5264989   0.5264090   0.5263199
#>  [895]   0.5262316   0.5261440   0.5260572   0.5259711   0.5258857   0.5258011
#>  [901]   0.5257173   0.5256342   0.5255518   0.5254702   0.5253893   0.5253091
#>  [907]   0.5252296   0.5251509   0.5250729   0.5249956   0.5249191   0.5248432
#>  [913]   0.5247681   0.5246937   0.5246200   0.5245470   0.5244747   0.5244031
#>  [919]   0.5243322   0.5242620   0.5241925   0.5241237   0.5240556   0.5239881
#>  [925]   0.5239214   0.5238553   0.5237899   0.5237252   0.5236612   0.5235979
#>  [931]   0.5235352   0.5234732   0.5234119   0.5233512   0.5232912   0.5232319
#>  [937]   0.5231732   0.5231152   0.5230578   0.5230011   0.5229450   0.5228896
#>  [943]   0.5228349   0.5227808   0.5227273   0.5226745   0.5226223   0.5225707
#>  [949]   0.5225198   0.5224695   0.5224199   0.5223708   0.5223224   0.5222747
#>  [955]   0.5222275   0.5221810   0.5221351   0.5220898   0.5220451   0.5220011
#>  [961]   0.5219576   0.5219148   0.5218726   0.5218309   0.5217899   0.5217495
#>  [967]   0.5217097   0.5216705   0.5216319   0.5215938   0.5215564   0.5215196
#>  [973]   0.5214833   0.5214477   0.5214126   0.5213781   0.5213442   0.5213109
#>  [979]   0.5212781   0.5212460   0.5212144   0.5211833   0.5211529   0.5211230
#>  [985]   0.5210937   0.5210649   0.5210368   0.5210091   0.5209821   0.5209556
#>  [991]   0.5209296   0.5209043   0.5208794   0.5208552   0.5208314   0.5208083
#>  [997]   0.5207856   0.5207636   0.5207420   0.5207210
bsmd1(20, 30, 0:2000 / 1000, 0.15, 0, 0.25)
#>    [1]        -Inf -51.2648036 -36.2334788 -29.5712749 -25.5980120 -22.8853050
#>    [7] -20.8819699 -19.3242820 -18.0681049 -17.0271305 -16.1461043 -15.3877938
#>   [13] -14.7260726 -14.1419943 -13.6214393 -13.1536410 -12.7302265 -12.3445732
#>   [19] -11.9913650 -11.6662779 -11.3657546 -11.0868378 -10.8270459 -10.5842793
#>   [25] -10.3567476 -10.1429135  -9.9414487  -9.7511986  -9.5711542  -9.4004293
#>   [31]  -9.2382419  -9.0838988  -8.9367835  -8.7963451  -8.6620898  -8.5335736
#>   [37]  -8.4103959  -8.2921944  -8.1786400  -8.0694335  -7.9643022  -7.8629964
#>   [43]  -7.7652878  -7.6709663  -7.5798390  -7.4917277  -7.4064681  -7.3239081
#>   [49]  -7.2439067  -7.1663330  -7.0910654  -7.0179907  -6.9470031  -6.8780040
#>   [55]  -6.8109013  -6.7456085  -6.6820449  -6.6201344  -6.5598058  -6.5009918
#>   [61]  -6.4436295  -6.3876592  -6.3330247  -6.2796731  -6.2275542  -6.1766205
#>   [67]  -6.1268273  -6.0781319  -6.0304940  -5.9838752  -5.9382393  -5.8935515
#>   [73]  -5.8497790  -5.8068906  -5.7648562  -5.7236475  -5.6832374  -5.6436000
#>   [79]  -5.6047104  -5.5665452  -5.5290816  -5.4922981  -5.4561739  -5.4206893
#>   [85]  -5.3858252  -5.3515634  -5.3178864  -5.2847776  -5.2522206  -5.2202002
#>   [91]  -5.1887014  -5.1577100  -5.1272121  -5.0971946  -5.0676447  -5.0385501
#>   [97]  -5.0098990  -4.9816800  -4.9538822  -4.9264949  -4.8995079  -4.8729113
#>  [103]  -4.8466956  -4.8208515  -4.7953702  -4.7702431  -4.7454618  -4.7210183
#>  [109]  -4.6969048  -4.6731138  -4.6496379  -4.6264701  -4.6036036  -4.5810316
#>  [115]  -4.5587479  -4.5367460  -4.5150201  -4.4935642  -4.4723726  -4.4514398
#>  [121]  -4.4307604  -4.4103293  -4.3901413  -4.3701916  -4.3504753  -4.3309878
#>  [127]  -4.3117247  -4.2926814  -4.2738538  -4.2552377  -4.2368290  -4.2186239
#>  [133]  -4.2006183  -4.1828088  -4.1651915  -4.1477630  -4.1305198  -4.1134585
#>  [139]  -4.0965760  -4.0798689  -4.0633342  -4.0469688  -4.0307698  -4.0147342
#>  [145]  -3.9988594  -3.9831424  -3.9675806  -3.9521715  -3.9369124  -3.9218008
#>  [151]  -3.9068343  -3.8920106  -3.8773272  -3.8627819  -3.8483725  -3.8340968
#>  [157]  -3.8199527  -3.8059381  -3.7920509  -3.7782892  -3.7646511  -3.7511345
#>  [163]  -3.7377378  -3.7244589  -3.7112963  -3.6982480  -3.6853124  -3.6724879
#>  [169]  -3.6597727  -3.6471653  -3.6346642  -3.6222677  -3.6099744  -3.5977827
#>  [175]  -3.5856913  -3.5736987  -3.5618036  -3.5500044  -3.5383001  -3.5266891
#>  [181]  -3.5151702  -3.5037423  -3.4924039  -3.4811540  -3.4699913  -3.4589146
#>  [187]  -3.4479229  -3.4370149  -3.4261897  -3.4154460  -3.4047828  -3.3941992
#>  [193]  -3.3836940  -3.3732663  -3.3629151  -3.3526393  -3.3424381  -3.3323105
#>  [199]  -3.3222556  -3.3122725  -3.3023603  -3.2925181  -3.2827451  -3.2730404
#>  [205]  -3.2634032  -3.2538327  -3.2443281  -3.2348886  -3.2255134  -3.2162017
#>  [211]  -3.2069529  -3.1977661  -3.1886407  -3.1795759  -3.1705711  -3.1616255
#>  [217]  -3.1527385  -3.1439094  -3.1351375  -3.1264222  -3.1177629  -3.1091590
#>  [223]  -3.1006097  -3.0921146  -3.0836730  -3.0752843  -3.0669480  -3.0586635
#>  [229]  -3.0504301  -3.0422475  -3.0341150  -3.0260321  -3.0179982  -3.0100129
#>  [235]  -3.0020756  -2.9941858  -2.9863431  -2.9785469  -2.9707967  -2.9630922
#>  [241]  -2.9554327  -2.9478179  -2.9402473  -2.9327205  -2.9252369  -2.9177962
#>  [247]  -2.9103979  -2.9030417  -2.8957270  -2.8884536  -2.8812209  -2.8740285
#>  [253]  -2.8668762  -2.8597635  -2.8526899  -2.8456552  -2.8386589  -2.8317007
#>  [259]  -2.8247803  -2.8178972  -2.8110511  -2.8042416  -2.7974685  -2.7907314
#>  [265]  -2.7840298  -2.7773636  -2.7707324  -2.7641358  -2.7575736  -2.7510453
#>  [271]  -2.7445508  -2.7380897  -2.7316617  -2.7252665  -2.7189038  -2.7125733
#>  [277]  -2.7062747  -2.7000078  -2.6937723  -2.6875678  -2.6813942  -2.6752511
#>  [283]  -2.6691383  -2.6630555  -2.6570024  -2.6509789  -2.6449845  -2.6390192
#>  [289]  -2.6330827  -2.6271746  -2.6212948  -2.6154430  -2.6096190  -2.6038225
#>  [295]  -2.5980534  -2.5923113  -2.5865961  -2.5809076  -2.5752455  -2.5696095
#>  [301]  -2.5639996  -2.5584155  -2.5528569  -2.5473237  -2.5418156  -2.5363325
#>  [307]  -2.5308742  -2.5254404  -2.5200310  -2.5146457  -2.5092844  -2.5039469
#>  [313]  -2.4986330  -2.4933425  -2.4880752  -2.4828310  -2.4776096  -2.4724110
#>  [319]  -2.4672348  -2.4620810  -2.4569493  -2.4518397  -2.4467519  -2.4416857
#>  [325]  -2.4366411  -2.4316178  -2.4266157  -2.4216346  -2.4166743  -2.4117348
#>  [331]  -2.4068158  -2.4019173  -2.3970389  -2.3921807  -2.3873424  -2.3825239
#>  [337]  -2.3777251  -2.3729459  -2.3681859  -2.3634453  -2.3587237  -2.3540211
#>  [343]  -2.3493373  -2.3446722  -2.3400256  -2.3353975  -2.3307877  -2.3261960
#>  [349]  -2.3216224  -2.3170666  -2.3125287  -2.3080084  -2.3035057  -2.2990204
#>  [355]  -2.2945523  -2.2901015  -2.2856677  -2.2812508  -2.2768508  -2.2724675
#>  [361]  -2.2681007  -2.2637505  -2.2594166  -2.2550990  -2.2507975  -2.2465121
#>  [367]  -2.2422426  -2.2379889  -2.2337510  -2.2295287  -2.2253218  -2.2211304
#>  [373]  -2.2169544  -2.2127935  -2.2086477  -2.2045170  -2.2004011  -2.1963001
#>  [379]  -2.1922138  -2.1881422  -2.1840850  -2.1800423  -2.1760140  -2.1719999
#>  [385]  -2.1680000  -2.1640141  -2.1600422  -2.1560842  -2.1521401  -2.1482096
#>  [391]  -2.1442928  -2.1403896  -2.1364998  -2.1326233  -2.1287602  -2.1249103
#>  [397]  -2.1210736  -2.1172499  -2.1134391  -2.1096413  -2.1058562  -2.1020840
#>  [403]  -2.0983243  -2.0945773  -2.0908427  -2.0871206  -2.0834109  -2.0797134
#>  [409]  -2.0760281  -2.0723550  -2.0686939  -2.0650449  -2.0614077  -2.0577824
#>  [415]  -2.0541689  -2.0505671  -2.0469769  -2.0433983  -2.0398312  -2.0362756
#>  [421]  -2.0327313  -2.0291983  -2.0256766  -2.0221660  -2.0186666  -2.0151782
#>  [427]  -2.0117007  -2.0082343  -2.0047786  -2.0013338  -1.9978997  -1.9944763
#>  [433]  -1.9910634  -1.9876612  -1.9842694  -1.9808881  -1.9775172  -1.9741565
#>  [439]  -1.9708062  -1.9674660  -1.9641360  -1.9608161  -1.9575062  -1.9542063
#>  [445]  -1.9509164  -1.9476363  -1.9443660  -1.9411055  -1.9378547  -1.9346135
#>  [451]  -1.9313820  -1.9281600  -1.9249475  -1.9217445  -1.9185509  -1.9153666
#>  [457]  -1.9121917  -1.9090260  -1.9058694  -1.9027221  -1.8995838  -1.8964546
#>  [463]  -1.8933344  -1.8902232  -1.8871209  -1.8840274  -1.8809428  -1.8778670
#>  [469]  -1.8747999  -1.8717414  -1.8686916  -1.8656504  -1.8626178  -1.8595936
#>  [475]  -1.8565779  -1.8535707  -1.8505718  -1.8475812  -1.8445990  -1.8416250
#>  [481]  -1.8386592  -1.8357015  -1.8327520  -1.8298106  -1.8268772  -1.8239518
#>  [487]  -1.8210344  -1.8181249  -1.8152232  -1.8123294  -1.8094435  -1.8065653
#>  [493]  -1.8036948  -1.8008320  -1.7979769  -1.7951293  -1.7922894  -1.7894570
#>  [499]  -1.7866321  -1.7838146  -1.7810046  -1.7782020  -1.7754067  -1.7726188
#>  [505]  -1.7698381  -1.7670647  -1.7642986  -1.7615396  -1.7587877  -1.7560430
#>  [511]  -1.7533053  -1.7505747  -1.7478511  -1.7451345  -1.7424248  -1.7397220
#>  [517]  -1.7370262  -1.7343371  -1.7316549  -1.7289795  -1.7263108  -1.7236489
#>  [523]  -1.7209936  -1.7183450  -1.7157030  -1.7130676  -1.7104388  -1.7078166
#>  [529]  -1.7052008  -1.7025915  -1.6999887  -1.6973922  -1.6948022  -1.6922185
#>  [535]  -1.6896412  -1.6870702  -1.6845054  -1.6819469  -1.6793946  -1.6768484
#>  [541]  -1.6743085  -1.6717747  -1.6692469  -1.6667253  -1.6642097  -1.6617001
#>  [547]  -1.6591966  -1.6566990  -1.6542073  -1.6517216  -1.6492418  -1.6467678
#>  [553]  -1.6442997  -1.6418374  -1.6393808  -1.6369301  -1.6344850  -1.6320457
#>  [559]  -1.6296121  -1.6271842  -1.6247618  -1.6223451  -1.6199340  -1.6175285
#>  [565]  -1.6151285  -1.6127340  -1.6103450  -1.6079615  -1.6055835  -1.6032109
#>  [571]  -1.6008436  -1.5984818  -1.5961253  -1.5937741  -1.5914283  -1.5890877
#>  [577]  -1.5867524  -1.5844224  -1.5820976  -1.5797780  -1.5774635  -1.5751543
#>  [583]  -1.5728501  -1.5705511  -1.5682572  -1.5659683  -1.5636845  -1.5614057
#>  [589]  -1.5591319  -1.5568632  -1.5545994  -1.5523405  -1.5500866  -1.5478376
#>  [595]  -1.5455934  -1.5433542  -1.5411198  -1.5388902  -1.5366654  -1.5344454
#>  [601]  -1.5322302  -1.5300198  -1.5278140  -1.5256130  -1.5234167  -1.5212251
#>  [607]  -1.5190381  -1.5168558  -1.5146780  -1.5125049  -1.5103364  -1.5081724
#>  [613]  -1.5060130  -1.5038581  -1.5017077  -1.4995619  -1.4974204  -1.4952835
#>  [619]  -1.4931510  -1.4910229  -1.4888992  -1.4867800  -1.4846650  -1.4825545
#>  [625]  -1.4804483  -1.4783464  -1.4762488  -1.4741555  -1.4720665  -1.4699817
#>  [631]  -1.4679012  -1.4658249  -1.4637528  -1.4616849  -1.4596211  -1.4575616
#>  [637]  -1.4555061  -1.4534548  -1.4514076  -1.4493645  -1.4473255  -1.4452906
#>  [643]  -1.4432597  -1.4412329  -1.4392100  -1.4371912  -1.4351763  -1.4331655
#>  [649]  -1.4311586  -1.4291556  -1.4271566  -1.4251615  -1.4231703  -1.4211830
#>  [655]  -1.4191995  -1.4172199  -1.4152442  -1.4132723  -1.4113042  -1.4093399
#>  [661]  -1.4073794  -1.4054227  -1.4034697  -1.4015205  -1.3995750  -1.3976333
#>  [667]  -1.3956952  -1.3937608  -1.3918302  -1.3899032  -1.3879798  -1.3860601
#>  [673]  -1.3841440  -1.3822315  -1.3803227  -1.3784174  -1.3765157  -1.3746175
#>  [679]  -1.3727229  -1.3708319  -1.3689444  -1.3670603  -1.3651798  -1.3633028
#>  [685]  -1.3614293  -1.3595592  -1.3576925  -1.3558293  -1.3539696  -1.3521132
#>  [691]  -1.3502603  -1.3484107  -1.3465646  -1.3447218  -1.3428823  -1.3410462
#>  [697]  -1.3392134  -1.3373840  -1.3355579  -1.3337350  -1.3319155  -1.3300992
#>  [703]  -1.3282862  -1.3264764  -1.3246699  -1.3228666  -1.3210666  -1.3192697
#>  [709]  -1.3174761  -1.3156856  -1.3138984  -1.3121142  -1.3103333  -1.3085555
#>  [715]  -1.3067808  -1.3050092  -1.3032408  -1.3014754  -1.2997132  -1.2979540
#>  [721]  -1.2961980  -1.2944449  -1.2926949  -1.2909480  -1.2892041  -1.2874632
#>  [727]  -1.2857254  -1.2839905  -1.2822586  -1.2805297  -1.2788038  -1.2770808
#>  [733]  -1.2753608  -1.2736437  -1.2719296  -1.2702184  -1.2685101  -1.2668047
#>  [739]  -1.2651022  -1.2634026  -1.2617059  -1.2600120  -1.2583210  -1.2566328
#>  [745]  -1.2549475  -1.2532650  -1.2515854  -1.2499085  -1.2482344  -1.2465632
#>  [751]  -1.2448947  -1.2432290  -1.2415661  -1.2399059  -1.2382485  -1.2365938
#>  [757]  -1.2349418  -1.2332926  -1.2316461  -1.2300023  -1.2283612  -1.2267227
#>  [763]  -1.2250870  -1.2234539  -1.2218235  -1.2201958  -1.2185707  -1.2169482
#>  [769]  -1.2153284  -1.2137112  -1.2120966  -1.2104846  -1.2088752  -1.2072684
#>  [775]  -1.2056642  -1.2040625  -1.2024635  -1.2008669  -1.1992730  -1.1976815
#>  [781]  -1.1960927  -1.1945063  -1.1929224  -1.1913411  -1.1897623  -1.1881859
#>  [787]  -1.1866121  -1.1850407  -1.1834718  -1.1819054  -1.1803415  -1.1787799
#>  [793]  -1.1772209  -1.1756642  -1.1741100  -1.1725582  -1.1710089  -1.1694619
#>  [799]  -1.1679173  -1.1663752  -1.1648354  -1.1632980  -1.1617629  -1.1602303
#>  [805]  -1.1586999  -1.1571720  -1.1556463  -1.1541231  -1.1526021  -1.1510835
#>  [811]  -1.1495671  -1.1480531  -1.1465414  -1.1450320  -1.1435249  -1.1420200
#>  [817]  -1.1405175  -1.1390172  -1.1375191  -1.1360233  -1.1345298  -1.1330385
#>  [823]  -1.1315494  -1.1300626  -1.1285780  -1.1270956  -1.1256154  -1.1241375
#>  [829]  -1.1226617  -1.1211881  -1.1197167  -1.1182474  -1.1167804  -1.1153155
#>  [835]  -1.1138527  -1.1123922  -1.1109337  -1.1094774  -1.1080233  -1.1065712
#>  [841]  -1.1051213  -1.1036735  -1.1022279  -1.1007843  -1.0993428  -1.0979034
#>  [847]  -1.0964661  -1.0950309  -1.0935977  -1.0921667  -1.0907377  -1.0893107
#>  [853]  -1.0878858  -1.0864629  -1.0850421  -1.0836233  -1.0822066  -1.0807919
#>  [859]  -1.0793791  -1.0779684  -1.0765597  -1.0751530  -1.0737483  -1.0723456
#>  [865]  -1.0709449  -1.0695462  -1.0681494  -1.0667546  -1.0653617  -1.0639709
#>  [871]  -1.0625819  -1.0611949  -1.0598099  -1.0584268  -1.0570456  -1.0556663
#>  [877]  -1.0542890  -1.0529136  -1.0515401  -1.0501685  -1.0487987  -1.0474309
#>  [883]  -1.0460650  -1.0447010  -1.0433388  -1.0419785  -1.0406201  -1.0392635
#>  [889]  -1.0379088  -1.0365560  -1.0352050  -1.0338558  -1.0325085  -1.0311630
#>  [895]  -1.0298194  -1.0284776  -1.0271376  -1.0257994  -1.0244630  -1.0231284
#>  [901]  -1.0217956  -1.0204646  -1.0191354  -1.0178080  -1.0164824  -1.0151585
#>  [907]  -1.0138365  -1.0125162  -1.0111976  -1.0098808  -1.0085658  -1.0072525
#>  [913]  -1.0059409  -1.0046311  -1.0033231  -1.0020167  -1.0007121  -0.9994092
#>  [919]  -0.9981080  -0.9968086  -0.9955108  -0.9942148  -0.9929204  -0.9916278
#>  [925]  -0.9903368  -0.9890475  -0.9877599  -0.9864740  -0.9851897  -0.9839072
#>  [931]  -0.9826262  -0.9813470  -0.9800694  -0.9787934  -0.9775191  -0.9762465
#>  [937]  -0.9749755  -0.9737061  -0.9724383  -0.9711722  -0.9699077  -0.9686448
#>  [943]  -0.9673836  -0.9661239  -0.9648659  -0.9636094  -0.9623546  -0.9611013
#>  [949]  -0.9598497  -0.9585996  -0.9573511  -0.9561042  -0.9548588  -0.9536151
#>  [955]  -0.9523729  -0.9511322  -0.9498932  -0.9486556  -0.9474197  -0.9461852
#>  [961]  -0.9449523  -0.9437210  -0.9424912  -0.9412629  -0.9400362  -0.9388110
#>  [967]  -0.9375873  -0.9363651  -0.9351444  -0.9339252  -0.9327076  -0.9314914
#>  [973]  -0.9302768  -0.9290636  -0.9278519  -0.9266418  -0.9254331  -0.9242259
#>  [979]  -0.9230201  -0.9218159  -0.9206131  -0.9194117  -0.9182119  -0.9170135
#>  [985]  -0.9158165  -0.9146210  -0.9134270  -0.9122344  -0.9110432  -0.9098535
#>  [991]  -0.9086652  -0.9074783  -0.9062929  -0.9051089  -0.9039263  -0.9027451
#>  [997]  -0.9015654  -0.9003870  -0.8992101  -0.8980346  -0.8968604  -0.8956877
#> [1003]  -0.8945164  -0.8933464  -0.8921779  -0.8910107  -0.8898449  -0.8886805
#> [1009]  -0.8875174  -0.8863558  -0.8851955  -0.8840365  -0.8828790  -0.8817227
#> [1015]  -0.8805679  -0.8794144  -0.8782622  -0.8771114  -0.8759619  -0.8748138
#> [1021]  -0.8736670  -0.8725216  -0.8713774  -0.8702346  -0.8690931  -0.8679530
#> [1027]  -0.8668142  -0.8656766  -0.8645404  -0.8634055  -0.8622719  -0.8611396
#> [1033]  -0.8600086  -0.8588790  -0.8577506  -0.8566234  -0.8554976  -0.8543731
#> [1039]  -0.8532498  -0.8521279  -0.8510072  -0.8498877  -0.8487696  -0.8476527
#> [1045]  -0.8465371  -0.8454227  -0.8443096  -0.8431978  -0.8420872  -0.8409779
#> [1051]  -0.8398698  -0.8387629  -0.8376573  -0.8365530  -0.8354499  -0.8343480
#> [1057]  -0.8332473  -0.8321479  -0.8310497  -0.8299527  -0.8288569  -0.8277624
#> [1063]  -0.8266691  -0.8255769  -0.8244860  -0.8233963  -0.8223078  -0.8212205
#> [1069]  -0.8201344  -0.8190495  -0.8179658  -0.8168833  -0.8158020  -0.8147218
#> [1075]  -0.8136429  -0.8125651  -0.8114885  -0.8104131  -0.8093388  -0.8082657
#> [1081]  -0.8071938  -0.8061231  -0.8050535  -0.8039851  -0.8029178  -0.8018517
#> [1087]  -0.8007867  -0.7997229  -0.7986602  -0.7975987  -0.7965383  -0.7954791
#> [1093]  -0.7944210  -0.7933640  -0.7923082  -0.7912535  -0.7901999  -0.7891475
#> [1099]  -0.7880961  -0.7870459  -0.7859968  -0.7849489  -0.7839020  -0.7828562
#> [1105]  -0.7818116  -0.7807681  -0.7797256  -0.7786843  -0.7776441  -0.7766049
#> [1111]  -0.7755669  -0.7745300  -0.7734941  -0.7724593  -0.7714256  -0.7703930
#> [1117]  -0.7693615  -0.7683311  -0.7673017  -0.7662734  -0.7652462  -0.7642200
#> [1123]  -0.7631949  -0.7621709  -0.7611480  -0.7601261  -0.7591052  -0.7580854
#> [1129]  -0.7570667  -0.7560490  -0.7550324  -0.7540168  -0.7530023  -0.7519888
#> [1135]  -0.7509763  -0.7499649  -0.7489545  -0.7479452  -0.7469368  -0.7459296
#> [1141]  -0.7449233  -0.7439181  -0.7429139  -0.7419107  -0.7409085  -0.7399073
#> [1147]  -0.7389072  -0.7379081  -0.7369100  -0.7359129  -0.7349168  -0.7339217
#> [1153]  -0.7329276  -0.7319345  -0.7309424  -0.7299513  -0.7289612  -0.7279721
#> [1159]  -0.7269840  -0.7259969  -0.7250107  -0.7240256  -0.7230414  -0.7220582
#> [1165]  -0.7210760  -0.7200947  -0.7191145  -0.7181352  -0.7171568  -0.7161795
#> [1171]  -0.7152031  -0.7142277  -0.7132532  -0.7122797  -0.7113072  -0.7103356
#> [1177]  -0.7093650  -0.7083953  -0.7074266  -0.7064588  -0.7054920  -0.7045261
#> [1183]  -0.7035611  -0.7025971  -0.7016341  -0.7006720  -0.6997108  -0.6987505
#> [1189]  -0.6977912  -0.6968328  -0.6958754  -0.6949189  -0.6939633  -0.6930086
#> [1195]  -0.6920548  -0.6911020  -0.6901501  -0.6891991  -0.6882490  -0.6872998
#> [1201]  -0.6863515  -0.6854042  -0.6844577  -0.6835122  -0.6825675  -0.6816238
#> [1207]  -0.6806810  -0.6797390  -0.6787980  -0.6778578  -0.6769186  -0.6759802
#> [1213]  -0.6750427  -0.6741062  -0.6731705  -0.6722356  -0.6713017  -0.6703687
#> [1219]  -0.6694365  -0.6685052  -0.6675748  -0.6666452  -0.6657166  -0.6647888
#> [1225]  -0.6638618  -0.6629358  -0.6620106  -0.6610862  -0.6601628  -0.6592402
#> [1231]  -0.6583184  -0.6573975  -0.6564775  -0.6555583  -0.6546400  -0.6537225
#> [1237]  -0.6528059  -0.6518901  -0.6509752  -0.6500611  -0.6491478  -0.6482354
#> [1243]  -0.6473239  -0.6464131  -0.6455033  -0.6445942  -0.6436860  -0.6427786
#> [1249]  -0.6418721  -0.6409663  -0.6400614  -0.6391574  -0.6382541  -0.6373517
#> [1255]  -0.6364501  -0.6355493  -0.6346493  -0.6337502  -0.6328519  -0.6319543
#> [1261]  -0.6310576  -0.6301617  -0.6292666  -0.6283724  -0.6274789  -0.6265862
#> [1267]  -0.6256944  -0.6248033  -0.6239130  -0.6230236  -0.6221349  -0.6212470
#> [1273]  -0.6203600  -0.6194737  -0.6185882  -0.6177035  -0.6168196  -0.6159364
#> [1279]  -0.6150541  -0.6141725  -0.6132918  -0.6124118  -0.6115326  -0.6106541
#> [1285]  -0.6097765  -0.6088996  -0.6080235  -0.6071482  -0.6062736  -0.6053998
#> [1291]  -0.6045268  -0.6036545  -0.6027830  -0.6019123  -0.6010423  -0.6001731
#> [1297]  -0.5993047  -0.5984370  -0.5975701  -0.5967039  -0.5958385  -0.5949739
#> [1303]  -0.5941099  -0.5932468  -0.5923844  -0.5915227  -0.5906618  -0.5898017
#> [1309]  -0.5889422  -0.5880835  -0.5872256  -0.5863684  -0.5855120  -0.5846562
#> [1315]  -0.5838012  -0.5829470  -0.5820935  -0.5812407  -0.5803886  -0.5795373
#> [1321]  -0.5786867  -0.5778369  -0.5769877  -0.5761393  -0.5752916  -0.5744446
#> [1327]  -0.5735984  -0.5727529  -0.5719080  -0.5710639  -0.5702206  -0.5693779
#> [1333]  -0.5685359  -0.5676947  -0.5668542  -0.5660143  -0.5651752  -0.5643368
#> [1339]  -0.5634991  -0.5626621  -0.5618258  -0.5609903  -0.5601554  -0.5593212
#> [1345]  -0.5584877  -0.5576549  -0.5568228  -0.5559914  -0.5551607  -0.5543307
#> [1351]  -0.5535013  -0.5526727  -0.5518447  -0.5510175  -0.5501909  -0.5493650
#> [1357]  -0.5485398  -0.5477153  -0.5468915  -0.5460683  -0.5452458  -0.5444240
#> [1363]  -0.5436029  -0.5427824  -0.5419627  -0.5411436  -0.5403251  -0.5395074
#> [1369]  -0.5386903  -0.5378739  -0.5370581  -0.5362431  -0.5354286  -0.5346149
#> [1375]  -0.5338018  -0.5329894  -0.5321776  -0.5313665  -0.5305560  -0.5297463
#> [1381]  -0.5289371  -0.5281286  -0.5273208  -0.5265137  -0.5257071  -0.5249013
#> [1387]  -0.5240961  -0.5232915  -0.5224876  -0.5216843  -0.5208817  -0.5200797
#> [1393]  -0.5192783  -0.5184777  -0.5176776  -0.5168782  -0.5160794  -0.5152813
#> [1399]  -0.5144838  -0.5136869  -0.5128907  -0.5120951  -0.5113001  -0.5105058
#> [1405]  -0.5097121  -0.5089190  -0.5081266  -0.5073348  -0.5065436  -0.5057530
#> [1411]  -0.5049631  -0.5041738  -0.5033851  -0.5025970  -0.5018096  -0.5010227
#> [1417]  -0.5002365  -0.4994509  -0.4986659  -0.4978816  -0.4970978  -0.4963147
#> [1423]  -0.4955322  -0.4947503  -0.4939690  -0.4931883  -0.4924082  -0.4916287
#> [1429]  -0.4908498  -0.4900716  -0.4892939  -0.4885169  -0.4877404  -0.4869646
#> [1435]  -0.4861893  -0.4854147  -0.4846406  -0.4838672  -0.4830943  -0.4823220
#> [1441]  -0.4815504  -0.4807793  -0.4800088  -0.4792389  -0.4784696  -0.4777009
#> [1447]  -0.4769328  -0.4761653  -0.4753983  -0.4746319  -0.4738662  -0.4731010
#> [1453]  -0.4723364  -0.4715724  -0.4708089  -0.4700460  -0.4692838  -0.4685221
#> [1459]  -0.4677609  -0.4670004  -0.4662404  -0.4654810  -0.4647222  -0.4639639
#> [1465]  -0.4632063  -0.4624492  -0.4616926  -0.4609366  -0.4601812  -0.4594264
#> [1471]  -0.4586722  -0.4579185  -0.4571653  -0.4564128  -0.4556607  -0.4549093
#> [1477]  -0.4541584  -0.4534081  -0.4526583  -0.4519091  -0.4511605  -0.4504124
#> [1483]  -0.4496649  -0.4489179  -0.4481715  -0.4474256  -0.4466803  -0.4459355
#> [1489]  -0.4451913  -0.4444477  -0.4437046  -0.4429620  -0.4422200  -0.4414785
#> [1495]  -0.4407376  -0.4399972  -0.4392574  -0.4385181  -0.4377793  -0.4370411
#> [1501]  -0.4363035  -0.4355663  -0.4348298  -0.4340937  -0.4333582  -0.4326232
#> [1507]  -0.4318888  -0.4311549  -0.4304215  -0.4296887  -0.4289564  -0.4282246
#> [1513]  -0.4274934  -0.4267627  -0.4260325  -0.4253028  -0.4245737  -0.4238451
#> [1519]  -0.4231171  -0.4223895  -0.4216625  -0.4209360  -0.4202100  -0.4194846
#> [1525]  -0.4187596  -0.4180352  -0.4173113  -0.4165880  -0.4158651  -0.4151428
#> [1531]  -0.4144210  -0.4136997  -0.4129789  -0.4122586  -0.4115389  -0.4108196
#> [1537]  -0.4101009  -0.4093827  -0.4086649  -0.4079477  -0.4072310  -0.4065149
#> [1543]  -0.4057992  -0.4050840  -0.4043693  -0.4036552  -0.4029415  -0.4022284
#> [1549]  -0.4015157  -0.4008036  -0.4000919  -0.3993808  -0.3986701  -0.3979600
#> [1555]  -0.3972503  -0.3965412  -0.3958325  -0.3951244  -0.3944167  -0.3937095
#> [1561]  -0.3930029  -0.3922967  -0.3915910  -0.3908858  -0.3901811  -0.3894769
#> [1567]  -0.3887732  -0.3880699  -0.3873672  -0.3866649  -0.3859631  -0.3852619
#> [1573]  -0.3845611  -0.3838607  -0.3831609  -0.3824615  -0.3817627  -0.3810643
#> [1579]  -0.3803664  -0.3796690  -0.3789720  -0.3782755  -0.3775796  -0.3768840
#> [1585]  -0.3761890  -0.3754944  -0.3748004  -0.3741068  -0.3734136  -0.3727210
#> [1591]  -0.3720288  -0.3713371  -0.3706458  -0.3699550  -0.3692647  -0.3685749
#> [1597]  -0.3678855  -0.3671966  -0.3665082  -0.3658202  -0.3651327  -0.3644457
#> [1603]  -0.3637591  -0.3630730  -0.3623874  -0.3617022  -0.3610175  -0.3603332
#> [1609]  -0.3596494  -0.3589661  -0.3582832  -0.3576008  -0.3569188  -0.3562373
#> [1615]  -0.3555563  -0.3548757  -0.3541956  -0.3535159  -0.3528366  -0.3521579
#> [1621]  -0.3514796  -0.3508017  -0.3501243  -0.3494473  -0.3487708  -0.3480947
#> [1627]  -0.3474191  -0.3467439  -0.3460692  -0.3453949  -0.3447211  -0.3440477
#> [1633]  -0.3433748  -0.3427023  -0.3420302  -0.3413586  -0.3406875  -0.3400167
#> [1639]  -0.3393465  -0.3386766  -0.3380072  -0.3373382  -0.3366697  -0.3360016
#> [1645]  -0.3353340  -0.3346668  -0.3340000  -0.3333337  -0.3326678  -0.3320023
#> [1651]  -0.3313373  -0.3306727  -0.3300085  -0.3293448  -0.3286815  -0.3280186
#> [1657]  -0.3273561  -0.3266941  -0.3260326  -0.3253714  -0.3247107  -0.3240504
#> [1663]  -0.3233905  -0.3227310  -0.3220720  -0.3214134  -0.3207553  -0.3200975
#> [1669]  -0.3194402  -0.3187833  -0.3181268  -0.3174707  -0.3168151  -0.3161599
#> [1675]  -0.3155051  -0.3148507  -0.3141968  -0.3135432  -0.3128901  -0.3122374
#> [1681]  -0.3115851  -0.3109332  -0.3102818  -0.3096307  -0.3089801  -0.3083299
#> [1687]  -0.3076801  -0.3070307  -0.3063817  -0.3057331  -0.3050849  -0.3044372
#> [1693]  -0.3037899  -0.3031429  -0.3024964  -0.3018503  -0.3012046  -0.3005593
#> [1699]  -0.2999144  -0.2992699  -0.2986258  -0.2979821  -0.2973389  -0.2966960
#> [1705]  -0.2960535  -0.2954115  -0.2947698  -0.2941285  -0.2934877  -0.2928472
#> [1711]  -0.2922072  -0.2915675  -0.2909282  -0.2902894  -0.2896509  -0.2890128
#> [1717]  -0.2883751  -0.2877379  -0.2871010  -0.2864645  -0.2858284  -0.2851927
#> [1723]  -0.2845574  -0.2839225  -0.2832880  -0.2826538  -0.2820201  -0.2813867
#> [1729]  -0.2807538  -0.2801212  -0.2794890  -0.2788572  -0.2782258  -0.2775948
#> [1735]  -0.2769642  -0.2763340  -0.2757041  -0.2750746  -0.2744456  -0.2738169
#> [1741]  -0.2731885  -0.2725606  -0.2719331  -0.2713059  -0.2706791  -0.2700527
#> [1747]  -0.2694267  -0.2688010  -0.2681758  -0.2675509  -0.2669264  -0.2663023
#> [1753]  -0.2656785  -0.2650552  -0.2644322  -0.2638096  -0.2631873  -0.2625655
#> [1759]  -0.2619440  -0.2613228  -0.2607021  -0.2600817  -0.2594618  -0.2588421
#> [1765]  -0.2582229  -0.2576040  -0.2569855  -0.2563674  -0.2557496  -0.2551322
#> [1771]  -0.2545152  -0.2538985  -0.2532822  -0.2526663  -0.2520508  -0.2514356
#> [1777]  -0.2508208  -0.2502063  -0.2495922  -0.2489785  -0.2483651  -0.2477522
#> [1783]  -0.2471395  -0.2465273  -0.2459153  -0.2453038  -0.2446926  -0.2440818
#> [1789]  -0.2434713  -0.2428612  -0.2422515  -0.2416421  -0.2410331  -0.2404245
#> [1795]  -0.2398161  -0.2392082  -0.2386006  -0.2379934  -0.2373865  -0.2367800
#> [1801]  -0.2361738  -0.2355680  -0.2349626  -0.2343575  -0.2337527  -0.2331483
#> [1807]  -0.2325443  -0.2319406  -0.2313372  -0.2307343  -0.2301316  -0.2295293
#> [1813]  -0.2289274  -0.2283258  -0.2277246  -0.2271237  -0.2265232  -0.2259230
#> [1819]  -0.2253231  -0.2247236  -0.2241245  -0.2235257  -0.2229272  -0.2223291
#> [1825]  -0.2217313  -0.2211339  -0.2205368  -0.2199401  -0.2193437  -0.2187476
#> [1831]  -0.2181519  -0.2175565  -0.2169615  -0.2163668  -0.2157725  -0.2151785
#> [1837]  -0.2145848  -0.2139915  -0.2133985  -0.2128058  -0.2122135  -0.2116216
#> [1843]  -0.2110299  -0.2104386  -0.2098476  -0.2092570  -0.2086667  -0.2080768
#> [1849]  -0.2074871  -0.2068978  -0.2063089  -0.2057203  -0.2051320  -0.2045440
#> [1855]  -0.2039564  -0.2033691  -0.2027821  -0.2021955  -0.2016092  -0.2010232
#> [1861]  -0.2004376  -0.1998523  -0.1992673  -0.1986826  -0.1980983  -0.1975143
#> [1867]  -0.1969306  -0.1963473  -0.1957643  -0.1951816  -0.1945992  -0.1940172
#> [1873]  -0.1934355  -0.1928541  -0.1922730  -0.1916923  -0.1911118  -0.1905317
#> [1879]  -0.1899520  -0.1893725  -0.1887934  -0.1882146  -0.1876361  -0.1870579
#> [1885]  -0.1864801  -0.1859025  -0.1853253  -0.1847484  -0.1841719  -0.1835956
#> [1891]  -0.1830197  -0.1824440  -0.1818687  -0.1812938  -0.1807191  -0.1801447
#> [1897]  -0.1795707  -0.1789970  -0.1784236  -0.1778505  -0.1772777  -0.1767052
#> [1903]  -0.1761331  -0.1755612  -0.1749897  -0.1744185  -0.1738476  -0.1732770
#> [1909]  -0.1727067  -0.1721367  -0.1715671  -0.1709977  -0.1704287  -0.1698600
#> [1915]  -0.1692915  -0.1687234  -0.1681556  -0.1675881  -0.1670209  -0.1664541
#> [1921]  -0.1658875  -0.1653212  -0.1647552  -0.1641896  -0.1636242  -0.1630592
#> [1927]  -0.1624944  -0.1619300  -0.1613659  -0.1608020  -0.1602385  -0.1596753
#> [1933]  -0.1591123  -0.1585497  -0.1579874  -0.1574254  -0.1568637  -0.1563022
#> [1939]  -0.1557411  -0.1551803  -0.1546198  -0.1540596  -0.1534996  -0.1529400
#> [1945]  -0.1523807  -0.1518217  -0.1512629  -0.1507045  -0.1501464  -0.1495885
#> [1951]  -0.1490310  -0.1484737  -0.1479168  -0.1473601  -0.1468038  -0.1462477
#> [1957]  -0.1456919  -0.1451364  -0.1445812  -0.1440263  -0.1434717  -0.1429174
#> [1963]  -0.1423634  -0.1418097  -0.1412562  -0.1407031  -0.1401502  -0.1395977
#> [1969]  -0.1390454  -0.1384934  -0.1379417  -0.1373903  -0.1368392  -0.1362884
#> [1975]  -0.1357378  -0.1351876  -0.1346376  -0.1340879  -0.1335385  -0.1329894
#> [1981]  -0.1324406  -0.1318921  -0.1313438  -0.1307958  -0.1302482  -0.1297008
#> [1987]  -0.1291536  -0.1286068  -0.1280603  -0.1275140  -0.1269680  -0.1264223
#> [1993]  -0.1258769  -0.1253318  -0.1247869  -0.1242423  -0.1236980  -0.1231540
#> [1999]  -0.1226103  -0.1220669  -0.1215237