Computes a measure of moneyness for Black-Scholes-Merton model.

Moneyness formula implementation, for Black-Scholes-Merton model.

Usage

bsmmoneyness(spot, strike, time, rate, yield)

Arguments

spot

current stock price.

strike

the strike price.

time

time to option expiration in years.

rate

the risk-free interest rate.

yield

the dividends that are expected to be paid.

Value

Vector with moneyness.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

Examples

bsmmoneyness(0:49, 1:50, 3, 0.1, 0)
#>  [1]  1.00000000  0.32507060  0.10009413 -0.01239411 -0.07988705 -0.12488234
#>  [7] -0.15702184 -0.18112646 -0.19987450 -0.21487293 -0.22714437 -0.23737057
#> [13] -0.24602351 -0.25344032 -0.25986822 -0.26549263 -0.27045535 -0.27486665
#> [19] -0.27881361 -0.28236587 -0.28557982 -0.28850159 -0.29116929 -0.29361469
#> [25] -0.29586446 -0.29794116 -0.29986404 -0.30164956 -0.30331195 -0.30486351
#> [31] -0.30631498 -0.30767572 -0.30895400 -0.31015708 -0.31129141 -0.31236273
#> [37] -0.31337614 -0.31433621 -0.31524704 -0.31611234 -0.31693542 -0.31771931
#> [43] -0.31846674 -0.31918020 -0.31986195 -0.32051405 -0.32113841 -0.32173675
#> [49] -0.32231067 -0.32286163
bsmmoneyness(10, 10, 1, 0.12, 0:1000 / 1000)
#>    [1] -0.1274968516 -0.1263699183 -0.1252441114 -0.1241194297 -0.1229958721
#>    [6] -0.1218734376 -0.1207521249 -0.1196319329 -0.1185128606 -0.1173949069
#>   [11] -0.1162780705 -0.1151623503 -0.1140477454 -0.1129342545 -0.1118218765
#>   [16] -0.1107106104 -0.1096004549 -0.1084914091 -0.1073834717 -0.1062766418
#>   [21] -0.1051709181 -0.1040662996 -0.1029627851 -0.1018603736 -0.1007590640
#>   [26] -0.0996588551 -0.0985597459 -0.0974617353 -0.0963648221 -0.0952690053
#>   [31] -0.0941742837 -0.0930806563 -0.0919881220 -0.0908966797 -0.0898063283
#>   [36] -0.0887170667 -0.0876288938 -0.0865418085 -0.0854558098 -0.0843708966
#>   [41] -0.0832870677 -0.0822043221 -0.0811226587 -0.0800420764 -0.0789625742
#>   [46] -0.0778841509 -0.0768068055 -0.0757305369 -0.0746553441 -0.0735812259
#>   [51] -0.0725081813 -0.0714362091 -0.0703653085 -0.0692954782 -0.0682267172
#>   [56] -0.0671590244 -0.0660923988 -0.0650268392 -0.0639623447 -0.0628989142
#>   [61] -0.0618365465 -0.0607752407 -0.0597149957 -0.0586558104 -0.0575976837
#>   [66] -0.0565406147 -0.0554846022 -0.0544296451 -0.0533757425 -0.0523228933
#>   [71] -0.0512710964 -0.0502203507 -0.0491706553 -0.0481220091 -0.0470744110
#>   [76] -0.0460278599 -0.0449823549 -0.0439378949 -0.0428944788 -0.0418521055
#>   [81] -0.0408107742 -0.0397704837 -0.0387312329 -0.0376930208 -0.0366558465
#>   [86] -0.0356197088 -0.0345846067 -0.0335505392 -0.0325175053 -0.0314855039
#>   [91] -0.0304545340 -0.0294245945 -0.0283956844 -0.0273678028 -0.0263409485
#>   [96] -0.0253151205 -0.0242903179 -0.0232665395 -0.0222437845 -0.0212220516
#>  [101] -0.0202013400 -0.0191816486 -0.0181629764 -0.0171453223 -0.0161286854
#>  [106] -0.0151130646 -0.0140984589 -0.0130848674 -0.0120722889 -0.0110607224
#>  [111] -0.0100501671 -0.0090406218 -0.0080320855 -0.0070245573 -0.0060180361
#>  [116] -0.0050125209 -0.0040080107 -0.0030045045 -0.0020020013 -0.0010005002
#>  [121]  0.0000000000  0.0009995002  0.0019980013  0.0029955045  0.0039920107
#>  [126]  0.0049875208  0.0059820359  0.0069755571  0.0079680852  0.0089596212
#>  [131]  0.0099501663  0.0109397212  0.0119282871  0.0129158650  0.0139024557
#>  [136]  0.0148880604  0.0158726799  0.0168563154  0.0178389676  0.0188206378
#>  [141]  0.0198013267  0.0207810354  0.0217597649  0.0227375162  0.0237142902
#>  [146]  0.0246900880  0.0256649104  0.0266387585  0.0276116332  0.0285835355
#>  [151]  0.0295544665  0.0305244269  0.0314934179  0.0324614404  0.0334284954
#>  [156]  0.0343945837  0.0353597065  0.0363238647  0.0372870591  0.0382492909
#>  [161]  0.0392105608  0.0401708701  0.0411302194  0.0420886099  0.0430460425
#>  [166]  0.0440025182  0.0449580378  0.0459126024  0.0468662129  0.0478188703
#>  [171]  0.0487705755  0.0497213295  0.0506711332  0.0516199875  0.0525678935
#>  [176]  0.0535148520  0.0544608641  0.0554059306  0.0563500526  0.0572932308
#>  [181]  0.0582354664  0.0591767602  0.0601171132  0.0610565263  0.0619950005
#>  [186]  0.0629325366  0.0638691357  0.0648047987  0.0657395264  0.0666733199
#>  [191]  0.0676061801  0.0685381079  0.0694691042  0.0703991700  0.0713283062
#>  [196]  0.0722565137  0.0731837934  0.0741101464  0.0750355735  0.0759600756
#>  [201]  0.0768836536  0.0778063086  0.0787280413  0.0796488528  0.0805687439
#>  [206]  0.0814877156  0.0824057688  0.0833229044  0.0842391233  0.0851544264
#>  [211]  0.0860688147  0.0869822891  0.0878948505  0.0888064997  0.0897172378
#>  [216]  0.0906270655  0.0915359839  0.0924439939  0.0933510962  0.0942572920
#>  [221]  0.0951625820  0.0960669671  0.0969704483  0.0978730265  0.0987747026
#>  [226]  0.0996754774  0.1005753519  0.1014743270  0.1023724036  0.1032695825
#>  [231]  0.1041658647  0.1050612511  0.1059557425  0.1068493399  0.1077420441
#>  [236]  0.1086338561  0.1095247767  0.1104148068  0.1113039474  0.1121921992
#>  [241]  0.1130795633  0.1139660404  0.1148516315  0.1157363374  0.1166201591
#>  [246]  0.1175030974  0.1183851532  0.1192663274  0.1201466209  0.1210260345
#>  [251]  0.1219045691  0.1227822256  0.1236590049  0.1245349079  0.1254099354
#>  [256]  0.1262840883  0.1271573675  0.1280297739  0.1289013083  0.1297719715
#>  [261]  0.1306417646  0.1315106883  0.1323787435  0.1332459311  0.1341122519
#>  [266]  0.1349777069  0.1358422968  0.1367060226  0.1375688851  0.1384308851
#>  [271]  0.1392920236  0.1401523013  0.1410117193  0.1418702782  0.1427279790
#>  [276]  0.1435848225  0.1444408096  0.1452959412  0.1461502180  0.1470036410
#>  [281]  0.1478562110  0.1487079289  0.1495587955  0.1504088116  0.1512579781
#>  [286]  0.1521062959  0.1529537658  0.1538003887  0.1546461653  0.1554910966
#>  [291]  0.1563351834  0.1571784265  0.1580208268  0.1588623852  0.1597031023
#>  [296]  0.1605429792  0.1613820167  0.1622202155  0.1630575765  0.1638941006
#>  [301]  0.1647297886  0.1655646413  0.1663986596  0.1672318443  0.1680641962
#>  [306]  0.1688957161  0.1697264050  0.1705562636  0.1713852928  0.1722134933
#>  [311]  0.1730408661  0.1738674118  0.1746931315  0.1755180259  0.1763420957
#>  [316]  0.1771653419  0.1779877653  0.1788093667  0.1796301469  0.1804501067
#>  [321]  0.1812692469  0.1820875684  0.1829050721  0.1837217586  0.1845376288
#>  [326]  0.1853526836  0.1861669237  0.1869803500  0.1877929633  0.1886047644
#>  [331]  0.1894157540  0.1902259331  0.1910353024  0.1918438628  0.1926516150
#>  [336]  0.1934585598  0.1942646981  0.1950700307  0.1958745583  0.1966782818
#>  [341]  0.1974812020  0.1982833197  0.1990846357  0.1998851507  0.2006848656
#>  [346]  0.2014837812  0.2022818983  0.2030792177  0.2038757402  0.2046714665
#>  [351]  0.2054663975  0.2062605340  0.2070538767  0.2078464265  0.2086381841
#>  [356]  0.2094291504  0.2102193261  0.2110087120  0.2117973089  0.2125851176
#>  [361]  0.2133721389  0.2141583736  0.2149438224  0.2157284862  0.2165123657
#>  [366]  0.2172954618  0.2180777751  0.2188593065  0.2196400567  0.2204200266
#>  [371]  0.2211992169  0.2219776284  0.2227552619  0.2235321182  0.2243081980
#>  [376]  0.2250835020  0.2258580312  0.2266317862  0.2274047679  0.2281769770
#>  [381]  0.2289484142  0.2297190804  0.2304889763  0.2312581027  0.2320264603
#>  [386]  0.2327940500  0.2335608725  0.2343269285  0.2350922189  0.2358567444
#>  [391]  0.2366205057  0.2373835036  0.2381457389  0.2389072124  0.2396679247
#>  [396]  0.2404278768  0.2411870692  0.2419455029  0.2427031785  0.2434600968
#>  [401]  0.2442162585  0.2449716645  0.2457263155  0.2464802121  0.2472333553
#>  [406]  0.2479857457  0.2487373841  0.2494882712  0.2502384078  0.2509877946
#>  [411]  0.2517364324  0.2524843220  0.2532314640  0.2539778593  0.2547235086
#>  [416]  0.2554684125  0.2562125720  0.2569559876  0.2576986603  0.2584405906
#>  [421]  0.2591817793  0.2599222273  0.2606619351  0.2614009036  0.2621391335
#>  [426]  0.2628766256  0.2636133805  0.2643493991  0.2650846820  0.2658192300
#>  [431]  0.2665530438  0.2672861241  0.2680184718  0.2687500874  0.2694809718
#>  [436]  0.2702111257  0.2709405498  0.2716692449  0.2723972116  0.2731244507
#>  [441]  0.2738509629  0.2745767490  0.2753018097  0.2760261456  0.2767497576
#>  [446]  0.2774726464  0.2781948126  0.2789162570  0.2796369803  0.2803569833
#>  [451]  0.2810762666  0.2817948310  0.2825126771  0.2832298058  0.2839462178
#>  [456]  0.2846619136  0.2853768942  0.2860911601  0.2868047121  0.2875175509
#>  [461]  0.2882296772  0.2889410918  0.2896517953  0.2903617884  0.2910710720
#>  [466]  0.2917796465  0.2924875129  0.2931946717  0.2939011238  0.2946068697
#>  [471]  0.2953119103  0.2960162461  0.2967198780  0.2974228066  0.2981250326
#>  [476]  0.2988265568  0.2995273798  0.3002275023  0.3009269250  0.3016256486
#>  [481]  0.3023236739  0.3030210015  0.3037176322  0.3044135665  0.3051088053
#>  [486]  0.3058033491  0.3064971988  0.3071903550  0.3078828183  0.3085745895
#>  [491]  0.3092656694  0.3099560584  0.3106457575  0.3113347672  0.3120230882
#>  [496]  0.3127107212  0.3133976670  0.3140839261  0.3147694993  0.3154543873
#>  [501]  0.3161385908  0.3168221104  0.3175049468  0.3181871007  0.3188685728
#>  [506]  0.3195493638  0.3202294743  0.3209089051  0.3215876567  0.3222657300
#>  [511]  0.3229431255  0.3236198440  0.3242958860  0.3249712524  0.3256459438
#>  [516]  0.3263199608  0.3269933041  0.3276659744  0.3283379723  0.3290092986
#>  [521]  0.3296799540  0.3303499390  0.3310192543  0.3316879007  0.3323558787
#>  [526]  0.3330231891  0.3336898326  0.3343558097  0.3350211212  0.3356857677
#>  [531]  0.3363497499  0.3370130684  0.3376757239  0.3383377172  0.3389990487
#>  [536]  0.3396597193  0.3403197295  0.3409790801  0.3416377716  0.3422958047
#>  [541]  0.3429531802  0.3436098986  0.3442659606  0.3449213669  0.3455761181
#>  [546]  0.3462302149  0.3468836579  0.3475364478  0.3481885852  0.3488400708
#>  [551]  0.3494909053  0.3501410892  0.3507906233  0.3514395082  0.3520877445
#>  [556]  0.3527353329  0.3533822741  0.3540285686  0.3546742171  0.3553192204
#>  [561]  0.3559635789  0.3566072934  0.3572503645  0.3578927929  0.3585345792
#>  [566]  0.3591757240  0.3598162279  0.3604560917  0.3610953160  0.3617339013
#>  [571]  0.3623718484  0.3630091578  0.3636458303  0.3642818664  0.3649172668
#>  [576]  0.3655520321  0.3661861629  0.3668196599  0.3674525238  0.3680847551
#>  [581]  0.3687163545  0.3693473226  0.3699776601  0.3706073675  0.3712364455
#>  [586]  0.3718648948  0.3724927160  0.3731199096  0.3737464763  0.3743724168
#>  [591]  0.3749977317  0.3756224216  0.3762464871  0.3768699288  0.3774927474
#>  [596]  0.3781149435  0.3787365178  0.3793574707  0.3799778030  0.3805975153
#>  [601]  0.3812166082  0.3818350823  0.3824529382  0.3830701766  0.3836867981
#>  [606]  0.3843028032  0.3849181927  0.3855329671  0.3861471270  0.3867606730
#>  [611]  0.3873736058  0.3879859260  0.3885976342  0.3892087309  0.3898192169
#>  [616]  0.3904290927  0.3910383589  0.3916470162  0.3922550651  0.3928625063
#>  [621]  0.3934693403  0.3940755678  0.3946811894  0.3952862056  0.3958906171
#>  [626]  0.3964944246  0.3970976285  0.3977002295  0.3983022282  0.3989036253
#>  [631]  0.3995044212  0.4001046166  0.4007042122  0.4013032084  0.4019016059
#>  [636]  0.4024994054  0.4030966073  0.4036932124  0.4042892211  0.4048846341
#>  [641]  0.4054794520  0.4060736754  0.4066673049  0.4072603410  0.4078527844
#>  [646]  0.4084446356  0.4090358953  0.4096265640  0.4102166424  0.4108061310
#>  [651]  0.4113950303  0.4119833411  0.4125710638  0.4131581992  0.4137447476
#>  [656]  0.4143307099  0.4149160864  0.4155008779  0.4160850848  0.4166687079
#>  [661]  0.4172517476  0.4178342046  0.4184160794  0.4189973726  0.4195780849
#>  [666]  0.4201582167  0.4207377686  0.4213167413  0.4218951353  0.4224729512
#>  [671]  0.4230501896  0.4236268511  0.4242029361  0.4247784454  0.4253533794
#>  [676]  0.4259277388  0.4265015241  0.4270747359  0.4276473748  0.4282194414
#>  [681]  0.4287909362  0.4293618597  0.4299322126  0.4305019955  0.4310712088
#>  [686]  0.4316398532  0.4322079293  0.4327754376  0.4333423786  0.4339087530
#>  [691]  0.4344745613  0.4350398041  0.4356044819  0.4361685953  0.4367321449
#>  [696]  0.4372951312  0.4378575548  0.4384194163  0.4389807162  0.4395414550
#>  [701]  0.4401016334  0.4406612519  0.4412203111  0.4417788115  0.4423367537
#>  [706]  0.4428941382  0.4434509656  0.4440072364  0.4445629513  0.4451181107
#>  [711]  0.4456727153  0.4462267655  0.4467802619  0.4473332051  0.4478855957
#>  [716]  0.4484374341  0.4489887210  0.4495394569  0.4500896423  0.4506392778
#>  [721]  0.4511883639  0.4517369012  0.4522848903  0.4528323316  0.4533792258
#>  [726]  0.4539255734  0.4544713748  0.4550166308  0.4555613418  0.4561055083
#>  [731]  0.4566491309  0.4571922102  0.4577347467  0.4582767409  0.4588181934
#>  [736]  0.4593591047  0.4598994754  0.4604393059  0.4609785969  0.4615173489
#>  [741]  0.4620555624  0.4625932380  0.4631303761  0.4636669774  0.4642030423
#>  [746]  0.4647385715  0.4652735654  0.4658080245  0.4663419495  0.4668753408
#>  [751]  0.4674081990  0.4679405246  0.4684723181  0.4690035801  0.4695343111
#>  [756]  0.4700645117  0.4705941823  0.4711233235  0.4716519358  0.4721800198
#>  [761]  0.4727075760  0.4732346048  0.4737611069  0.4742870828  0.4748125329
#>  [766]  0.4753374579  0.4758618582  0.4763857344  0.4769090869  0.4774319164
#>  [771]  0.4779542232  0.4784760081  0.4789972714  0.4795180137  0.4800382355
#>  [776]  0.4805579374  0.4810771198  0.4815957833  0.4821139284  0.4826315557
#>  [781]  0.4831486655  0.4836652585  0.4841813352  0.4846968960  0.4852119415
#>  [786]  0.4857264723  0.4862404888  0.4867539915  0.4872669810  0.4877794577
#>  [791]  0.4882914222  0.4888028750  0.4893138166  0.4898242476  0.4903341683
#>  [796]  0.4908435794  0.4913524813  0.4918608746  0.4923687597  0.4928761372
#>  [801]  0.4933830076  0.4938893714  0.4943952291  0.4949005811  0.4954054281
#>  [806]  0.4959097704  0.4964136087  0.4969169434  0.4974197750  0.4979221040
#>  [811]  0.4984239309  0.4989252563  0.4994260806  0.4999264043  0.5004262279
#>  [816]  0.5009255520  0.5014243770  0.5019227034  0.5024205318  0.5029178625
#>  [821]  0.5034146962  0.5039110333  0.5044068743  0.5049022197  0.5053970700
#>  [826]  0.5058914257  0.5063852873  0.5068786553  0.5073715302  0.5078639124
#>  [831]  0.5083558025  0.5088472010  0.5093381083  0.5098285249  0.5103184514
#>  [836]  0.5108078882  0.5112968358  0.5117852947  0.5122732654  0.5127607483
#>  [841]  0.5132477440  0.5137342530  0.5142202757  0.5147058126  0.5151908642
#>  [846]  0.5156754310  0.5161595135  0.5166431122  0.5171262275  0.5176088599
#>  [851]  0.5180910099  0.5185726780  0.5190538647  0.5195345705  0.5200147957
#>  [856]  0.5204945410  0.5209738068  0.5214525936  0.5219309018  0.5224087319
#>  [861]  0.5228860845  0.5233629599  0.5238393587  0.5243152814  0.5247907283
#>  [866]  0.5252657001  0.5257401971  0.5262142198  0.5266877688  0.5271608444
#>  [871]  0.5276334473  0.5281055777  0.5285772363  0.5290484234  0.5295191396
#>  [876]  0.5299893853  0.5304591610  0.5309284671  0.5313973042  0.5318656727
#>  [881]  0.5323335730  0.5328010057  0.5332679711  0.5337344699  0.5342005024
#>  [886]  0.5346660690  0.5351311704  0.5355958068  0.5360599789  0.5365236870
#>  [891]  0.5369869317  0.5374497133  0.5379120324  0.5383738894  0.5388352848
#>  [896]  0.5392962190  0.5397566925  0.5402167058  0.5406762592  0.5411353534
#>  [901]  0.5415939887  0.5420521656  0.5425098845  0.5429671460  0.5434239504
#>  [906]  0.5438802982  0.5443361899  0.5447916260  0.5452466068  0.5457011329
#>  [911]  0.5461552047  0.5466088227  0.5470619872  0.5475146988  0.5479669580
#>  [916]  0.5484187651  0.5488701206  0.5493210250  0.5497714787  0.5502214822
#>  [921]  0.5506710359  0.5511201403  0.5515687958  0.5520170028  0.5524647619
#>  [926]  0.5529120734  0.5533589379  0.5538053557  0.5542513273  0.5546968532
#>  [931]  0.5551419338  0.5555865695  0.5560307608  0.5564745081  0.5569178119
#>  [936]  0.5573606726  0.5578030907  0.5582450666  0.5586866007  0.5591276936
#>  [941]  0.5595683455  0.5600085570  0.5604483285  0.5608876605  0.5613265534
#>  [946]  0.5617650075  0.5622030235  0.5626406016  0.5630777424  0.5635144463
#>  [951]  0.5639507137  0.5643865450  0.5648219407  0.5652569013  0.5656914271
#>  [956]  0.5661255186  0.5665591762  0.5669924004  0.5674251915  0.5678575501
#>  [961]  0.5682894766  0.5687209713  0.5691520348  0.5695826674  0.5700128696
#>  [966]  0.5704426418  0.5708719844  0.5713008980  0.5717293828  0.5721574393
#>  [971]  0.5725850681  0.5730122693  0.5734390437  0.5738653914  0.5742913130
#>  [976]  0.5747168089  0.5751418795  0.5755665253  0.5759907466  0.5764145439
#>  [981]  0.5768379177  0.5772608683  0.5776833961  0.5781055016  0.5785271852
#>  [986]  0.5789484474  0.5793692885  0.5797897089  0.5802097092  0.5806292897
#>  [991]  0.5810484508  0.5814671929  0.5818855165  0.5823034220  0.5827209098
#>  [996]  0.5831379803  0.5835546340  0.5839708712  0.5843866924  0.5848020979
#> [1001]  0.5852170883
bsmmoneyness(10, 10, 1, 0:1000 / 1000, 0)
#>    [1]  0.000000000 -0.001000500 -0.002002001 -0.003004505 -0.004008011
#>    [6] -0.005012521 -0.006018036 -0.007024557 -0.008032086 -0.009040622
#>   [11] -0.010050167 -0.011060722 -0.012072289 -0.013084867 -0.014098459
#>   [16] -0.015113065 -0.016128685 -0.017145322 -0.018162976 -0.019181649
#>   [21] -0.020201340 -0.021222052 -0.022243784 -0.023266540 -0.024290318
#>   [26] -0.025315121 -0.026340948 -0.027367803 -0.028395684 -0.029424594
#>   [31] -0.030454534 -0.031485504 -0.032517505 -0.033550539 -0.034584607
#>   [36] -0.035619709 -0.036655846 -0.037693021 -0.038731233 -0.039770484
#>   [41] -0.040810774 -0.041852106 -0.042894479 -0.043937895 -0.044982355
#>   [46] -0.046027860 -0.047074411 -0.048122009 -0.049170655 -0.050220351
#>   [51] -0.051271096 -0.052322893 -0.053375743 -0.054429645 -0.055484602
#>   [56] -0.056540615 -0.057597684 -0.058655810 -0.059714996 -0.060775241
#>   [61] -0.061836547 -0.062898914 -0.063962345 -0.065026839 -0.066092399
#>   [66] -0.067159024 -0.068226717 -0.069295478 -0.070365308 -0.071436209
#>   [71] -0.072508181 -0.073581226 -0.074655344 -0.075730537 -0.076806805
#>   [76] -0.077884151 -0.078962574 -0.080042076 -0.081122659 -0.082204322
#>   [81] -0.083287068 -0.084370897 -0.085455810 -0.086541809 -0.087628894
#>   [86] -0.088717067 -0.089806328 -0.090896680 -0.091988122 -0.093080656
#>   [91] -0.094174284 -0.095269005 -0.096364822 -0.097461735 -0.098559746
#>   [96] -0.099658855 -0.100759064 -0.101860374 -0.102962785 -0.104066300
#>  [101] -0.105170918 -0.106276642 -0.107383472 -0.108491409 -0.109600455
#>  [106] -0.110710610 -0.111821877 -0.112934254 -0.114047745 -0.115162350
#>  [111] -0.116278070 -0.117394907 -0.118512861 -0.119631933 -0.120752125
#>  [116] -0.121873438 -0.122995872 -0.124119430 -0.125244111 -0.126369918
#>  [121] -0.127496852 -0.128624912 -0.129754102 -0.130884421 -0.132015871
#>  [126] -0.133148453 -0.134282168 -0.135417018 -0.136553003 -0.137690124
#>  [131] -0.138828383 -0.139967781 -0.141108319 -0.142249998 -0.143392820
#>  [136] -0.144536784 -0.145681894 -0.146828149 -0.147975550 -0.149124100
#>  [141] -0.150273799 -0.151424648 -0.152576649 -0.153729802 -0.154884109
#>  [146] -0.156039570 -0.157196188 -0.158353963 -0.159512896 -0.160672989
#>  [151] -0.161834243 -0.162996658 -0.164160236 -0.165324979 -0.166490887
#>  [156] -0.167657961 -0.168826203 -0.169995614 -0.171166195 -0.172337947
#>  [161] -0.173510871 -0.174684969 -0.175860241 -0.177036690 -0.178214315
#>  [166] -0.179393119 -0.180573102 -0.181754265 -0.182936611 -0.184120139
#>  [171] -0.185304851 -0.186490749 -0.187677833 -0.188866105 -0.190055566
#>  [176] -0.191246217 -0.192438059 -0.193631093 -0.194825321 -0.196020744
#>  [181] -0.197217363 -0.198415179 -0.199614194 -0.200814408 -0.202015823
#>  [186] -0.203218440 -0.204422260 -0.205627285 -0.206833515 -0.208040952
#>  [191] -0.209249598 -0.210459452 -0.211670517 -0.212882794 -0.214096283
#>  [196] -0.215310986 -0.216526905 -0.217744041 -0.218962394 -0.220181966
#>  [201] -0.221402758 -0.222624772 -0.223848008 -0.225072468 -0.226298153
#>  [206] -0.227525065 -0.228753204 -0.229982572 -0.231213170 -0.232444999
#>  [211] -0.233678060 -0.234912355 -0.236147885 -0.237384651 -0.238622655
#>  [216] -0.239861897 -0.241102379 -0.242344102 -0.243587068 -0.244831277
#>  [221] -0.246076731 -0.247323431 -0.248571378 -0.249820574 -0.251071019
#>  [226] -0.252322716 -0.253575665 -0.254829868 -0.256085325 -0.257342039
#>  [231] -0.258600010 -0.259859239 -0.261119729 -0.262381479 -0.263644492
#>  [236] -0.264908769 -0.266174310 -0.267441118 -0.268709193 -0.269978537
#>  [241] -0.271249150 -0.272521035 -0.273794193 -0.275068624 -0.276344330
#>  [246] -0.277621313 -0.278899574 -0.280179113 -0.281459932 -0.282742033
#>  [251] -0.284025417 -0.285310084 -0.286596037 -0.287883277 -0.289171804
#>  [256] -0.290461621 -0.291752728 -0.293045127 -0.294338819 -0.295633805
#>  [261] -0.296930087 -0.298227665 -0.299526542 -0.300826719 -0.302128196
#>  [266] -0.303430976 -0.304735059 -0.306040446 -0.307347140 -0.308655141
#>  [271] -0.309964451 -0.311275070 -0.312587001 -0.313900245 -0.315214802
#>  [276] -0.316530675 -0.317847864 -0.319166371 -0.320486197 -0.321807344
#>  [281] -0.323129812 -0.324453604 -0.325778720 -0.327105162 -0.328432931
#>  [286] -0.329762028 -0.331092455 -0.332424213 -0.333757304 -0.335091729
#>  [291] -0.336427488 -0.337764584 -0.339103018 -0.340442790 -0.341783904
#>  [296] -0.343126359 -0.344470157 -0.345815299 -0.347161788 -0.348509623
#>  [301] -0.349858808 -0.351209342 -0.352561227 -0.353914464 -0.355269056
#>  [306] -0.356625003 -0.357982307 -0.359340968 -0.360700989 -0.362062371
#>  [311] -0.363425114 -0.364789221 -0.366154693 -0.367521531 -0.368889737
#>  [316] -0.370259311 -0.371630256 -0.373002572 -0.374376261 -0.375751325
#>  [321] -0.377127764 -0.378505581 -0.379884776 -0.381265351 -0.382647307
#>  [326] -0.384030646 -0.385415369 -0.386801477 -0.388188972 -0.389577856
#>  [331] -0.390968128 -0.392359792 -0.393752849 -0.395147298 -0.396543144
#>  [336] -0.397940385 -0.399339025 -0.400739064 -0.402140503 -0.403543345
#>  [341] -0.404947591 -0.406353241 -0.407760298 -0.409168762 -0.410578636
#>  [346] -0.411989920 -0.413402616 -0.414816725 -0.416232250 -0.417649190
#>  [351] -0.419067549 -0.420487326 -0.421908524 -0.423331143 -0.424755186
#>  [356] -0.426180654 -0.427607548 -0.429035870 -0.430465620 -0.431896802
#>  [361] -0.433329415 -0.434763461 -0.436198942 -0.437635859 -0.439074214
#>  [366] -0.440514008 -0.441955243 -0.443397919 -0.444842039 -0.446287604
#>  [371] -0.447734615 -0.449183073 -0.450632981 -0.452084340 -0.453537150
#>  [376] -0.454991415 -0.456447134 -0.457904309 -0.459362943 -0.460823036
#>  [381] -0.462284589 -0.463747605 -0.465212085 -0.466678030 -0.468145442
#>  [386] -0.469614321 -0.471084671 -0.472556491 -0.474029784 -0.475504551
#>  [391] -0.476980794 -0.478458513 -0.479937711 -0.481418389 -0.482900549
#>  [396] -0.484384191 -0.485869318 -0.487355930 -0.488844030 -0.490333619
#>  [401] -0.491824698 -0.493317268 -0.494811333 -0.496306892 -0.497803947
#>  [406] -0.499302500 -0.500802552 -0.502304106 -0.503807161 -0.505311720
#>  [411] -0.506817785 -0.508325357 -0.509834436 -0.511345026 -0.512857127
#>  [416] -0.514370741 -0.515885869 -0.517402513 -0.518920674 -0.520440355
#>  [421] -0.521961556 -0.523484278 -0.525008525 -0.526534296 -0.528061594
#>  [426] -0.529590420 -0.531120775 -0.532652662 -0.534186081 -0.535721034
#>  [431] -0.537257524 -0.538795550 -0.540335115 -0.541876221 -0.543418868
#>  [436] -0.544963059 -0.546508795 -0.548056077 -0.549604907 -0.551155287
#>  [441] -0.552707219 -0.554260702 -0.555815740 -0.557372334 -0.558930486
#>  [446] -0.560490196 -0.562051467 -0.563614299 -0.565178696 -0.566744657
#>  [451] -0.568312185 -0.569881282 -0.571451949 -0.573024187 -0.574597997
#>  [456] -0.576173383 -0.577750345 -0.579328884 -0.580909003 -0.582490703
#>  [461] -0.584073985 -0.585658851 -0.587245303 -0.588833342 -0.590422970
#>  [466] -0.592014189 -0.593606999 -0.595201403 -0.596797403 -0.598394999
#>  [471] -0.599994193 -0.601594988 -0.603197384 -0.604801383 -0.606406987
#>  [476] -0.608014197 -0.609623016 -0.611233444 -0.612845483 -0.614459136
#>  [481] -0.616074402 -0.617691285 -0.619309785 -0.620929905 -0.622551646
#>  [486] -0.624175009 -0.625799996 -0.627426609 -0.629054850 -0.630684720
#>  [491] -0.632316220 -0.633949353 -0.635584119 -0.637220521 -0.638858561
#>  [496] -0.640498239 -0.642139558 -0.643782519 -0.645427123 -0.647073374
#>  [501] -0.648721271 -0.650370817 -0.652022013 -0.653674861 -0.655329363
#>  [506] -0.656985520 -0.658643335 -0.660302808 -0.661963941 -0.663626736
#>  [511] -0.665291195 -0.666957319 -0.668625110 -0.670294570 -0.671965700
#>  [516] -0.673638502 -0.675312977 -0.676989128 -0.678666956 -0.680346463
#>  [521] -0.682027650 -0.683710519 -0.685395071 -0.687081309 -0.688769234
#>  [526] -0.690458848 -0.692150153 -0.693843149 -0.695537840 -0.697234225
#>  [531] -0.698932309 -0.700632091 -0.702333573 -0.704036758 -0.705741647
#>  [536] -0.707448242 -0.709156545 -0.710866556 -0.712578278 -0.714291713
#>  [541] -0.716006862 -0.717723727 -0.719442310 -0.721162613 -0.722884636
#>  [546] -0.724608382 -0.726333853 -0.728061051 -0.729789976 -0.731520631
#>  [551] -0.733253018 -0.734987138 -0.736722993 -0.738460584 -0.740199914
#>  [556] -0.741940985 -0.743683797 -0.745428353 -0.747174654 -0.748922703
#>  [561] -0.750672500 -0.752424048 -0.754177349 -0.755932404 -0.757689214
#>  [566] -0.759447783 -0.761208111 -0.762970200 -0.764734052 -0.766499668
#>  [571] -0.768267051 -0.770036203 -0.771807124 -0.773579818 -0.775354285
#>  [576] -0.777130527 -0.778908546 -0.780688345 -0.782469924 -0.784253285
#>  [581] -0.786038431 -0.787825362 -0.789614082 -0.791404591 -0.793196892
#>  [586] -0.794990986 -0.796786874 -0.798584560 -0.800384044 -0.802185329
#>  [591] -0.803988415 -0.805793306 -0.807600003 -0.809408507 -0.811218820
#>  [596] -0.813030945 -0.814844883 -0.816660635 -0.818478205 -0.820297592
#>  [601] -0.822118800 -0.823941831 -0.825766685 -0.827593365 -0.829421872
#>  [606] -0.831252209 -0.833084377 -0.834918378 -0.836754214 -0.838591887
#>  [611] -0.840431399 -0.842272751 -0.844115945 -0.845960983 -0.847807867
#>  [616] -0.849656600 -0.851507181 -0.853359615 -0.855213901 -0.857070043
#>  [621] -0.858928042 -0.860787900 -0.862649618 -0.864513200 -0.866378645
#>  [626] -0.868245957 -0.870115138 -0.871986188 -0.873859111 -0.875733907
#>  [631] -0.877610579 -0.879489129 -0.881369558 -0.883251869 -0.885136063
#>  [636] -0.887022141 -0.888910107 -0.890799962 -0.892691708 -0.894585346
#>  [641] -0.896480879 -0.898378309 -0.900277637 -0.902178865 -0.904081995
#>  [646] -0.905987029 -0.907893970 -0.909802818 -0.911713576 -0.913626246
#>  [651] -0.915540829 -0.917457328 -0.919375744 -0.921296080 -0.923218337
#>  [656] -0.925142517 -0.927068623 -0.928996655 -0.930926617 -0.932858509
#>  [661] -0.934792334 -0.936728094 -0.938665791 -0.940605427 -0.942547003
#>  [666] -0.944490521 -0.946435984 -0.948383394 -0.950332752 -0.952284060
#>  [671] -0.954237321 -0.956192535 -0.958149706 -0.960108835 -0.962069925
#>  [676] -0.964032976 -0.965997991 -0.967964973 -0.969933922 -0.971904841
#>  [681] -0.973877732 -0.975852597 -0.977829438 -0.979808257 -0.981789055
#>  [686] -0.983771836 -0.985756600 -0.987743349 -0.989732087 -0.991722814
#>  [691] -0.993715533 -0.995710246 -0.997706954 -0.999705661 -1.001706366
#>  [696] -1.003709074 -1.005713785 -1.007720502 -1.009729227 -1.011739961
#>  [701] -1.013752707 -1.015767467 -1.017784243 -1.019803037 -1.021823850
#>  [706] -1.023846685 -1.025871544 -1.027898429 -1.029927341 -1.031958284
#>  [711] -1.033991259 -1.036026267 -1.038063312 -1.040102395 -1.042143517
#>  [716] -1.044186682 -1.046231891 -1.048279147 -1.050328450 -1.052379804
#>  [721] -1.054433211 -1.056488671 -1.058546189 -1.060605764 -1.062667401
#>  [726] -1.064731100 -1.066796864 -1.068864694 -1.070934594 -1.073006564
#>  [731] -1.075080608 -1.077156726 -1.079234922 -1.081315197 -1.083397553
#>  [736] -1.085481993 -1.087568518 -1.089657130 -1.091747833 -1.093840627
#>  [741] -1.095935514 -1.098032498 -1.100131580 -1.102232762 -1.104336046
#>  [746] -1.106441435 -1.108548930 -1.110658534 -1.112770248 -1.114884075
#>  [751] -1.117000017 -1.119118075 -1.121238253 -1.123360553 -1.125484975
#>  [756] -1.127611523 -1.129740199 -1.131871004 -1.134003942 -1.136139013
#>  [761] -1.138276220 -1.140415566 -1.142557052 -1.144700681 -1.146846454
#>  [766] -1.148994375 -1.151144444 -1.153296664 -1.155451038 -1.157607567
#>  [771] -1.159766254 -1.161927100 -1.164090109 -1.166255281 -1.168422620
#>  [776] -1.170592127 -1.172763805 -1.174937656 -1.177113681 -1.179291884
#>  [781] -1.181472265 -1.183654829 -1.185839576 -1.188026509 -1.190215630
#>  [786] -1.192406941 -1.194600444 -1.196796142 -1.198994037 -1.201194131
#>  [791] -1.203396426 -1.205600925 -1.207807629 -1.210016541 -1.212227663
#>  [796] -1.214440997 -1.216656545 -1.218874311 -1.221094295 -1.223316500
#>  [801] -1.225540928 -1.227767583 -1.229996464 -1.232227576 -1.234460920
#>  [806] -1.236696499 -1.238934314 -1.241174368 -1.243416664 -1.245661202
#>  [811] -1.247907987 -1.250157019 -1.252408301 -1.254661836 -1.256917626
#>  [816] -1.259175672 -1.261435978 -1.263698545 -1.265963376 -1.268230473
#>  [821] -1.270499838 -1.272771473 -1.275045381 -1.277321565 -1.279600025
#>  [826] -1.281880765 -1.284163787 -1.286449094 -1.288736686 -1.291026568
#>  [831] -1.293318740 -1.295613206 -1.297909967 -1.300209027 -1.302510386
#>  [836] -1.304814048 -1.307120015 -1.309428289 -1.311738872 -1.314051768
#>  [841] -1.316366977 -1.318684502 -1.321004347 -1.323326512 -1.325651000
#>  [846] -1.327977815 -1.330306957 -1.332638429 -1.334972234 -1.337308375
#>  [851] -1.339646852 -1.341987669 -1.344330828 -1.346676331 -1.349024181
#>  [856] -1.351374381 -1.353726931 -1.356081835 -1.358439095 -1.360798714
#>  [861] -1.363160694 -1.365525036 -1.367891745 -1.370260821 -1.372632267
#>  [866] -1.375006086 -1.377382280 -1.379760851 -1.382141802 -1.384525136
#>  [871] -1.386910854 -1.389298958 -1.391689452 -1.394082338 -1.396477618
#>  [876] -1.398875294 -1.401275369 -1.403677846 -1.406082726 -1.408490012
#>  [881] -1.410899706 -1.413311812 -1.415726331 -1.418143265 -1.420562618
#>  [886] -1.422984391 -1.425408588 -1.427835209 -1.430264259 -1.432695739
#>  [891] -1.435129651 -1.437565999 -1.440004784 -1.442446009 -1.444889677
#>  [896] -1.447335789 -1.449784349 -1.452235359 -1.454688821 -1.457144737
#>  [901] -1.459603111 -1.462063944 -1.464527240 -1.466993000 -1.469461227
#>  [906] -1.471931923 -1.474405091 -1.476880734 -1.479358854 -1.481839453
#>  [911] -1.484322533 -1.486808098 -1.489296150 -1.491786692 -1.494279725
#>  [916] -1.496775252 -1.499273276 -1.501773799 -1.504276824 -1.506782354
#>  [921] -1.509290390 -1.511800935 -1.514313993 -1.516829564 -1.519347653
#>  [926] -1.521868260 -1.524391390 -1.526917044 -1.529445225 -1.531975935
#>  [931] -1.534509178 -1.537044954 -1.539583268 -1.542124122 -1.544667517
#>  [936] -1.547213458 -1.549761945 -1.552312982 -1.554866572 -1.557422716
#>  [941] -1.559981418 -1.562542680 -1.565106505 -1.567672894 -1.570241851
#>  [946] -1.572813379 -1.575387479 -1.577964154 -1.580543408 -1.583125242
#>  [951] -1.585709659 -1.588296662 -1.590886254 -1.593478436 -1.596073211
#>  [956] -1.598670583 -1.601270553 -1.603873125 -1.606478300 -1.609086082
#>  [961] -1.611696473 -1.614309476 -1.616925093 -1.619543327 -1.622164181
#>  [966] -1.624787656 -1.627413757 -1.630042485 -1.632673843 -1.635307833
#>  [971] -1.637944459 -1.640583723 -1.643225628 -1.645870175 -1.648517369
#>  [976] -1.651167211 -1.653819704 -1.656474851 -1.659132655 -1.661793117
#>  [981] -1.664456242 -1.667122031 -1.669790487 -1.672461613 -1.675135411
#>  [986] -1.677811884 -1.680491036 -1.683172867 -1.685857382 -1.688544583
#>  [991] -1.691234472 -1.693927053 -1.696622327 -1.699320298 -1.702020969
#>  [996] -1.704724341 -1.707430418 -1.710139203 -1.712850698 -1.715564905
#> [1001] -1.718281828