Computes option price via Black-Scholes-Merton formula.

Black-Scholes-Merton formula implementation, pricing european options.

Usage

bsmprice(type, spot, strike, time, rate, yield, sigma)

Arguments

type

'call' for call option, any other value for put.

spot

current stock price.

strike

the strike price.

time

time to option expiration in years.

rate

the risk-free interest rate.

yield

the dividends that are expected to be paid.

sigma

volatility of the stock price.

Value

Vector with price of european options on stocks, according to Black-Scholes-Merton model.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

Examples

bsmprice(c("call", "put"), c(50, 49, 48, 47), 45, 0.25, 0.13, 0.01, 0.2)
#> [1] 6.5031266 0.2854164 4.7405597 0.6079959
bsmprice("call", 50, 45, 0.25, 0.13, 0.01, 1:1000 / 1000)
#>    [1]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>    [8]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [15]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [22]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [29]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [36]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [43]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [50]  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146  6.314146
#>   [57]  6.314146  6.314146  6.314146  6.314147  6.314147  6.314148  6.314149
#>   [64]  6.314150  6.314151  6.314153  6.314155  6.314158  6.314162  6.314167
#>   [71]  6.314173  6.314180  6.314188  6.314199  6.314212  6.314227  6.314244
#>   [78]  6.314265  6.314290  6.314318  6.314350  6.314388  6.314430  6.314479
#>   [85]  6.314534  6.314596  6.314665  6.314743  6.314829  6.314925  6.315031
#>   [92]  6.315148  6.315276  6.315417  6.315570  6.315737  6.315919  6.316116
#>   [99]  6.316329  6.316558  6.316805  6.317070  6.317354  6.317658  6.317983
#>  [106]  6.318328  6.318695  6.319085  6.319499  6.319936  6.320399  6.320886
#>  [113]  6.321400  6.321940  6.322508  6.323104  6.323729  6.324383  6.325067
#>  [120]  6.325781  6.326526  6.327302  6.328110  6.328951  6.329825  6.330732
#>  [127]  6.331674  6.332649  6.333659  6.334704  6.335785  6.336901  6.338053
#>  [134]  6.339242  6.340468  6.341731  6.343031  6.344369  6.345744  6.347158
#>  [141]  6.348609  6.350100  6.351629  6.353196  6.354803  6.356448  6.358133
#>  [148]  6.359857  6.361621  6.363424  6.365267  6.367149  6.369071  6.371032
#>  [155]  6.373034  6.375075  6.377156  6.379277  6.381437  6.383637  6.385877
#>  [162]  6.388157  6.390476  6.392834  6.395233  6.397670  6.400147  6.402663
#>  [169]  6.405219  6.407813  6.410447  6.413119  6.415830  6.418580  6.421368
#>  [176]  6.424194  6.427059  6.429962  6.432902  6.435881  6.438897  6.441951
#>  [183]  6.445042  6.448170  6.451335  6.454537  6.457775  6.461050  6.464361
#>  [190]  6.467709  6.471092  6.474511  6.477966  6.481455  6.484981  6.488541
#>  [197]  6.492135  6.495765  6.499429  6.503127  6.506859  6.510624  6.514424
#>  [204]  6.518256  6.522122  6.526021  6.529953  6.533917  6.537914  6.541943
#>  [211]  6.546003  6.550096  6.554220  6.558375  6.562562  6.566779  6.571027
#>  [218]  6.575306  6.579615  6.583954  6.588323  6.592722  6.597150  6.601608
#>  [225]  6.606094  6.610610  6.615154  6.619727  6.624328  6.628957  6.633614
#>  [232]  6.638299  6.643011  6.647750  6.652517  6.657311  6.662131  6.666978
#>  [239]  6.671851  6.676750  6.681676  6.686627  6.691604  6.696606  6.701633
#>  [246]  6.706685  6.711763  6.716865  6.721991  6.727142  6.732317  6.737515
#>  [253]  6.742738  6.747984  6.753254  6.758547  6.763863  6.769202  6.774563
#>  [260]  6.779948  6.785354  6.790783  6.796234  6.801707  6.807202  6.812718
#>  [267]  6.818256  6.823814  6.829395  6.834996  6.840617  6.846260  6.851923
#>  [274]  6.857606  6.863309  6.869033  6.874776  6.880539  6.886322  6.892124
#>  [281]  6.897946  6.903786  6.909646  6.915525  6.921422  6.927338  6.933272
#>  [288]  6.939225  6.945195  6.951184  6.957191  6.963216  6.969258  6.975317
#>  [295]  6.981395  6.987489  6.993600  6.999729  7.005874  7.012037  7.018215
#>  [302]  7.024411  7.030622  7.036850  7.043094  7.049355  7.055631  7.061922
#>  [309]  7.068230  7.074553  7.080892  7.087245  7.093615  7.099999  7.106398
#>  [316]  7.112812  7.119241  7.125685  7.132143  7.138616  7.145103  7.151604
#>  [323]  7.158119  7.164649  7.171192  7.177749  7.184320  7.190905  7.197503
#>  [330]  7.204115  7.210740  7.217378  7.224030  7.230695  7.237372  7.244063
#>  [337]  7.250766  7.257482  7.264210  7.270951  7.277705  7.284471  7.291249
#>  [344]  7.298039  7.304842  7.311656  7.318482  7.325321  7.332171  7.339032
#>  [351]  7.345905  7.352790  7.359686  7.366594  7.373512  7.380442  7.387383
#>  [358]  7.394335  7.401298  7.408272  7.415257  7.422253  7.429259  7.436275
#>  [365]  7.443303  7.450340  7.457388  7.464447  7.471515  7.478594  7.485683
#>  [372]  7.492781  7.499890  7.507009  7.514137  7.521275  7.528423  7.535581
#>  [379]  7.542748  7.549924  7.557110  7.564305  7.571510  7.578724  7.585947
#>  [386]  7.593179  7.600420  7.607670  7.614929  7.622197  7.629474  7.636759
#>  [393]  7.644053  7.651356  7.658668  7.665988  7.673316  7.680653  7.687998
#>  [400]  7.695351  7.702713  7.710083  7.717461  7.724847  7.732241  7.739643
#>  [407]  7.747053  7.754471  7.761896  7.769330  7.776771  7.784220  7.791676
#>  [414]  7.799140  7.806612  7.814091  7.821577  7.829071  7.836572  7.844081
#>  [421]  7.851596  7.859119  7.866649  7.874186  7.881730  7.889281  7.896839
#>  [428]  7.904404  7.911976  7.919555  7.927141  7.934733  7.942332  7.949937
#>  [435]  7.957549  7.965168  7.972794  7.980425  7.988064  7.995708  8.003359
#>  [442]  8.011017  8.018680  8.026350  8.034026  8.041709  8.049397  8.057092
#>  [449]  8.064792  8.072499  8.080211  8.087930  8.095654  8.103385  8.111121
#>  [456]  8.118863  8.126610  8.134364  8.142123  8.149888  8.157658  8.165434
#>  [463]  8.173216  8.181003  8.188796  8.196594  8.204397  8.212206  8.220020
#>  [470]  8.227840  8.235665  8.243495  8.251330  8.259171  8.267016  8.274867
#>  [477]  8.282723  8.290584  8.298450  8.306321  8.314197  8.322078  8.329964
#>  [484]  8.337855  8.345750  8.353651  8.361556  8.369466  8.377381  8.385301
#>  [491]  8.393225  8.401154  8.409087  8.417025  8.424968  8.432915  8.440867
#>  [498]  8.448824  8.456784  8.464750  8.472719  8.480693  8.488672  8.496654
#>  [505]  8.504642  8.512633  8.520629  8.528628  8.536633  8.544641  8.552653
#>  [512]  8.560670  8.568691  8.576716  8.584745  8.592778  8.600815  8.608856
#>  [519]  8.616901  8.624949  8.633002  8.641059  8.649120  8.657184  8.665253
#>  [526]  8.673325  8.681401  8.689481  8.697564  8.705652  8.713743  8.721837
#>  [533]  8.729936  8.738038  8.746143  8.754253  8.762365  8.770482  8.778602
#>  [540]  8.786725  8.794852  8.802983  8.811117  8.819254  8.827395  8.835539
#>  [547]  8.843687  8.851838  8.859992  8.868150  8.876311  8.884475  8.892643
#>  [554]  8.900814  8.908988  8.917165  8.925345  8.933529  8.941716  8.949906
#>  [561]  8.958099  8.966295  8.974495  8.982697  8.990903  8.999111  9.007323
#>  [568]  9.015537  9.023755  9.031975  9.040199  9.048425  9.056655  9.064887
#>  [575]  9.073122  9.081360  9.089601  9.097845  9.106091  9.114341  9.122593
#>  [582]  9.130848  9.139106  9.147366  9.155629  9.163895  9.172164  9.180436
#>  [589]  9.188710  9.196986  9.205266  9.213548  9.221832  9.230120  9.238409
#>  [596]  9.246702  9.254997  9.263294  9.271594  9.279897  9.288202  9.296509
#>  [603]  9.304820  9.313132  9.321447  9.329764  9.338084  9.346406  9.354731
#>  [610]  9.363058  9.371387  9.379719  9.388053  9.396390  9.404728  9.413070
#>  [617]  9.421413  9.429758  9.438106  9.446457  9.454809  9.463164  9.471520
#>  [624]  9.479879  9.488241  9.496604  9.504970  9.513338  9.521707  9.530079
#>  [631]  9.538454  9.546830  9.555208  9.563589  9.571971  9.580356  9.588743
#>  [638]  9.597131  9.605522  9.613915  9.622310  9.630707  9.639105  9.647506
#>  [645]  9.655909  9.664313  9.672720  9.681129  9.689539  9.697952  9.706366
#>  [652]  9.714782  9.723200  9.731620  9.740042  9.748466  9.756891  9.765318
#>  [659]  9.773748  9.782179  9.790611  9.799046  9.807482  9.815920  9.824360
#>  [666]  9.832802  9.841245  9.849690  9.858137  9.866586  9.875036  9.883488
#>  [673]  9.891942  9.900397  9.908854  9.917313  9.925773  9.934235  9.942698
#>  [680]  9.951164  9.959630  9.968099  9.976569  9.985040  9.993514 10.001988
#>  [687] 10.010465 10.018942 10.027422 10.035903 10.044385 10.052869 10.061355
#>  [694] 10.069842 10.078330 10.086820 10.095311 10.103804 10.112299 10.120795
#>  [701] 10.129292 10.137791 10.146291 10.154792 10.163295 10.171799 10.180305
#>  [708] 10.188812 10.197321 10.205831 10.214342 10.222855 10.231369 10.239884
#>  [715] 10.248401 10.256919 10.265438 10.273958 10.282480 10.291004 10.299528
#>  [722] 10.308054 10.316581 10.325109 10.333639 10.342170 10.350702 10.359235
#>  [729] 10.367770 10.376305 10.384842 10.393381 10.401920 10.410461 10.419002
#>  [736] 10.427545 10.436090 10.444635 10.453182 10.461729 10.470278 10.478828
#>  [743] 10.487379 10.495931 10.504485 10.513039 10.521595 10.530151 10.538709
#>  [750] 10.547268 10.555828 10.564389 10.572951 10.581514 10.590079 10.598644
#>  [757] 10.607210 10.615778 10.624346 10.632915 10.641486 10.650057 10.658630
#>  [764] 10.667203 10.675778 10.684353 10.692930 10.701507 10.710086 10.718665
#>  [771] 10.727246 10.735827 10.744409 10.752993 10.761577 10.770162 10.778748
#>  [778] 10.787335 10.795923 10.804512 10.813101 10.821692 10.830283 10.838876
#>  [785] 10.847469 10.856063 10.864658 10.873254 10.881851 10.890449 10.899047
#>  [792] 10.907647 10.916247 10.924848 10.933450 10.942052 10.950656 10.959260
#>  [799] 10.967865 10.976471 10.985078 10.993685 11.002294 11.010903 11.019513
#>  [806] 11.028123 11.036735 11.045347 11.053960 11.062574 11.071188 11.079804
#>  [813] 11.088419 11.097036 11.105654 11.114272 11.122891 11.131510 11.140131
#>  [820] 11.148752 11.157373 11.165996 11.174619 11.183243 11.191867 11.200492
#>  [827] 11.209118 11.217745 11.226372 11.235000 11.243628 11.252258 11.260888
#>  [834] 11.269518 11.278149 11.286781 11.295413 11.304046 11.312680 11.321314
#>  [841] 11.329949 11.338584 11.347221 11.355857 11.364494 11.373132 11.381771
#>  [848] 11.390410 11.399050 11.407690 11.416330 11.424972 11.433614 11.442256
#>  [855] 11.450899 11.459543 11.468187 11.476832 11.485477 11.494123 11.502769
#>  [862] 11.511416 11.520063 11.528711 11.537359 11.546008 11.554657 11.563307
#>  [869] 11.571958 11.580609 11.589260 11.597912 11.606564 11.615217 11.623870
#>  [876] 11.632524 11.641178 11.649833 11.658488 11.667144 11.675800 11.684456
#>  [883] 11.693113 11.701771 11.710429 11.719087 11.727746 11.736405 11.745064
#>  [890] 11.753724 11.762385 11.771046 11.779707 11.788369 11.797031 11.805693
#>  [897] 11.814356 11.823019 11.831683 11.840347 11.849012 11.857676 11.866342
#>  [904] 11.875007 11.883673 11.892339 11.901006 11.909673 11.918340 11.927008
#>  [911] 11.935676 11.944345 11.953013 11.961682 11.970352 11.979022 11.987692
#>  [918] 11.996362 12.005033 12.013704 12.022375 12.031047 12.039719 12.048392
#>  [925] 12.057064 12.065737 12.074410 12.083084 12.091758 12.100432 12.109106
#>  [932] 12.117781 12.126456 12.135131 12.143807 12.152482 12.161159 12.169835
#>  [939] 12.178511 12.187188 12.195865 12.204543 12.213220 12.221898 12.230576
#>  [946] 12.239255 12.247933 12.256612 12.265291 12.273970 12.282650 12.291329
#>  [953] 12.300009 12.308690 12.317370 12.326050 12.334731 12.343412 12.352093
#>  [960] 12.360775 12.369456 12.378138 12.386820 12.395502 12.404185 12.412867
#>  [967] 12.421550 12.430233 12.438916 12.447599 12.456283 12.464966 12.473650
#>  [974] 12.482334 12.491018 12.499702 12.508387 12.517071 12.525756 12.534441
#>  [981] 12.543126 12.551811 12.560496 12.569182 12.577867 12.586553 12.595239
#>  [988] 12.603925 12.612611 12.621297 12.629983 12.638670 12.647356 12.656043
#>  [995] 12.664730 12.673416 12.682103 12.690790 12.699478 12.708165
bsmprice("put", 20, 30, 0:2000 / 1000, 0.15, 0, 0.25)
#>    [1] 10.000000  9.995500  9.991001  9.986503  9.982005  9.977508  9.973012
#>    [8]  9.968517  9.964022  9.959527  9.955034  9.950541  9.946049  9.941557
#>   [15]  9.937066  9.932576  9.928086  9.923597  9.919109  9.914622  9.910135
#>   [22]  9.905649  9.901163  9.896678  9.892194  9.887711  9.883228  9.878746
#>   [29]  9.874264  9.869783  9.865303  9.860824  9.856345  9.851867  9.847389
#>   [36]  9.842913  9.838437  9.833961  9.829486  9.825012  9.820539  9.816066
#>   [43]  9.811594  9.807123  9.802652  9.798182  9.793713  9.789244  9.784776
#>   [50]  9.780308  9.775842  9.771376  9.766910  9.762446  9.757981  9.753518
#>   [57]  9.749055  9.744593  9.740132  9.735671  9.731211  9.726752  9.722293
#>   [64]  9.717835  9.713378  9.708921  9.704465  9.700010  9.695555  9.691101
#>   [71]  9.686648  9.682195  9.677743  9.673292  9.668841  9.664391  9.659942
#>   [78]  9.655493  9.651045  9.646598  9.642151  9.637705  9.633260  9.628815
#>   [85]  9.624371  9.619928  9.615485  9.611043  9.606602  9.602162  9.597722
#>   [92]  9.593282  9.588844  9.584406  9.579968  9.575532  9.571096  9.566660
#>   [99]  9.562226  9.557792  9.553358  9.548926  9.544494  9.540062  9.535632
#>  [106]  9.531202  9.526772  9.522344  9.517916  9.513489  9.509062  9.504636
#>  [113]  9.500211  9.495786  9.491362  9.486939  9.482516  9.478094  9.473673
#>  [120]  9.469253  9.464833  9.460413  9.455995  9.451577  9.447160  9.442743
#>  [127]  9.438328  9.433912  9.429498  9.425084  9.420671  9.416259  9.411847
#>  [134]  9.407436  9.403026  9.398616  9.394207  9.389799  9.385392  9.380985
#>  [141]  9.376579  9.372173  9.367769  9.363365  9.358962  9.354559  9.350157
#>  [148]  9.345756  9.341356  9.336956  9.332558  9.328159  9.323762  9.319365
#>  [155]  9.314969  9.310574  9.306180  9.301786  9.297393  9.293001  9.288610
#>  [162]  9.284219  9.279830  9.275441  9.271052  9.266665  9.262278  9.257892
#>  [169]  9.253507  9.249123  9.244740  9.240357  9.235975  9.231594  9.227214
#>  [176]  9.222835  9.218456  9.214079  9.209702  9.205326  9.200951  9.196577
#>  [183]  9.192204  9.187831  9.183460  9.179089  9.174719  9.170350  9.165983
#>  [190]  9.161616  9.157249  9.152884  9.148520  9.144157  9.139794  9.135433
#>  [197]  9.131072  9.126713  9.122354  9.117997  9.113640  9.109285  9.104930
#>  [204]  9.100577  9.096224  9.091873  9.087522  9.083173  9.078824  9.074477
#>  [211]  9.070131  9.065785  9.061441  9.057098  9.052756  9.048415  9.044075
#>  [218]  9.039737  9.035399  9.031062  9.026727  9.022393  9.018060  9.013728
#>  [225]  9.009397  9.005067  9.000739  8.996412  8.992086  8.987761  8.983437
#>  [232]  8.979115  8.974793  8.970473  8.966155  8.961837  8.957521  8.953206
#>  [239]  8.948892  8.944580  8.940269  8.935959  8.931650  8.927343  8.923037
#>  [246]  8.918732  8.914429  8.910127  8.905826  8.901527  8.897229  8.892933
#>  [253]  8.888637  8.884344  8.880051  8.875760  8.871471  8.867183  8.862896
#>  [260]  8.858611  8.854327  8.850044  8.845763  8.841484  8.837206  8.832930
#>  [267]  8.828654  8.824381  8.820109  8.815838  8.811569  8.807302  8.803036
#>  [274]  8.798772  8.794509  8.790247  8.785988  8.781730  8.777473  8.773218
#>  [281]  8.768965  8.764713  8.760463  8.756214  8.751967  8.747722  8.743478
#>  [288]  8.739236  8.734996  8.730757  8.726520  8.722285  8.718051  8.713819
#>  [295]  8.709589  8.705360  8.701133  8.696908  8.692685  8.688463  8.684243
#>  [302]  8.680025  8.675809  8.671594  8.667381  8.663170  8.658960  8.654753
#>  [309]  8.650547  8.646343  8.642141  8.637941  8.633742  8.629545  8.625350
#>  [316]  8.621157  8.616966  8.612777  8.608589  8.604404  8.600220  8.596038
#>  [323]  8.591858  8.587680  8.583504  8.579329  8.575157  8.570986  8.566818
#>  [330]  8.562651  8.558486  8.554324  8.550163  8.546004  8.541847  8.537692
#>  [337]  8.533539  8.529388  8.525239  8.521092  8.516947  8.512803  8.508662
#>  [344]  8.504523  8.500386  8.496251  8.492118  8.487987  8.483858  8.479731
#>  [351]  8.475606  8.471483  8.467362  8.463243  8.459127  8.455012  8.450899
#>  [358]  8.446789  8.442680  8.438574  8.434470  8.430368  8.426268  8.422170
#>  [365]  8.418074  8.413980  8.409888  8.405799  8.401711  8.397626  8.393543
#>  [372]  8.389462  8.385383  8.381306  8.377232  8.373159  8.369089  8.365021
#>  [379]  8.360955  8.356891  8.352829  8.348770  8.344713  8.340658  8.336605
#>  [386]  8.332554  8.328506  8.324459  8.320415  8.316373  8.312334  8.308296
#>  [393]  8.304261  8.300228  8.296197  8.292169  8.288142  8.284118  8.280096
#>  [400]  8.276077  8.272059  8.268044  8.264031  8.260021  8.256012  8.252006
#>  [407]  8.248003  8.244001  8.240002  8.236005  8.232010  8.228017  8.224027
#>  [414]  8.220039  8.216054  8.212070  8.208089  8.204110  8.200134  8.196160
#>  [421]  8.192188  8.188218  8.184251  8.180286  8.176323  8.172363  8.168405
#>  [428]  8.164449  8.160496  8.156545  8.152596  8.148649  8.144705  8.140764
#>  [435]  8.136824  8.132887  8.128952  8.125020  8.121090  8.117162  8.113237
#>  [442]  8.109314  8.105393  8.101474  8.097558  8.093645  8.089733  8.085825
#>  [449]  8.081918  8.078014  8.074112  8.070212  8.066315  8.062420  8.058528
#>  [456]  8.054638  8.050750  8.046865  8.042982  8.039101  8.035223  8.031347
#>  [463]  8.027474  8.023603  8.019734  8.015868  8.012004  8.008143  8.004283
#>  [470]  8.000427  7.996572  7.992720  7.988871  7.985024  7.981179  7.977336
#>  [477]  7.973496  7.969659  7.965823  7.961991  7.958160  7.954332  7.950506
#>  [484]  7.946683  7.942862  7.939044  7.935228  7.931414  7.927603  7.923794
#>  [491]  7.919988  7.916184  7.912382  7.908583  7.904786  7.900992  7.897200
#>  [498]  7.893410  7.889623  7.885838  7.882056  7.878276  7.874498  7.870723
#>  [505]  7.866951  7.863180  7.859412  7.855647  7.851884  7.848123  7.844365
#>  [512]  7.840609  7.836856  7.833105  7.829356  7.825610  7.821866  7.818125
#>  [519]  7.814386  7.810649  7.806915  7.803183  7.799454  7.795727  7.792003
#>  [526]  7.788281  7.784561  7.780844  7.777129  7.773417  7.769707  7.765999
#>  [533]  7.762294  7.758591  7.754891  7.751193  7.747497  7.743804  7.740113
#>  [540]  7.736425  7.732739  7.729056  7.725375  7.721696  7.718020  7.714346
#>  [547]  7.710675  7.707006  7.703339  7.699675  7.696013  7.692354  7.688697
#>  [554]  7.685042  7.681390  7.677740  7.674093  7.670448  7.666805  7.663165
#>  [561]  7.659527  7.655892  7.652259  7.648629  7.645000  7.641375  7.637751
#>  [568]  7.634130  7.630512  7.626896  7.623282  7.619670  7.616061  7.612455
#>  [575]  7.608851  7.605249  7.601649  7.598052  7.594458  7.590866  7.587276
#>  [582]  7.583688  7.580103  7.576520  7.572940  7.569362  7.565787  7.562213
#>  [589]  7.558643  7.555074  7.551508  7.547945  7.544383  7.540824  7.537268
#>  [596]  7.533714  7.530162  7.526612  7.523065  7.519521  7.515978  7.512438
#>  [603]  7.508901  7.505366  7.501833  7.498302  7.494774  7.491249  7.487725
#>  [610]  7.484204  7.480685  7.477169  7.473655  7.470144  7.466634  7.463127
#>  [617]  7.459623  7.456121  7.452621  7.449123  7.445628  7.442135  7.438645
#>  [624]  7.435157  7.431671  7.428188  7.424707  7.421228  7.417751  7.414277
#>  [631]  7.410806  7.407336  7.403869  7.400404  7.396942  7.393482  7.390024
#>  [638]  7.386569  7.383115  7.379665  7.376216  7.372770  7.369326  7.365885
#>  [645]  7.362446  7.359009  7.355574  7.352142  7.348712  7.345284  7.341859
#>  [652]  7.338436  7.335015  7.331597  7.328181  7.324767  7.321355  7.317946
#>  [659]  7.314539  7.311135  7.307732  7.304332  7.300934  7.297539  7.294146
#>  [666]  7.290755  7.287366  7.283980  7.280596  7.277214  7.273835  7.270458
#>  [673]  7.267083  7.263710  7.260340  7.256972  7.253606  7.250242  7.246881
#>  [680]  7.243522  7.240165  7.236811  7.233458  7.230108  7.226761  7.223415
#>  [687]  7.220072  7.216731  7.213392  7.210056  7.206722  7.203390  7.200060
#>  [694]  7.196733  7.193407  7.190084  7.186764  7.183445  7.180129  7.176815
#>  [701]  7.173503  7.170193  7.166886  7.163581  7.160278  7.156977  7.153679
#>  [708]  7.150383  7.147089  7.143797  7.140507  7.137220  7.133935  7.130652
#>  [715]  7.127371  7.124093  7.120816  7.117542  7.114270  7.111001  7.107733
#>  [722]  7.104468  7.101205  7.097944  7.094685  7.091429  7.088174  7.084922
#>  [729]  7.081672  7.078424  7.075179  7.071935  7.068694  7.065455  7.062218
#>  [736]  7.058983  7.055751  7.052521  7.049292  7.046066  7.042842  7.039621
#>  [743]  7.036401  7.033184  7.029969  7.026756  7.023545  7.020336  7.017129
#>  [750]  7.013925  7.010723  7.007522  7.004324  7.001129  6.997935  6.994743
#>  [757]  6.991554  6.988367  6.985181  6.981998  6.978817  6.975639  6.972462
#>  [764]  6.969287  6.966115  6.962945  6.959777  6.956611  6.953447  6.950285
#>  [771]  6.947125  6.943968  6.940812  6.937659  6.934508  6.931358  6.928211
#>  [778]  6.925066  6.921924  6.918783  6.915644  6.912508  6.909373  6.906241
#>  [785]  6.903111  6.899982  6.896856  6.893732  6.890610  6.887491  6.884373
#>  [792]  6.881257  6.878143  6.875032  6.871922  6.868815  6.865710  6.862606
#>  [799]  6.859505  6.856406  6.853309  6.850214  6.847121  6.844030  6.840941
#>  [806]  6.837854  6.834770  6.831687  6.828606  6.825528  6.822451  6.819376
#>  [813]  6.816304  6.813234  6.810165  6.807099  6.804034  6.800972  6.797912
#>  [820]  6.794853  6.791797  6.788743  6.785691  6.782641  6.779592  6.776546
#>  [827]  6.773502  6.770460  6.767420  6.764382  6.761346  6.758312  6.755280
#>  [834]  6.752250  6.749222  6.746195  6.743171  6.740149  6.737129  6.734111
#>  [841]  6.731095  6.728081  6.725069  6.722059  6.719050  6.716044  6.713040
#>  [848]  6.710038  6.707038  6.704039  6.701043  6.698049  6.695056  6.692066
#>  [855]  6.689078  6.686091  6.683107  6.680124  6.677143  6.674165  6.671188
#>  [862]  6.668213  6.665241  6.662270  6.659301  6.656334  6.653369  6.650406
#>  [869]  6.647445  6.644486  6.641528  6.638573  6.635620  6.632668  6.629719
#>  [876]  6.626771  6.623826  6.620882  6.617940  6.615000  6.612062  6.609126
#>  [883]  6.606192  6.603260  6.600330  6.597401  6.594475  6.591550  6.588628
#>  [890]  6.585707  6.582788  6.579871  6.576956  6.574043  6.571132  6.568222
#>  [897]  6.565315  6.562409  6.559506  6.556604  6.553704  6.550806  6.547910
#>  [904]  6.545015  6.542123  6.539232  6.536344  6.533457  6.530572  6.527689
#>  [911]  6.524808  6.521928  6.519051  6.516175  6.513302  6.510430  6.507560
#>  [918]  6.504692  6.501825  6.498961  6.496098  6.493238  6.490379  6.487522
#>  [925]  6.484666  6.481813  6.478962  6.476112  6.473264  6.470418  6.467574
#>  [932]  6.464731  6.461891  6.459052  6.456215  6.453380  6.450547  6.447716
#>  [939]  6.444886  6.442058  6.439232  6.436408  6.433586  6.430765  6.427947
#>  [946]  6.425130  6.422315  6.419501  6.416690  6.413880  6.411072  6.408266
#>  [953]  6.405462  6.402659  6.399859  6.397060  6.394263  6.391467  6.388674
#>  [960]  6.385882  6.383092  6.380304  6.377518  6.374733  6.371950  6.369169
#>  [967]  6.366390  6.363612  6.360836  6.358062  6.355290  6.352520  6.349751
#>  [974]  6.346984  6.344219  6.341455  6.338694  6.335934  6.333176  6.330419
#>  [981]  6.327664  6.324912  6.322160  6.319411  6.316663  6.313917  6.311173
#>  [988]  6.308431  6.305690  6.302951  6.300214  6.297478  6.294744  6.292012
#>  [995]  6.289282  6.286553  6.283826  6.281101  6.278378  6.275656  6.272936
#> [1002]  6.270218  6.267501  6.264786  6.262073  6.259362  6.256652  6.253944
#> [1009]  6.251237  6.248533  6.245830  6.243129  6.240429  6.237731  6.235035
#> [1016]  6.232341  6.229648  6.226957  6.224268  6.221580  6.218894  6.216210
#> [1023]  6.213527  6.210846  6.208167  6.205489  6.202813  6.200139  6.197466
#> [1030]  6.194795  6.192126  6.189459  6.186793  6.184128  6.181466  6.178805
#> [1037]  6.176146  6.173488  6.170832  6.168178  6.165525  6.162874  6.160225
#> [1044]  6.157577  6.154931  6.152287  6.149644  6.147003  6.144364  6.141726
#> [1051]  6.139090  6.136455  6.133822  6.131191  6.128561  6.125933  6.123307
#> [1058]  6.120682  6.118059  6.115438  6.112818  6.110200  6.107583  6.104968
#> [1065]  6.102355  6.099743  6.097133  6.094524  6.091917  6.089312  6.086708
#> [1072]  6.084106  6.081506  6.078907  6.076309  6.073714  6.071120  6.068527
#> [1079]  6.065936  6.063347  6.060759  6.058173  6.055589  6.053006  6.050424
#> [1086]  6.047845  6.045266  6.042690  6.040115  6.037541  6.034969  6.032399
#> [1093]  6.029831  6.027263  6.024698  6.022134  6.019571  6.017011  6.014451
#> [1100]  6.011894  6.009337  6.006783  6.004230  6.001678  5.999129  5.996580
#> [1107]  5.994033  5.991488  5.988944  5.986402  5.983862  5.981323  5.978785
#> [1114]  5.976249  5.973715  5.971182  5.968651  5.966121  5.963593  5.961066
#> [1121]  5.958541  5.956017  5.953495  5.950975  5.948456  5.945938  5.943422
#> [1128]  5.940908  5.938395  5.935883  5.933373  5.930865  5.928358  5.925853
#> [1135]  5.923349  5.920847  5.918346  5.915847  5.913349  5.910853  5.908358
#> [1142]  5.905865  5.903373  5.900883  5.898394  5.895907  5.893421  5.890937
#> [1149]  5.888454  5.885973  5.883493  5.881015  5.878538  5.876063  5.873589
#> [1156]  5.871117  5.868646  5.866177  5.863709  5.861242  5.858777  5.856314
#> [1163]  5.853852  5.851392  5.848933  5.846475  5.844019  5.841565  5.839111
#> [1170]  5.836660  5.834210  5.831761  5.829314  5.826868  5.824424  5.821981
#> [1177]  5.819539  5.817099  5.814661  5.812224  5.809788  5.807354  5.804922
#> [1184]  5.802490  5.800061  5.797632  5.795205  5.792780  5.790356  5.787933
#> [1191]  5.785512  5.783093  5.780674  5.778258  5.775842  5.773428  5.771016
#> [1198]  5.768605  5.766195  5.763787  5.761380  5.758975  5.756571  5.754168
#> [1205]  5.751767  5.749367  5.746969  5.744572  5.742177  5.739783  5.737390
#> [1212]  5.734999  5.732609  5.730221  5.727834  5.725449  5.723064  5.720682
#> [1219]  5.718300  5.715920  5.713542  5.711165  5.708789  5.706415  5.704042
#> [1226]  5.701670  5.699300  5.696931  5.694564  5.692198  5.689833  5.687470
#> [1233]  5.685108  5.682748  5.680389  5.678031  5.675675  5.673320  5.670967
#> [1240]  5.668614  5.666264  5.663914  5.661566  5.659220  5.656874  5.654530
#> [1247]  5.652188  5.649847  5.647507  5.645168  5.642831  5.640496  5.638161
#> [1254]  5.635828  5.633497  5.631166  5.628838  5.626510  5.624184  5.621859
#> [1261]  5.619535  5.617213  5.614892  5.612573  5.610255  5.607938  5.605623
#> [1268]  5.603309  5.600996  5.598684  5.596374  5.594066  5.591758  5.589452
#> [1275]  5.587148  5.584844  5.582542  5.580241  5.577942  5.575644  5.573347
#> [1282]  5.571052  5.568758  5.566465  5.564174  5.561884  5.559595  5.557307
#> [1289]  5.555021  5.552736  5.550453  5.548171  5.545890  5.543610  5.541332
#> [1296]  5.539055  5.536779  5.534505  5.532232  5.529960  5.527690  5.525421
#> [1303]  5.523153  5.520886  5.518621  5.516357  5.514095  5.511833  5.509573
#> [1310]  5.507315  5.505057  5.502801  5.500546  5.498293  5.496040  5.493789
#> [1317]  5.491540  5.489291  5.487044  5.484798  5.482554  5.480310  5.478068
#> [1324]  5.475828  5.473588  5.471350  5.469113  5.466877  5.464643  5.462410
#> [1331]  5.460178  5.457948  5.455718  5.453490  5.451264  5.449038  5.446814
#> [1338]  5.444591  5.442369  5.440149  5.437930  5.435712  5.433495  5.431280
#> [1345]  5.429065  5.426853  5.424641  5.422431  5.420221  5.418014  5.415807
#> [1352]  5.413601  5.411397  5.409194  5.406993  5.404792  5.402593  5.400395
#> [1359]  5.398199  5.396003  5.393809  5.391616  5.389424  5.387234  5.385044
#> [1366]  5.382856  5.380670  5.378484  5.376300  5.374116  5.371935  5.369754
#> [1373]  5.367574  5.365396  5.363219  5.361043  5.358869  5.356696  5.354523
#> [1380]  5.352352  5.350183  5.348014  5.345847  5.343681  5.341516  5.339353
#> [1387]  5.337190  5.335029  5.332869  5.330710  5.328553  5.326396  5.324241
#> [1394]  5.322087  5.319934  5.317783  5.315632  5.313483  5.311335  5.309188
#> [1401]  5.307043  5.304898  5.302755  5.300613  5.298472  5.296333  5.294194
#> [1408]  5.292057  5.289921  5.287786  5.285653  5.283520  5.281389  5.279259
#> [1415]  5.277130  5.275002  5.272875  5.270750  5.268626  5.266503  5.264381
#> [1422]  5.262260  5.260141  5.258022  5.255905  5.253789  5.251674  5.249561
#> [1429]  5.247448  5.245337  5.243226  5.241117  5.239010  5.236903  5.234797
#> [1436]  5.232693  5.230590  5.228488  5.226387  5.224287  5.222189  5.220091
#> [1443]  5.217995  5.215900  5.213806  5.211713  5.209621  5.207531  5.205441
#> [1450]  5.203353  5.201266  5.199180  5.197095  5.195012  5.192929  5.190848
#> [1457]  5.188768  5.186689  5.184611  5.182534  5.180458  5.178384  5.176310
#> [1464]  5.174238  5.172167  5.170097  5.168028  5.165960  5.163894  5.161828
#> [1471]  5.159764  5.157701  5.155639  5.153578  5.151518  5.149459  5.147401
#> [1478]  5.145345  5.143290  5.141235  5.139182  5.137130  5.135079  5.133029
#> [1485]  5.130981  5.128933  5.126887  5.124841  5.122797  5.120754  5.118712
#> [1492]  5.116671  5.114631  5.112593  5.110555  5.108519  5.106483  5.104449
#> [1499]  5.102416  5.100384  5.098353  5.096323  5.094294  5.092266  5.090240
#> [1506]  5.088214  5.086190  5.084167  5.082144  5.080123  5.078103  5.076084
#> [1513]  5.074066  5.072050  5.070034  5.068019  5.066006  5.063994  5.061982
#> [1520]  5.059972  5.057963  5.055955  5.053948  5.051942  5.049937  5.047933
#> [1527]  5.045930  5.043929  5.041928  5.039929  5.037930  5.035933  5.033937
#> [1534]  5.031941  5.029947  5.027954  5.025962  5.023971  5.021981  5.019993
#> [1541]  5.018005  5.016018  5.014033  5.012048  5.010065  5.008082  5.006101
#> [1548]  5.004120  5.002141  5.000163  4.998186  4.996210  4.994235  4.992261
#> [1555]  4.990288  4.988316  4.986345  4.984375  4.982407  4.980439  4.978472
#> [1562]  4.976507  4.974542  4.972579  4.970616  4.968655  4.966694  4.964735
#> [1569]  4.962777  4.960819  4.958863  4.956908  4.954954  4.953001  4.951049
#> [1576]  4.949098  4.947148  4.945199  4.943251  4.941304  4.939358  4.937413
#> [1583]  4.935469  4.933527  4.931585  4.929644  4.927704  4.925766  4.923828
#> [1590]  4.921891  4.919956  4.918021  4.916087  4.914155  4.912223  4.910293
#> [1597]  4.908363  4.906435  4.904507  4.902581  4.900655  4.898731  4.896807
#> [1604]  4.894885  4.892964  4.891043  4.889124  4.887206  4.885288  4.883372
#> [1611]  4.881456  4.879542  4.877629  4.875716  4.873805  4.871895  4.869985
#> [1618]  4.868077  4.866170  4.864263  4.862358  4.860454  4.858550  4.856648
#> [1625]  4.854746  4.852846  4.850947  4.849048  4.847151  4.845255  4.843359
#> [1632]  4.841465  4.839571  4.837679  4.835787  4.833897  4.832007  4.830119
#> [1639]  4.828231  4.826345  4.824459  4.822575  4.820691  4.818808  4.816927
#> [1646]  4.815046  4.813166  4.811288  4.809410  4.807533  4.805657  4.803783
#> [1653]  4.801909  4.800036  4.798164  4.796293  4.794423  4.792554  4.790686
#> [1660]  4.788819  4.786953  4.785088  4.783224  4.781361  4.779498  4.777637
#> [1667]  4.775777  4.773917  4.772059  4.770202  4.768345  4.766490  4.764635
#> [1674]  4.762782  4.760929  4.759077  4.757227  4.755377  4.753528  4.751680
#> [1681]  4.749833  4.747987  4.746142  4.744298  4.742455  4.740613  4.738772
#> [1688]  4.736932  4.735092  4.733254  4.731417  4.729580  4.727744  4.725910
#> [1695]  4.724076  4.722244  4.720412  4.718581  4.716751  4.714922  4.713094
#> [1702]  4.711267  4.709441  4.707616  4.705791  4.703968  4.702146  4.700324
#> [1709]  4.698504  4.696684  4.694865  4.693048  4.691231  4.689415  4.687600
#> [1716]  4.685786  4.683973  4.682161  4.680349  4.678539  4.676730  4.674921
#> [1723]  4.673113  4.671307  4.669501  4.667696  4.665892  4.664089  4.662287
#> [1730]  4.660486  4.658686  4.656887  4.655088  4.653291  4.651494  4.649699
#> [1737]  4.647904  4.646110  4.644317  4.642525  4.640734  4.638944  4.637154
#> [1744]  4.635366  4.633579  4.631792  4.630006  4.628222  4.626438  4.624655
#> [1751]  4.622873  4.621092  4.619311  4.617532  4.615754  4.613976  4.612199
#> [1758]  4.610424  4.608649  4.606875  4.605102  4.603330  4.601558  4.599788
#> [1765]  4.598018  4.596250  4.594482  4.592715  4.590949  4.589184  4.587420
#> [1772]  4.585657  4.583895  4.582133  4.580373  4.578613  4.576854  4.575096
#> [1779]  4.573339  4.571583  4.569828  4.568073  4.566320  4.564567  4.562815
#> [1786]  4.561064  4.559314  4.557565  4.555817  4.554069  4.552323  4.550577
#> [1793]  4.548833  4.547089  4.545346  4.543604  4.541862  4.540122  4.538382
#> [1800]  4.536644  4.534906  4.533169  4.531433  4.529698  4.527963  4.526230
#> [1807]  4.524497  4.522766  4.521035  4.519305  4.517576  4.515847  4.514120
#> [1814]  4.512394  4.510668  4.508943  4.507219  4.505496  4.503774  4.502052
#> [1821]  4.500332  4.498612  4.496893  4.495175  4.493458  4.491742  4.490027
#> [1828]  4.488312  4.486598  4.484885  4.483173  4.481462  4.479752  4.478042
#> [1835]  4.476334  4.474626  4.472919  4.471213  4.469508  4.467804  4.466100
#> [1842]  4.464397  4.462695  4.460994  4.459294  4.457595  4.455896  4.454199
#> [1849]  4.452502  4.450806  4.449111  4.447417  4.445723  4.444031  4.442339
#> [1856]  4.440648  4.438958  4.437268  4.435580  4.433892  4.432206  4.430520
#> [1863]  4.428834  4.427150  4.425467  4.423784  4.422102  4.420421  4.418741
#> [1870]  4.417062  4.415383  4.413706  4.412029  4.410353  4.408678  4.407003
#> [1877]  4.405330  4.403657  4.401985  4.400314  4.398644  4.396974  4.395305
#> [1884]  4.393638  4.391971  4.390304  4.388639  4.386975  4.385311  4.383648
#> [1891]  4.381986  4.380324  4.378664  4.377004  4.375345  4.373687  4.372030
#> [1898]  4.370374  4.368718  4.367063  4.365409  4.363756  4.362104  4.360452
#> [1905]  4.358801  4.357151  4.355502  4.353854  4.352206  4.350560  4.348914
#> [1912]  4.347268  4.345624  4.343981  4.342338  4.340696  4.339055  4.337414
#> [1919]  4.335775  4.334136  4.332498  4.330861  4.329225  4.327589  4.325954
#> [1926]  4.324320  4.322687  4.321055  4.319423  4.317792  4.316162  4.314533
#> [1933]  4.312904  4.311277  4.309650  4.308024  4.306399  4.304774  4.303150
#> [1940]  4.301527  4.299905  4.298284  4.296663  4.295044  4.293425  4.291806
#> [1947]  4.290189  4.288572  4.286956  4.285341  4.283727  4.282113  4.280501
#> [1954]  4.278889  4.277277  4.275667  4.274057  4.272448  4.270840  4.269233
#> [1961]  4.267626  4.266021  4.264416  4.262811  4.261208  4.259605  4.258003
#> [1968]  4.256402  4.254802  4.253202  4.251603  4.250005  4.248408  4.246812
#> [1975]  4.245216  4.243621  4.242027  4.240433  4.238840  4.237248  4.235657
#> [1982]  4.234067  4.232477  4.230888  4.229300  4.227713  4.226126  4.224540
#> [1989]  4.222955  4.221371  4.219787  4.218205  4.216623  4.215041  4.213461
#> [1996]  4.211881  4.210302  4.208724  4.207146  4.205569  4.203993