Computes d from modified Corrado-Su model.

d formula implementation from modified Corrado-Su model.

Usage

csmd(spot, strike, time, rate, yield, sigma, w)

Arguments

spot

current stock price.

strike

the strike price.

time

time to option expiration in years.

rate

the risk-free interest rate.

yield

the dividends that are expected to be paid.

sigma

volatility of the stock price.

w

constant from modified Corrado-Su model.

Value

Vector with d.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

See also

Examples

csmd(14, 120:160 / 10, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, 0, 3))
#>  [1]  1.47092255  1.40572099  1.34105609  1.27691906  1.21330137  1.15019467
#>  [7]  1.08759081  1.02548186  0.96386004  0.90271777  0.84204764  0.78184243
#> [13]  0.72209505  0.66279861  0.60394633  0.54553162  0.48754802  0.42998922
#> [19]  0.37284902  0.31612140  0.25980042  0.20388032  0.14835541  0.09322015
#> [25]  0.03846911 -0.01590302 -0.06990146 -0.12353131 -0.17679756 -0.22970511
#> [31] -0.28225877 -0.33446323 -0.38632310 -0.43784290 -0.48902707 -0.53987995
#> [37] -0.59040579 -0.64060879 -0.69049303 -0.74006255 -0.78932128
csmd(14, 13, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, -1000:1000 / 100, 3))
#>    [1] 0.8690933 0.8690662 0.8690391 0.8690120 0.8689849 0.8689578 0.8689307
#>    [8] 0.8689036 0.8688765 0.8688494 0.8688223 0.8687952 0.8687682 0.8687411
#>   [15] 0.8687140 0.8686869 0.8686598 0.8686327 0.8686056 0.8685785 0.8685514
#>   [22] 0.8685243 0.8684972 0.8684702 0.8684431 0.8684160 0.8683889 0.8683618
#>   [29] 0.8683347 0.8683076 0.8682805 0.8682534 0.8682264 0.8681993 0.8681722
#>   [36] 0.8681451 0.8681180 0.8680909 0.8680638 0.8680367 0.8680096 0.8679826
#>   [43] 0.8679555 0.8679284 0.8679013 0.8678742 0.8678471 0.8678200 0.8677929
#>   [50] 0.8677658 0.8677388 0.8677117 0.8676846 0.8676575 0.8676304 0.8676033
#>   [57] 0.8675762 0.8675491 0.8675221 0.8674950 0.8674679 0.8674408 0.8674137
#>   [64] 0.8673866 0.8673595 0.8673325 0.8673054 0.8672783 0.8672512 0.8672241
#>   [71] 0.8671970 0.8671699 0.8671429 0.8671158 0.8670887 0.8670616 0.8670345
#>   [78] 0.8670074 0.8669803 0.8669533 0.8669262 0.8668991 0.8668720 0.8668449
#>   [85] 0.8668178 0.8667908 0.8667637 0.8667366 0.8667095 0.8666824 0.8666553
#>   [92] 0.8666282 0.8666012 0.8665741 0.8665470 0.8665199 0.8664928 0.8664657
#>   [99] 0.8664387 0.8664116 0.8663845 0.8663574 0.8663303 0.8663033 0.8662762
#>  [106] 0.8662491 0.8662220 0.8661949 0.8661678 0.8661408 0.8661137 0.8660866
#>  [113] 0.8660595 0.8660324 0.8660053 0.8659783 0.8659512 0.8659241 0.8658970
#>  [120] 0.8658699 0.8658429 0.8658158 0.8657887 0.8657616 0.8657345 0.8657075
#>  [127] 0.8656804 0.8656533 0.8656262 0.8655991 0.8655721 0.8655450 0.8655179
#>  [134] 0.8654908 0.8654637 0.8654367 0.8654096 0.8653825 0.8653554 0.8653283
#>  [141] 0.8653013 0.8652742 0.8652471 0.8652200 0.8651929 0.8651659 0.8651388
#>  [148] 0.8651117 0.8650846 0.8650576 0.8650305 0.8650034 0.8649763 0.8649492
#>  [155] 0.8649222 0.8648951 0.8648680 0.8648409 0.8648138 0.8647868 0.8647597
#>  [162] 0.8647326 0.8647055 0.8646785 0.8646514 0.8646243 0.8645972 0.8645702
#>  [169] 0.8645431 0.8645160 0.8644889 0.8644619 0.8644348 0.8644077 0.8643806
#>  [176] 0.8643535 0.8643265 0.8642994 0.8642723 0.8642452 0.8642182 0.8641911
#>  [183] 0.8641640 0.8641369 0.8641099 0.8640828 0.8640557 0.8640286 0.8640016
#>  [190] 0.8639745 0.8639474 0.8639203 0.8638933 0.8638662 0.8638391 0.8638120
#>  [197] 0.8637850 0.8637579 0.8637308 0.8637038 0.8636767 0.8636496 0.8636225
#>  [204] 0.8635955 0.8635684 0.8635413 0.8635142 0.8634872 0.8634601 0.8634330
#>  [211] 0.8634059 0.8633789 0.8633518 0.8633247 0.8632977 0.8632706 0.8632435
#>  [218] 0.8632164 0.8631894 0.8631623 0.8631352 0.8631082 0.8630811 0.8630540
#>  [225] 0.8630269 0.8629999 0.8629728 0.8629457 0.8629187 0.8628916 0.8628645
#>  [232] 0.8628374 0.8628104 0.8627833 0.8627562 0.8627292 0.8627021 0.8626750
#>  [239] 0.8626480 0.8626209 0.8625938 0.8625667 0.8625397 0.8625126 0.8624855
#>  [246] 0.8624585 0.8624314 0.8624043 0.8623773 0.8623502 0.8623231 0.8622960
#>  [253] 0.8622690 0.8622419 0.8622148 0.8621878 0.8621607 0.8621336 0.8621066
#>  [260] 0.8620795 0.8620524 0.8620254 0.8619983 0.8619712 0.8619442 0.8619171
#>  [267] 0.8618900 0.8618630 0.8618359 0.8618088 0.8617818 0.8617547 0.8617276
#>  [274] 0.8617006 0.8616735 0.8616464 0.8616194 0.8615923 0.8615652 0.8615382
#>  [281] 0.8615111 0.8614840 0.8614570 0.8614299 0.8614028 0.8613758 0.8613487
#>  [288] 0.8613216 0.8612946 0.8612675 0.8612404 0.8612134 0.8611863 0.8611592
#>  [295] 0.8611322 0.8611051 0.8610780 0.8610510 0.8610239 0.8609969 0.8609698
#>  [302] 0.8609427 0.8609157 0.8608886 0.8608615 0.8608345 0.8608074 0.8607803
#>  [309] 0.8607533 0.8607262 0.8606992 0.8606721 0.8606450 0.8606180 0.8605909
#>  [316] 0.8605638 0.8605368 0.8605097 0.8604826 0.8604556 0.8604285 0.8604015
#>  [323] 0.8603744 0.8603473 0.8603203 0.8602932 0.8602662 0.8602391 0.8602120
#>  [330] 0.8601850 0.8601579 0.8601308 0.8601038 0.8600767 0.8600497 0.8600226
#>  [337] 0.8599955 0.8599685 0.8599414 0.8599144 0.8598873 0.8598602 0.8598332
#>  [344] 0.8598061 0.8597791 0.8597520 0.8597249 0.8596979 0.8596708 0.8596438
#>  [351] 0.8596167 0.8595896 0.8595626 0.8595355 0.8595085 0.8594814 0.8594543
#>  [358] 0.8594273 0.8594002 0.8593732 0.8593461 0.8593190 0.8592920 0.8592649
#>  [365] 0.8592379 0.8592108 0.8591838 0.8591567 0.8591296 0.8591026 0.8590755
#>  [372] 0.8590485 0.8590214 0.8589944 0.8589673 0.8589402 0.8589132 0.8588861
#>  [379] 0.8588591 0.8588320 0.8588050 0.8587779 0.8587508 0.8587238 0.8586967
#>  [386] 0.8586697 0.8586426 0.8586156 0.8585885 0.8585614 0.8585344 0.8585073
#>  [393] 0.8584803 0.8584532 0.8584262 0.8583991 0.8583721 0.8583450 0.8583179
#>  [400] 0.8582909 0.8582638 0.8582368 0.8582097 0.8581827 0.8581556 0.8581286
#>  [407] 0.8581015 0.8580745 0.8580474 0.8580203 0.8579933 0.8579662 0.8579392
#>  [414] 0.8579121 0.8578851 0.8578580 0.8578310 0.8578039 0.8577769 0.8577498
#>  [421] 0.8577228 0.8576957 0.8576686 0.8576416 0.8576145 0.8575875 0.8575604
#>  [428] 0.8575334 0.8575063 0.8574793 0.8574522 0.8574252 0.8573981 0.8573711
#>  [435] 0.8573440 0.8573170 0.8572899 0.8572629 0.8572358 0.8572088 0.8571817
#>  [442] 0.8571547 0.8571276 0.8571006 0.8570735 0.8570465 0.8570194 0.8569924
#>  [449] 0.8569653 0.8569383 0.8569112 0.8568842 0.8568571 0.8568301 0.8568030
#>  [456] 0.8567760 0.8567489 0.8567219 0.8566948 0.8566678 0.8566407 0.8566137
#>  [463] 0.8565866 0.8565596 0.8565325 0.8565055 0.8564784 0.8564514 0.8564243
#>  [470] 0.8563973 0.8563702 0.8563432 0.8563161 0.8562891 0.8562620 0.8562350
#>  [477] 0.8562079 0.8561809 0.8561538 0.8561268 0.8560997 0.8560727 0.8560456
#>  [484] 0.8560186 0.8559916 0.8559645 0.8559375 0.8559104 0.8558834 0.8558563
#>  [491] 0.8558293 0.8558022 0.8557752 0.8557481 0.8557211 0.8556940 0.8556670
#>  [498] 0.8556399 0.8556129 0.8555859 0.8555588 0.8555318 0.8555047 0.8554777
#>  [505] 0.8554506 0.8554236 0.8553965 0.8553695 0.8553424 0.8553154 0.8552884
#>  [512] 0.8552613 0.8552343 0.8552072 0.8551802 0.8551531 0.8551261 0.8550990
#>  [519] 0.8550720 0.8550450 0.8550179 0.8549909 0.8549638 0.8549368 0.8549097
#>  [526] 0.8548827 0.8548557 0.8548286 0.8548016 0.8547745 0.8547475 0.8547204
#>  [533] 0.8546934 0.8546664 0.8546393 0.8546123 0.8545852 0.8545582 0.8545311
#>  [540] 0.8545041 0.8544771 0.8544500 0.8544230 0.8543959 0.8543689 0.8543419
#>  [547] 0.8543148 0.8542878 0.8542607 0.8542337 0.8542066 0.8541796 0.8541526
#>  [554] 0.8541255 0.8540985 0.8540714 0.8540444 0.8540174 0.8539903 0.8539633
#>  [561] 0.8539362 0.8539092 0.8538822 0.8538551 0.8538281 0.8538010 0.8537740
#>  [568] 0.8537470 0.8537199 0.8536929 0.8536658 0.8536388 0.8536118 0.8535847
#>  [575] 0.8535577 0.8535307 0.8535036 0.8534766 0.8534495 0.8534225 0.8533955
#>  [582] 0.8533684 0.8533414 0.8533143 0.8532873 0.8532603 0.8532332 0.8532062
#>  [589] 0.8531792 0.8531521 0.8531251 0.8530980 0.8530710 0.8530440 0.8530169
#>  [596] 0.8529899 0.8529629 0.8529358 0.8529088 0.8528818 0.8528547 0.8528277
#>  [603] 0.8528006 0.8527736 0.8527466 0.8527195 0.8526925 0.8526655 0.8526384
#>  [610] 0.8526114 0.8525844 0.8525573 0.8525303 0.8525033 0.8524762 0.8524492
#>  [617] 0.8524221 0.8523951 0.8523681 0.8523410 0.8523140 0.8522870 0.8522599
#>  [624] 0.8522329 0.8522059 0.8521788 0.8521518 0.8521248 0.8520977 0.8520707
#>  [631] 0.8520437 0.8520166 0.8519896 0.8519626 0.8519355 0.8519085 0.8518815
#>  [638] 0.8518544 0.8518274 0.8518004 0.8517733 0.8517463 0.8517193 0.8516922
#>  [645] 0.8516652 0.8516382 0.8516111 0.8515841 0.8515571 0.8515301 0.8515030
#>  [652] 0.8514760 0.8514490 0.8514219 0.8513949 0.8513679 0.8513408 0.8513138
#>  [659] 0.8512868 0.8512597 0.8512327 0.8512057 0.8511786 0.8511516 0.8511246
#>  [666] 0.8510976 0.8510705 0.8510435 0.8510165 0.8509894 0.8509624 0.8509354
#>  [673] 0.8509083 0.8508813 0.8508543 0.8508273 0.8508002 0.8507732 0.8507462
#>  [680] 0.8507191 0.8506921 0.8506651 0.8506381 0.8506110 0.8505840 0.8505570
#>  [687] 0.8505299 0.8505029 0.8504759 0.8504489 0.8504218 0.8503948 0.8503678
#>  [694] 0.8503407 0.8503137 0.8502867 0.8502597 0.8502326 0.8502056 0.8501786
#>  [701] 0.8501516 0.8501245 0.8500975 0.8500705 0.8500434 0.8500164 0.8499894
#>  [708] 0.8499624 0.8499353 0.8499083 0.8498813 0.8498543 0.8498272 0.8498002
#>  [715] 0.8497732 0.8497462 0.8497191 0.8496921 0.8496651 0.8496381 0.8496110
#>  [722] 0.8495840 0.8495570 0.8495300 0.8495029 0.8494759 0.8494489 0.8494219
#>  [729] 0.8493948 0.8493678 0.8493408 0.8493138 0.8492867 0.8492597 0.8492327
#>  [736] 0.8492057 0.8491786 0.8491516 0.8491246 0.8490976 0.8490706 0.8490435
#>  [743] 0.8490165 0.8489895 0.8489625 0.8489354 0.8489084 0.8488814 0.8488544
#>  [750] 0.8488273 0.8488003 0.8487733 0.8487463 0.8487193 0.8486922 0.8486652
#>  [757] 0.8486382 0.8486112 0.8485841 0.8485571 0.8485301 0.8485031 0.8484761
#>  [764] 0.8484490 0.8484220 0.8483950 0.8483680 0.8483410 0.8483139 0.8482869
#>  [771] 0.8482599 0.8482329 0.8482059 0.8481788 0.8481518 0.8481248 0.8480978
#>  [778] 0.8480708 0.8480437 0.8480167 0.8479897 0.8479627 0.8479357 0.8479086
#>  [785] 0.8478816 0.8478546 0.8478276 0.8478006 0.8477735 0.8477465 0.8477195
#>  [792] 0.8476925 0.8476655 0.8476384 0.8476114 0.8475844 0.8475574 0.8475304
#>  [799] 0.8475034 0.8474763 0.8474493 0.8474223 0.8473953 0.8473683 0.8473413
#>  [806] 0.8473142 0.8472872 0.8472602 0.8472332 0.8472062 0.8471791 0.8471521
#>  [813] 0.8471251 0.8470981 0.8470711 0.8470441 0.8470171 0.8469900 0.8469630
#>  [820] 0.8469360 0.8469090 0.8468820 0.8468550 0.8468279 0.8468009 0.8467739
#>  [827] 0.8467469 0.8467199 0.8466929 0.8466658 0.8466388 0.8466118 0.8465848
#>  [834] 0.8465578 0.8465308 0.8465038 0.8464767 0.8464497 0.8464227 0.8463957
#>  [841] 0.8463687 0.8463417 0.8463147 0.8462876 0.8462606 0.8462336 0.8462066
#>  [848] 0.8461796 0.8461526 0.8461256 0.8460986 0.8460715 0.8460445 0.8460175
#>  [855] 0.8459905 0.8459635 0.8459365 0.8459095 0.8458825 0.8458554 0.8458284
#>  [862] 0.8458014 0.8457744 0.8457474 0.8457204 0.8456934 0.8456664 0.8456393
#>  [869] 0.8456123 0.8455853 0.8455583 0.8455313 0.8455043 0.8454773 0.8454503
#>  [876] 0.8454233 0.8453962 0.8453692 0.8453422 0.8453152 0.8452882 0.8452612
#>  [883] 0.8452342 0.8452072 0.8451802 0.8451532 0.8451261 0.8450991 0.8450721
#>  [890] 0.8450451 0.8450181 0.8449911 0.8449641 0.8449371 0.8449101 0.8448831
#>  [897] 0.8448561 0.8448290 0.8448020 0.8447750 0.8447480 0.8447210 0.8446940
#>  [904] 0.8446670 0.8446400 0.8446130 0.8445860 0.8445590 0.8445320 0.8445049
#>  [911] 0.8444779 0.8444509 0.8444239 0.8443969 0.8443699 0.8443429 0.8443159
#>  [918] 0.8442889 0.8442619 0.8442349 0.8442079 0.8441809 0.8441539 0.8441268
#>  [925] 0.8440998 0.8440728 0.8440458 0.8440188 0.8439918 0.8439648 0.8439378
#>  [932] 0.8439108 0.8438838 0.8438568 0.8438298 0.8438028 0.8437758 0.8437488
#>  [939] 0.8437218 0.8436948 0.8436678 0.8436408 0.8436137 0.8435867 0.8435597
#>  [946] 0.8435327 0.8435057 0.8434787 0.8434517 0.8434247 0.8433977 0.8433707
#>  [953] 0.8433437 0.8433167 0.8432897 0.8432627 0.8432357 0.8432087 0.8431817
#>  [960] 0.8431547 0.8431277 0.8431007 0.8430737 0.8430467 0.8430197 0.8429927
#>  [967] 0.8429657 0.8429387 0.8429117 0.8428847 0.8428577 0.8428307 0.8428037
#>  [974] 0.8427767 0.8427497 0.8427226 0.8426956 0.8426686 0.8426416 0.8426146
#>  [981] 0.8425876 0.8425606 0.8425336 0.8425066 0.8424796 0.8424526 0.8424256
#>  [988] 0.8423986 0.8423716 0.8423446 0.8423176 0.8422906 0.8422636 0.8422366
#>  [995] 0.8422096 0.8421826 0.8421556 0.8421286 0.8421016 0.8420746 0.8420476
#> [1002] 0.8420206 0.8419936 0.8419666 0.8419396 0.8419126 0.8418856 0.8418587
#> [1009] 0.8418317 0.8418047 0.8417777 0.8417507 0.8417237 0.8416967 0.8416697
#> [1016] 0.8416427 0.8416157 0.8415887 0.8415617 0.8415347 0.8415077 0.8414807
#> [1023] 0.8414537 0.8414267 0.8413997 0.8413727 0.8413457 0.8413187 0.8412917
#> [1030] 0.8412647 0.8412377 0.8412107 0.8411837 0.8411567 0.8411297 0.8411027
#> [1037] 0.8410757 0.8410487 0.8410217 0.8409947 0.8409678 0.8409408 0.8409138
#> [1044] 0.8408868 0.8408598 0.8408328 0.8408058 0.8407788 0.8407518 0.8407248
#> [1051] 0.8406978 0.8406708 0.8406438 0.8406168 0.8405898 0.8405628 0.8405358
#> [1058] 0.8405088 0.8404819 0.8404549 0.8404279 0.8404009 0.8403739 0.8403469
#> [1065] 0.8403199 0.8402929 0.8402659 0.8402389 0.8402119 0.8401849 0.8401579
#> [1072] 0.8401309 0.8401039 0.8400770 0.8400500 0.8400230 0.8399960 0.8399690
#> [1079] 0.8399420 0.8399150 0.8398880 0.8398610 0.8398340 0.8398070 0.8397800
#> [1086] 0.8397531 0.8397261 0.8396991 0.8396721 0.8396451 0.8396181 0.8395911
#> [1093] 0.8395641 0.8395371 0.8395101 0.8394831 0.8394562 0.8394292 0.8394022
#> [1100] 0.8393752 0.8393482 0.8393212 0.8392942 0.8392672 0.8392402 0.8392132
#> [1107] 0.8391863 0.8391593 0.8391323 0.8391053 0.8390783 0.8390513 0.8390243
#> [1114] 0.8389973 0.8389703 0.8389434 0.8389164 0.8388894 0.8388624 0.8388354
#> [1121] 0.8388084 0.8387814 0.8387544 0.8387275 0.8387005 0.8386735 0.8386465
#> [1128] 0.8386195 0.8385925 0.8385655 0.8385385 0.8385116 0.8384846 0.8384576
#> [1135] 0.8384306 0.8384036 0.8383766 0.8383496 0.8383226 0.8382957 0.8382687
#> [1142] 0.8382417 0.8382147 0.8381877 0.8381607 0.8381337 0.8381068 0.8380798
#> [1149] 0.8380528 0.8380258 0.8379988 0.8379718 0.8379448 0.8379179 0.8378909
#> [1156] 0.8378639 0.8378369 0.8378099 0.8377829 0.8377560 0.8377290 0.8377020
#> [1163] 0.8376750 0.8376480 0.8376210 0.8375941 0.8375671 0.8375401 0.8375131
#> [1170] 0.8374861 0.8374591 0.8374322 0.8374052 0.8373782 0.8373512 0.8373242
#> [1177] 0.8372972 0.8372703 0.8372433 0.8372163 0.8371893 0.8371623 0.8371353
#> [1184] 0.8371084 0.8370814 0.8370544 0.8370274 0.8370004 0.8369734 0.8369465
#> [1191] 0.8369195 0.8368925 0.8368655 0.8368385 0.8368116 0.8367846 0.8367576
#> [1198] 0.8367306 0.8367036 0.8366767 0.8366497 0.8366227 0.8365957 0.8365687
#> [1205] 0.8365418 0.8365148 0.8364878 0.8364608 0.8364338 0.8364069 0.8363799
#> [1212] 0.8363529 0.8363259 0.8362989 0.8362720 0.8362450 0.8362180 0.8361910
#> [1219] 0.8361640 0.8361371 0.8361101 0.8360831 0.8360561 0.8360291 0.8360022
#> [1226] 0.8359752 0.8359482 0.8359212 0.8358943 0.8358673 0.8358403 0.8358133
#> [1233] 0.8357863 0.8357594 0.8357324 0.8357054 0.8356784 0.8356515 0.8356245
#> [1240] 0.8355975 0.8355705 0.8355435 0.8355166 0.8354896 0.8354626 0.8354356
#> [1247] 0.8354087 0.8353817 0.8353547 0.8353277 0.8353008 0.8352738 0.8352468
#> [1254] 0.8352198 0.8351929 0.8351659 0.8351389 0.8351119 0.8350850 0.8350580
#> [1261] 0.8350310 0.8350040 0.8349771 0.8349501 0.8349231 0.8348961 0.8348692
#> [1268] 0.8348422 0.8348152 0.8347882 0.8347613 0.8347343 0.8347073 0.8346803
#> [1275] 0.8346534 0.8346264 0.8345994 0.8345724 0.8345455 0.8345185 0.8344915
#> [1282] 0.8344646 0.8344376 0.8344106 0.8343836 0.8343567 0.8343297 0.8343027
#> [1289] 0.8342757 0.8342488 0.8342218 0.8341948 0.8341679 0.8341409 0.8341139
#> [1296] 0.8340869 0.8340600 0.8340330 0.8340060 0.8339791 0.8339521 0.8339251
#> [1303] 0.8338981 0.8338712 0.8338442 0.8338172 0.8337903 0.8337633 0.8337363
#> [1310] 0.8337093 0.8336824 0.8336554 0.8336284 0.8336015 0.8335745 0.8335475
#> [1317] 0.8335206 0.8334936 0.8334666 0.8334396 0.8334127 0.8333857 0.8333587
#> [1324] 0.8333318 0.8333048 0.8332778 0.8332509 0.8332239 0.8331969 0.8331700
#> [1331] 0.8331430 0.8331160 0.8330890 0.8330621 0.8330351 0.8330081 0.8329812
#> [1338] 0.8329542 0.8329272 0.8329003 0.8328733 0.8328463 0.8328194 0.8327924
#> [1345] 0.8327654 0.8327385 0.8327115 0.8326845 0.8326576 0.8326306 0.8326036
#> [1352] 0.8325767 0.8325497 0.8325227 0.8324958 0.8324688 0.8324418 0.8324149
#> [1359] 0.8323879 0.8323609 0.8323340 0.8323070 0.8322800 0.8322531 0.8322261
#> [1366] 0.8321991 0.8321722 0.8321452 0.8321182 0.8320913 0.8320643 0.8320374
#> [1373] 0.8320104 0.8319834 0.8319565 0.8319295 0.8319025 0.8318756 0.8318486
#> [1380] 0.8318216 0.8317947 0.8317677 0.8317407 0.8317138 0.8316868 0.8316599
#> [1387] 0.8316329 0.8316059 0.8315790 0.8315520 0.8315250 0.8314981 0.8314711
#> [1394] 0.8314441 0.8314172 0.8313902 0.8313633 0.8313363 0.8313093 0.8312824
#> [1401] 0.8312554 0.8312285 0.8312015 0.8311745 0.8311476 0.8311206 0.8310936
#> [1408] 0.8310667 0.8310397 0.8310128 0.8309858 0.8309588 0.8309319 0.8309049
#> [1415] 0.8308780 0.8308510 0.8308240 0.8307971 0.8307701 0.8307432 0.8307162
#> [1422] 0.8306892 0.8306623 0.8306353 0.8306084 0.8305814 0.8305544 0.8305275
#> [1429] 0.8305005 0.8304736 0.8304466 0.8304196 0.8303927 0.8303657 0.8303388
#> [1436] 0.8303118 0.8302848 0.8302579 0.8302309 0.8302040 0.8301770 0.8301500
#> [1443] 0.8301231 0.8300961 0.8300692 0.8300422 0.8300153 0.8299883 0.8299613
#> [1450] 0.8299344 0.8299074 0.8298805 0.8298535 0.8298266 0.8297996 0.8297726
#> [1457] 0.8297457 0.8297187 0.8296918 0.8296648 0.8296379 0.8296109 0.8295839
#> [1464] 0.8295570 0.8295300 0.8295031 0.8294761 0.8294492 0.8294222 0.8293953
#> [1471] 0.8293683 0.8293413 0.8293144 0.8292874 0.8292605 0.8292335 0.8292066
#> [1478] 0.8291796 0.8291527 0.8291257 0.8290987 0.8290718 0.8290448 0.8290179
#> [1485] 0.8289909 0.8289640 0.8289370 0.8289101 0.8288831 0.8288562 0.8288292
#> [1492] 0.8288022 0.8287753 0.8287483 0.8287214 0.8286944 0.8286675 0.8286405
#> [1499] 0.8286136 0.8285866 0.8285597 0.8285327 0.8285058 0.8284788 0.8284519
#> [1506] 0.8284249 0.8283980 0.8283710 0.8283441 0.8283171 0.8282901 0.8282632
#> [1513] 0.8282362 0.8282093 0.8281823 0.8281554 0.8281284 0.8281015 0.8280745
#> [1520] 0.8280476 0.8280206 0.8279937 0.8279667 0.8279398 0.8279128 0.8278859
#> [1527] 0.8278589 0.8278320 0.8278050 0.8277781 0.8277511 0.8277242 0.8276972
#> [1534] 0.8276703 0.8276433 0.8276164 0.8275894 0.8275625 0.8275355 0.8275086
#> [1541] 0.8274816 0.8274547 0.8274277 0.8274008 0.8273738 0.8273469 0.8273199
#> [1548] 0.8272930 0.8272660 0.8272391 0.8272121 0.8271852 0.8271582 0.8271313
#> [1555] 0.8271044 0.8270774 0.8270505 0.8270235 0.8269966 0.8269696 0.8269427
#> [1562] 0.8269157 0.8268888 0.8268618 0.8268349 0.8268079 0.8267810 0.8267540
#> [1569] 0.8267271 0.8267001 0.8266732 0.8266463 0.8266193 0.8265924 0.8265654
#> [1576] 0.8265385 0.8265115 0.8264846 0.8264576 0.8264307 0.8264037 0.8263768
#> [1583] 0.8263499 0.8263229 0.8262960 0.8262690 0.8262421 0.8262151 0.8261882
#> [1590] 0.8261612 0.8261343 0.8261073 0.8260804 0.8260535 0.8260265 0.8259996
#> [1597] 0.8259726 0.8259457 0.8259187 0.8258918 0.8258649 0.8258379 0.8258110
#> [1604] 0.8257840 0.8257571 0.8257301 0.8257032 0.8256762 0.8256493 0.8256224
#> [1611] 0.8255954 0.8255685 0.8255415 0.8255146 0.8254877 0.8254607 0.8254338
#> [1618] 0.8254068 0.8253799 0.8253529 0.8253260 0.8252991 0.8252721 0.8252452
#> [1625] 0.8252182 0.8251913 0.8251644 0.8251374 0.8251105 0.8250835 0.8250566
#> [1632] 0.8250296 0.8250027 0.8249758 0.8249488 0.8249219 0.8248949 0.8248680
#> [1639] 0.8248411 0.8248141 0.8247872 0.8247602 0.8247333 0.8247064 0.8246794
#> [1646] 0.8246525 0.8246255 0.8245986 0.8245717 0.8245447 0.8245178 0.8244909
#> [1653] 0.8244639 0.8244370 0.8244100 0.8243831 0.8243562 0.8243292 0.8243023
#> [1660] 0.8242753 0.8242484 0.8242215 0.8241945 0.8241676 0.8241407 0.8241137
#> [1667] 0.8240868 0.8240598 0.8240329 0.8240060 0.8239790 0.8239521 0.8239252
#> [1674] 0.8238982 0.8238713 0.8238443 0.8238174 0.8237905 0.8237635 0.8237366
#> [1681] 0.8237097 0.8236827 0.8236558 0.8236289 0.8236019 0.8235750 0.8235480
#> [1688] 0.8235211 0.8234942 0.8234672 0.8234403 0.8234134 0.8233864 0.8233595
#> [1695] 0.8233326 0.8233056 0.8232787 0.8232518 0.8232248 0.8231979 0.8231710
#> [1702] 0.8231440 0.8231171 0.8230902 0.8230632 0.8230363 0.8230094 0.8229824
#> [1709] 0.8229555 0.8229286 0.8229016 0.8228747 0.8228478 0.8228208 0.8227939
#> [1716] 0.8227670 0.8227400 0.8227131 0.8226862 0.8226592 0.8226323 0.8226054
#> [1723] 0.8225784 0.8225515 0.8225246 0.8224976 0.8224707 0.8224438 0.8224168
#> [1730] 0.8223899 0.8223630 0.8223360 0.8223091 0.8222822 0.8222552 0.8222283
#> [1737] 0.8222014 0.8221745 0.8221475 0.8221206 0.8220937 0.8220667 0.8220398
#> [1744] 0.8220129 0.8219859 0.8219590 0.8219321 0.8219051 0.8218782 0.8218513
#> [1751] 0.8218244 0.8217974 0.8217705 0.8217436 0.8217166 0.8216897 0.8216628
#> [1758] 0.8216359 0.8216089 0.8215820 0.8215551 0.8215281 0.8215012 0.8214743
#> [1765] 0.8214473 0.8214204 0.8213935 0.8213666 0.8213396 0.8213127 0.8212858
#> [1772] 0.8212589 0.8212319 0.8212050 0.8211781 0.8211511 0.8211242 0.8210973
#> [1779] 0.8210704 0.8210434 0.8210165 0.8209896 0.8209627 0.8209357 0.8209088
#> [1786] 0.8208819 0.8208549 0.8208280 0.8208011 0.8207742 0.8207472 0.8207203
#> [1793] 0.8206934 0.8206665 0.8206395 0.8206126 0.8205857 0.8205588 0.8205318
#> [1800] 0.8205049 0.8204780 0.8204511 0.8204241 0.8203972 0.8203703 0.8203434
#> [1807] 0.8203164 0.8202895 0.8202626 0.8202357 0.8202087 0.8201818 0.8201549
#> [1814] 0.8201280 0.8201010 0.8200741 0.8200472 0.8200203 0.8199934 0.8199664
#> [1821] 0.8199395 0.8199126 0.8198857 0.8198587 0.8198318 0.8198049 0.8197780
#> [1828] 0.8197510 0.8197241 0.8196972 0.8196703 0.8196434 0.8196164 0.8195895
#> [1835] 0.8195626 0.8195357 0.8195087 0.8194818 0.8194549 0.8194280 0.8194011
#> [1842] 0.8193741 0.8193472 0.8193203 0.8192934 0.8192665 0.8192395 0.8192126
#> [1849] 0.8191857 0.8191588 0.8191318 0.8191049 0.8190780 0.8190511 0.8190242
#> [1856] 0.8189972 0.8189703 0.8189434 0.8189165 0.8188896 0.8188626 0.8188357
#> [1863] 0.8188088 0.8187819 0.8187550 0.8187281 0.8187011 0.8186742 0.8186473
#> [1870] 0.8186204 0.8185935 0.8185665 0.8185396 0.8185127 0.8184858 0.8184589
#> [1877] 0.8184319 0.8184050 0.8183781 0.8183512 0.8183243 0.8182974 0.8182704
#> [1884] 0.8182435 0.8182166 0.8181897 0.8181628 0.8181359 0.8181089 0.8180820
#> [1891] 0.8180551 0.8180282 0.8180013 0.8179744 0.8179474 0.8179205 0.8178936
#> [1898] 0.8178667 0.8178398 0.8178129 0.8177859 0.8177590 0.8177321 0.8177052
#> [1905] 0.8176783 0.8176514 0.8176244 0.8175975 0.8175706 0.8175437 0.8175168
#> [1912] 0.8174899 0.8174630 0.8174360 0.8174091 0.8173822 0.8173553 0.8173284
#> [1919] 0.8173015 0.8172746 0.8172476 0.8172207 0.8171938 0.8171669 0.8171400
#> [1926] 0.8171131 0.8170862 0.8170592 0.8170323 0.8170054 0.8169785 0.8169516
#> [1933] 0.8169247 0.8168978 0.8168709 0.8168439 0.8168170 0.8167901 0.8167632
#> [1940] 0.8167363 0.8167094 0.8166825 0.8166556 0.8166286 0.8166017 0.8165748
#> [1947] 0.8165479 0.8165210 0.8164941 0.8164672 0.8164403 0.8164133 0.8163864
#> [1954] 0.8163595 0.8163326 0.8163057 0.8162788 0.8162519 0.8162250 0.8161981
#> [1961] 0.8161712 0.8161442 0.8161173 0.8160904 0.8160635 0.8160366 0.8160097
#> [1968] 0.8159828 0.8159559 0.8159290 0.8159021 0.8158751 0.8158482 0.8158213
#> [1975] 0.8157944 0.8157675 0.8157406 0.8157137 0.8156868 0.8156599 0.8156330
#> [1982] 0.8156061 0.8155791 0.8155522 0.8155253 0.8154984 0.8154715 0.8154446
#> [1989] 0.8154177 0.8153908 0.8153639 0.8153370 0.8153101 0.8152832 0.8152563
#> [1996] 0.8152293 0.8152024 0.8151755 0.8151486 0.8151217 0.8150948
csmd(14, 13, 0.5, 0.15, 0.1, 0.18, csmw(0.18, 0.5, 0, -1000:1000 / 100))
#>    [1] 0.8431646 0.8431637 0.8431628 0.8431620 0.8431611 0.8431603 0.8431594
#>    [8] 0.8431585 0.8431577 0.8431568 0.8431560 0.8431551 0.8431542 0.8431534
#>   [15] 0.8431525 0.8431517 0.8431508 0.8431500 0.8431491 0.8431482 0.8431474
#>   [22] 0.8431465 0.8431457 0.8431448 0.8431439 0.8431431 0.8431422 0.8431414
#>   [29] 0.8431405 0.8431396 0.8431388 0.8431379 0.8431371 0.8431362 0.8431353
#>   [36] 0.8431345 0.8431336 0.8431328 0.8431319 0.8431311 0.8431302 0.8431293
#>   [43] 0.8431285 0.8431276 0.8431268 0.8431259 0.8431250 0.8431242 0.8431233
#>   [50] 0.8431225 0.8431216 0.8431207 0.8431199 0.8431190 0.8431182 0.8431173
#>   [57] 0.8431164 0.8431156 0.8431147 0.8431139 0.8431130 0.8431121 0.8431113
#>   [64] 0.8431104 0.8431096 0.8431087 0.8431079 0.8431070 0.8431061 0.8431053
#>   [71] 0.8431044 0.8431036 0.8431027 0.8431018 0.8431010 0.8431001 0.8430993
#>   [78] 0.8430984 0.8430975 0.8430967 0.8430958 0.8430950 0.8430941 0.8430932
#>   [85] 0.8430924 0.8430915 0.8430907 0.8430898 0.8430889 0.8430881 0.8430872
#>   [92] 0.8430864 0.8430855 0.8430847 0.8430838 0.8430829 0.8430821 0.8430812
#>   [99] 0.8430804 0.8430795 0.8430786 0.8430778 0.8430769 0.8430761 0.8430752
#>  [106] 0.8430743 0.8430735 0.8430726 0.8430718 0.8430709 0.8430700 0.8430692
#>  [113] 0.8430683 0.8430675 0.8430666 0.8430658 0.8430649 0.8430640 0.8430632
#>  [120] 0.8430623 0.8430615 0.8430606 0.8430597 0.8430589 0.8430580 0.8430572
#>  [127] 0.8430563 0.8430554 0.8430546 0.8430537 0.8430529 0.8430520 0.8430511
#>  [134] 0.8430503 0.8430494 0.8430486 0.8430477 0.8430468 0.8430460 0.8430451
#>  [141] 0.8430443 0.8430434 0.8430426 0.8430417 0.8430408 0.8430400 0.8430391
#>  [148] 0.8430383 0.8430374 0.8430365 0.8430357 0.8430348 0.8430340 0.8430331
#>  [155] 0.8430322 0.8430314 0.8430305 0.8430297 0.8430288 0.8430279 0.8430271
#>  [162] 0.8430262 0.8430254 0.8430245 0.8430236 0.8430228 0.8430219 0.8430211
#>  [169] 0.8430202 0.8430194 0.8430185 0.8430176 0.8430168 0.8430159 0.8430151
#>  [176] 0.8430142 0.8430133 0.8430125 0.8430116 0.8430108 0.8430099 0.8430090
#>  [183] 0.8430082 0.8430073 0.8430065 0.8430056 0.8430047 0.8430039 0.8430030
#>  [190] 0.8430022 0.8430013 0.8430004 0.8429996 0.8429987 0.8429979 0.8429970
#>  [197] 0.8429962 0.8429953 0.8429944 0.8429936 0.8429927 0.8429919 0.8429910
#>  [204] 0.8429901 0.8429893 0.8429884 0.8429876 0.8429867 0.8429858 0.8429850
#>  [211] 0.8429841 0.8429833 0.8429824 0.8429815 0.8429807 0.8429798 0.8429790
#>  [218] 0.8429781 0.8429773 0.8429764 0.8429755 0.8429747 0.8429738 0.8429730
#>  [225] 0.8429721 0.8429712 0.8429704 0.8429695 0.8429687 0.8429678 0.8429669
#>  [232] 0.8429661 0.8429652 0.8429644 0.8429635 0.8429626 0.8429618 0.8429609
#>  [239] 0.8429601 0.8429592 0.8429583 0.8429575 0.8429566 0.8429558 0.8429549
#>  [246] 0.8429541 0.8429532 0.8429523 0.8429515 0.8429506 0.8429498 0.8429489
#>  [253] 0.8429480 0.8429472 0.8429463 0.8429455 0.8429446 0.8429437 0.8429429
#>  [260] 0.8429420 0.8429412 0.8429403 0.8429394 0.8429386 0.8429377 0.8429369
#>  [267] 0.8429360 0.8429351 0.8429343 0.8429334 0.8429326 0.8429317 0.8429309
#>  [274] 0.8429300 0.8429291 0.8429283 0.8429274 0.8429266 0.8429257 0.8429248
#>  [281] 0.8429240 0.8429231 0.8429223 0.8429214 0.8429205 0.8429197 0.8429188
#>  [288] 0.8429180 0.8429171 0.8429162 0.8429154 0.8429145 0.8429137 0.8429128
#>  [295] 0.8429120 0.8429111 0.8429102 0.8429094 0.8429085 0.8429077 0.8429068
#>  [302] 0.8429059 0.8429051 0.8429042 0.8429034 0.8429025 0.8429016 0.8429008
#>  [309] 0.8428999 0.8428991 0.8428982 0.8428973 0.8428965 0.8428956 0.8428948
#>  [316] 0.8428939 0.8428930 0.8428922 0.8428913 0.8428905 0.8428896 0.8428888
#>  [323] 0.8428879 0.8428870 0.8428862 0.8428853 0.8428845 0.8428836 0.8428827
#>  [330] 0.8428819 0.8428810 0.8428802 0.8428793 0.8428784 0.8428776 0.8428767
#>  [337] 0.8428759 0.8428750 0.8428741 0.8428733 0.8428724 0.8428716 0.8428707
#>  [344] 0.8428699 0.8428690 0.8428681 0.8428673 0.8428664 0.8428656 0.8428647
#>  [351] 0.8428638 0.8428630 0.8428621 0.8428613 0.8428604 0.8428595 0.8428587
#>  [358] 0.8428578 0.8428570 0.8428561 0.8428552 0.8428544 0.8428535 0.8428527
#>  [365] 0.8428518 0.8428509 0.8428501 0.8428492 0.8428484 0.8428475 0.8428467
#>  [372] 0.8428458 0.8428449 0.8428441 0.8428432 0.8428424 0.8428415 0.8428406
#>  [379] 0.8428398 0.8428389 0.8428381 0.8428372 0.8428363 0.8428355 0.8428346
#>  [386] 0.8428338 0.8428329 0.8428320 0.8428312 0.8428303 0.8428295 0.8428286
#>  [393] 0.8428278 0.8428269 0.8428260 0.8428252 0.8428243 0.8428235 0.8428226
#>  [400] 0.8428217 0.8428209 0.8428200 0.8428192 0.8428183 0.8428174 0.8428166
#>  [407] 0.8428157 0.8428149 0.8428140 0.8428131 0.8428123 0.8428114 0.8428106
#>  [414] 0.8428097 0.8428088 0.8428080 0.8428071 0.8428063 0.8428054 0.8428046
#>  [421] 0.8428037 0.8428028 0.8428020 0.8428011 0.8428003 0.8427994 0.8427985
#>  [428] 0.8427977 0.8427968 0.8427960 0.8427951 0.8427942 0.8427934 0.8427925
#>  [435] 0.8427917 0.8427908 0.8427899 0.8427891 0.8427882 0.8427874 0.8427865
#>  [442] 0.8427856 0.8427848 0.8427839 0.8427831 0.8427822 0.8427814 0.8427805
#>  [449] 0.8427796 0.8427788 0.8427779 0.8427771 0.8427762 0.8427753 0.8427745
#>  [456] 0.8427736 0.8427728 0.8427719 0.8427710 0.8427702 0.8427693 0.8427685
#>  [463] 0.8427676 0.8427667 0.8427659 0.8427650 0.8427642 0.8427633 0.8427625
#>  [470] 0.8427616 0.8427607 0.8427599 0.8427590 0.8427582 0.8427573 0.8427564
#>  [477] 0.8427556 0.8427547 0.8427539 0.8427530 0.8427521 0.8427513 0.8427504
#>  [484] 0.8427496 0.8427487 0.8427478 0.8427470 0.8427461 0.8427453 0.8427444
#>  [491] 0.8427435 0.8427427 0.8427418 0.8427410 0.8427401 0.8427393 0.8427384
#>  [498] 0.8427375 0.8427367 0.8427358 0.8427350 0.8427341 0.8427332 0.8427324
#>  [505] 0.8427315 0.8427307 0.8427298 0.8427289 0.8427281 0.8427272 0.8427264
#>  [512] 0.8427255 0.8427246 0.8427238 0.8427229 0.8427221 0.8427212 0.8427204
#>  [519] 0.8427195 0.8427186 0.8427178 0.8427169 0.8427161 0.8427152 0.8427143
#>  [526] 0.8427135 0.8427126 0.8427118 0.8427109 0.8427100 0.8427092 0.8427083
#>  [533] 0.8427075 0.8427066 0.8427057 0.8427049 0.8427040 0.8427032 0.8427023
#>  [540] 0.8427015 0.8427006 0.8426997 0.8426989 0.8426980 0.8426972 0.8426963
#>  [547] 0.8426954 0.8426946 0.8426937 0.8426929 0.8426920 0.8426911 0.8426903
#>  [554] 0.8426894 0.8426886 0.8426877 0.8426868 0.8426860 0.8426851 0.8426843
#>  [561] 0.8426834 0.8426825 0.8426817 0.8426808 0.8426800 0.8426791 0.8426783
#>  [568] 0.8426774 0.8426765 0.8426757 0.8426748 0.8426740 0.8426731 0.8426722
#>  [575] 0.8426714 0.8426705 0.8426697 0.8426688 0.8426679 0.8426671 0.8426662
#>  [582] 0.8426654 0.8426645 0.8426636 0.8426628 0.8426619 0.8426611 0.8426602
#>  [589] 0.8426594 0.8426585 0.8426576 0.8426568 0.8426559 0.8426551 0.8426542
#>  [596] 0.8426533 0.8426525 0.8426516 0.8426508 0.8426499 0.8426490 0.8426482
#>  [603] 0.8426473 0.8426465 0.8426456 0.8426447 0.8426439 0.8426430 0.8426422
#>  [610] 0.8426413 0.8426404 0.8426396 0.8426387 0.8426379 0.8426370 0.8426362
#>  [617] 0.8426353 0.8426344 0.8426336 0.8426327 0.8426319 0.8426310 0.8426301
#>  [624] 0.8426293 0.8426284 0.8426276 0.8426267 0.8426258 0.8426250 0.8426241
#>  [631] 0.8426233 0.8426224 0.8426215 0.8426207 0.8426198 0.8426190 0.8426181
#>  [638] 0.8426173 0.8426164 0.8426155 0.8426147 0.8426138 0.8426130 0.8426121
#>  [645] 0.8426112 0.8426104 0.8426095 0.8426087 0.8426078 0.8426069 0.8426061
#>  [652] 0.8426052 0.8426044 0.8426035 0.8426026 0.8426018 0.8426009 0.8426001
#>  [659] 0.8425992 0.8425983 0.8425975 0.8425966 0.8425958 0.8425949 0.8425941
#>  [666] 0.8425932 0.8425923 0.8425915 0.8425906 0.8425898 0.8425889 0.8425880
#>  [673] 0.8425872 0.8425863 0.8425855 0.8425846 0.8425837 0.8425829 0.8425820
#>  [680] 0.8425812 0.8425803 0.8425794 0.8425786 0.8425777 0.8425769 0.8425760
#>  [687] 0.8425752 0.8425743 0.8425734 0.8425726 0.8425717 0.8425709 0.8425700
#>  [694] 0.8425691 0.8425683 0.8425674 0.8425666 0.8425657 0.8425648 0.8425640
#>  [701] 0.8425631 0.8425623 0.8425614 0.8425605 0.8425597 0.8425588 0.8425580
#>  [708] 0.8425571 0.8425563 0.8425554 0.8425545 0.8425537 0.8425528 0.8425520
#>  [715] 0.8425511 0.8425502 0.8425494 0.8425485 0.8425477 0.8425468 0.8425459
#>  [722] 0.8425451 0.8425442 0.8425434 0.8425425 0.8425416 0.8425408 0.8425399
#>  [729] 0.8425391 0.8425382 0.8425373 0.8425365 0.8425356 0.8425348 0.8425339
#>  [736] 0.8425331 0.8425322 0.8425313 0.8425305 0.8425296 0.8425288 0.8425279
#>  [743] 0.8425270 0.8425262 0.8425253 0.8425245 0.8425236 0.8425227 0.8425219
#>  [750] 0.8425210 0.8425202 0.8425193 0.8425184 0.8425176 0.8425167 0.8425159
#>  [757] 0.8425150 0.8425142 0.8425133 0.8425124 0.8425116 0.8425107 0.8425099
#>  [764] 0.8425090 0.8425081 0.8425073 0.8425064 0.8425056 0.8425047 0.8425038
#>  [771] 0.8425030 0.8425021 0.8425013 0.8425004 0.8424995 0.8424987 0.8424978
#>  [778] 0.8424970 0.8424961 0.8424953 0.8424944 0.8424935 0.8424927 0.8424918
#>  [785] 0.8424910 0.8424901 0.8424892 0.8424884 0.8424875 0.8424867 0.8424858
#>  [792] 0.8424849 0.8424841 0.8424832 0.8424824 0.8424815 0.8424806 0.8424798
#>  [799] 0.8424789 0.8424781 0.8424772 0.8424763 0.8424755 0.8424746 0.8424738
#>  [806] 0.8424729 0.8424721 0.8424712 0.8424703 0.8424695 0.8424686 0.8424678
#>  [813] 0.8424669 0.8424660 0.8424652 0.8424643 0.8424635 0.8424626 0.8424617
#>  [820] 0.8424609 0.8424600 0.8424592 0.8424583 0.8424574 0.8424566 0.8424557
#>  [827] 0.8424549 0.8424540 0.8424532 0.8424523 0.8424514 0.8424506 0.8424497
#>  [834] 0.8424489 0.8424480 0.8424471 0.8424463 0.8424454 0.8424446 0.8424437
#>  [841] 0.8424428 0.8424420 0.8424411 0.8424403 0.8424394 0.8424385 0.8424377
#>  [848] 0.8424368 0.8424360 0.8424351 0.8424343 0.8424334 0.8424325 0.8424317
#>  [855] 0.8424308 0.8424300 0.8424291 0.8424282 0.8424274 0.8424265 0.8424257
#>  [862] 0.8424248 0.8424239 0.8424231 0.8424222 0.8424214 0.8424205 0.8424196
#>  [869] 0.8424188 0.8424179 0.8424171 0.8424162 0.8424153 0.8424145 0.8424136
#>  [876] 0.8424128 0.8424119 0.8424111 0.8424102 0.8424093 0.8424085 0.8424076
#>  [883] 0.8424068 0.8424059 0.8424050 0.8424042 0.8424033 0.8424025 0.8424016
#>  [890] 0.8424007 0.8423999 0.8423990 0.8423982 0.8423973 0.8423964 0.8423956
#>  [897] 0.8423947 0.8423939 0.8423930 0.8423922 0.8423913 0.8423904 0.8423896
#>  [904] 0.8423887 0.8423879 0.8423870 0.8423861 0.8423853 0.8423844 0.8423836
#>  [911] 0.8423827 0.8423818 0.8423810 0.8423801 0.8423793 0.8423784 0.8423775
#>  [918] 0.8423767 0.8423758 0.8423750 0.8423741 0.8423733 0.8423724 0.8423715
#>  [925] 0.8423707 0.8423698 0.8423690 0.8423681 0.8423672 0.8423664 0.8423655
#>  [932] 0.8423647 0.8423638 0.8423629 0.8423621 0.8423612 0.8423604 0.8423595
#>  [939] 0.8423586 0.8423578 0.8423569 0.8423561 0.8423552 0.8423543 0.8423535
#>  [946] 0.8423526 0.8423518 0.8423509 0.8423501 0.8423492 0.8423483 0.8423475
#>  [953] 0.8423466 0.8423458 0.8423449 0.8423440 0.8423432 0.8423423 0.8423415
#>  [960] 0.8423406 0.8423397 0.8423389 0.8423380 0.8423372 0.8423363 0.8423354
#>  [967] 0.8423346 0.8423337 0.8423329 0.8423320 0.8423312 0.8423303 0.8423294
#>  [974] 0.8423286 0.8423277 0.8423269 0.8423260 0.8423251 0.8423243 0.8423234
#>  [981] 0.8423226 0.8423217 0.8423208 0.8423200 0.8423191 0.8423183 0.8423174
#>  [988] 0.8423165 0.8423157 0.8423148 0.8423140 0.8423131 0.8423123 0.8423114
#>  [995] 0.8423105 0.8423097 0.8423088 0.8423080 0.8423071 0.8423062 0.8423054
#> [1002] 0.8423045 0.8423037 0.8423028 0.8423019 0.8423011 0.8423002 0.8422994
#> [1009] 0.8422985 0.8422976 0.8422968 0.8422959 0.8422951 0.8422942 0.8422934
#> [1016] 0.8422925 0.8422916 0.8422908 0.8422899 0.8422891 0.8422882 0.8422873
#> [1023] 0.8422865 0.8422856 0.8422848 0.8422839 0.8422830 0.8422822 0.8422813
#> [1030] 0.8422805 0.8422796 0.8422787 0.8422779 0.8422770 0.8422762 0.8422753
#> [1037] 0.8422745 0.8422736 0.8422727 0.8422719 0.8422710 0.8422702 0.8422693
#> [1044] 0.8422684 0.8422676 0.8422667 0.8422659 0.8422650 0.8422641 0.8422633
#> [1051] 0.8422624 0.8422616 0.8422607 0.8422598 0.8422590 0.8422581 0.8422573
#> [1058] 0.8422564 0.8422555 0.8422547 0.8422538 0.8422530 0.8422521 0.8422513
#> [1065] 0.8422504 0.8422495 0.8422487 0.8422478 0.8422470 0.8422461 0.8422452
#> [1072] 0.8422444 0.8422435 0.8422427 0.8422418 0.8422409 0.8422401 0.8422392
#> [1079] 0.8422384 0.8422375 0.8422366 0.8422358 0.8422349 0.8422341 0.8422332
#> [1086] 0.8422324 0.8422315 0.8422306 0.8422298 0.8422289 0.8422281 0.8422272
#> [1093] 0.8422263 0.8422255 0.8422246 0.8422238 0.8422229 0.8422220 0.8422212
#> [1100] 0.8422203 0.8422195 0.8422186 0.8422177 0.8422169 0.8422160 0.8422152
#> [1107] 0.8422143 0.8422135 0.8422126 0.8422117 0.8422109 0.8422100 0.8422092
#> [1114] 0.8422083 0.8422074 0.8422066 0.8422057 0.8422049 0.8422040 0.8422031
#> [1121] 0.8422023 0.8422014 0.8422006 0.8421997 0.8421988 0.8421980 0.8421971
#> [1128] 0.8421963 0.8421954 0.8421946 0.8421937 0.8421928 0.8421920 0.8421911
#> [1135] 0.8421903 0.8421894 0.8421885 0.8421877 0.8421868 0.8421860 0.8421851
#> [1142] 0.8421842 0.8421834 0.8421825 0.8421817 0.8421808 0.8421799 0.8421791
#> [1149] 0.8421782 0.8421774 0.8421765 0.8421757 0.8421748 0.8421739 0.8421731
#> [1156] 0.8421722 0.8421714 0.8421705 0.8421696 0.8421688 0.8421679 0.8421671
#> [1163] 0.8421662 0.8421653 0.8421645 0.8421636 0.8421628 0.8421619 0.8421610
#> [1170] 0.8421602 0.8421593 0.8421585 0.8421576 0.8421568 0.8421559 0.8421550
#> [1177] 0.8421542 0.8421533 0.8421525 0.8421516 0.8421507 0.8421499 0.8421490
#> [1184] 0.8421482 0.8421473 0.8421464 0.8421456 0.8421447 0.8421439 0.8421430
#> [1191] 0.8421421 0.8421413 0.8421404 0.8421396 0.8421387 0.8421378 0.8421370
#> [1198] 0.8421361 0.8421353 0.8421344 0.8421336 0.8421327 0.8421318 0.8421310
#> [1205] 0.8421301 0.8421293 0.8421284 0.8421275 0.8421267 0.8421258 0.8421250
#> [1212] 0.8421241 0.8421232 0.8421224 0.8421215 0.8421207 0.8421198 0.8421189
#> [1219] 0.8421181 0.8421172 0.8421164 0.8421155 0.8421147 0.8421138 0.8421129
#> [1226] 0.8421121 0.8421112 0.8421104 0.8421095 0.8421086 0.8421078 0.8421069
#> [1233] 0.8421061 0.8421052 0.8421043 0.8421035 0.8421026 0.8421018 0.8421009
#> [1240] 0.8421000 0.8420992 0.8420983 0.8420975 0.8420966 0.8420958 0.8420949
#> [1247] 0.8420940 0.8420932 0.8420923 0.8420915 0.8420906 0.8420897 0.8420889
#> [1254] 0.8420880 0.8420872 0.8420863 0.8420854 0.8420846 0.8420837 0.8420829
#> [1261] 0.8420820 0.8420811 0.8420803 0.8420794 0.8420786 0.8420777 0.8420769
#> [1268] 0.8420760 0.8420751 0.8420743 0.8420734 0.8420726 0.8420717 0.8420708
#> [1275] 0.8420700 0.8420691 0.8420683 0.8420674 0.8420665 0.8420657 0.8420648
#> [1282] 0.8420640 0.8420631 0.8420622 0.8420614 0.8420605 0.8420597 0.8420588
#> [1289] 0.8420580 0.8420571 0.8420562 0.8420554 0.8420545 0.8420537 0.8420528
#> [1296] 0.8420519 0.8420511 0.8420502 0.8420494 0.8420485 0.8420476 0.8420468
#> [1303] 0.8420459 0.8420451 0.8420442 0.8420433 0.8420425 0.8420416 0.8420408
#> [1310] 0.8420399 0.8420391 0.8420382 0.8420373 0.8420365 0.8420356 0.8420348
#> [1317] 0.8420339 0.8420330 0.8420322 0.8420313 0.8420305 0.8420296 0.8420287
#> [1324] 0.8420279 0.8420270 0.8420262 0.8420253 0.8420244 0.8420236 0.8420227
#> [1331] 0.8420219 0.8420210 0.8420202 0.8420193 0.8420184 0.8420176 0.8420167
#> [1338] 0.8420159 0.8420150 0.8420141 0.8420133 0.8420124 0.8420116 0.8420107
#> [1345] 0.8420098 0.8420090 0.8420081 0.8420073 0.8420064 0.8420055 0.8420047
#> [1352] 0.8420038 0.8420030 0.8420021 0.8420013 0.8420004 0.8419995 0.8419987
#> [1359] 0.8419978 0.8419970 0.8419961 0.8419952 0.8419944 0.8419935 0.8419927
#> [1366] 0.8419918 0.8419909 0.8419901 0.8419892 0.8419884 0.8419875 0.8419866
#> [1373] 0.8419858 0.8419849 0.8419841 0.8419832 0.8419824 0.8419815 0.8419806
#> [1380] 0.8419798 0.8419789 0.8419781 0.8419772 0.8419763 0.8419755 0.8419746
#> [1387] 0.8419738 0.8419729 0.8419720 0.8419712 0.8419703 0.8419695 0.8419686
#> [1394] 0.8419677 0.8419669 0.8419660 0.8419652 0.8419643 0.8419635 0.8419626
#> [1401] 0.8419617 0.8419609 0.8419600 0.8419592 0.8419583 0.8419574 0.8419566
#> [1408] 0.8419557 0.8419549 0.8419540 0.8419531 0.8419523 0.8419514 0.8419506
#> [1415] 0.8419497 0.8419488 0.8419480 0.8419471 0.8419463 0.8419454 0.8419446
#> [1422] 0.8419437 0.8419428 0.8419420 0.8419411 0.8419403 0.8419394 0.8419385
#> [1429] 0.8419377 0.8419368 0.8419360 0.8419351 0.8419342 0.8419334 0.8419325
#> [1436] 0.8419317 0.8419308 0.8419299 0.8419291 0.8419282 0.8419274 0.8419265
#> [1443] 0.8419257 0.8419248 0.8419239 0.8419231 0.8419222 0.8419214 0.8419205
#> [1450] 0.8419196 0.8419188 0.8419179 0.8419171 0.8419162 0.8419153 0.8419145
#> [1457] 0.8419136 0.8419128 0.8419119 0.8419110 0.8419102 0.8419093 0.8419085
#> [1464] 0.8419076 0.8419068 0.8419059 0.8419050 0.8419042 0.8419033 0.8419025
#> [1471] 0.8419016 0.8419007 0.8418999 0.8418990 0.8418982 0.8418973 0.8418964
#> [1478] 0.8418956 0.8418947 0.8418939 0.8418930 0.8418921 0.8418913 0.8418904
#> [1485] 0.8418896 0.8418887 0.8418879 0.8418870 0.8418861 0.8418853 0.8418844
#> [1492] 0.8418836 0.8418827 0.8418818 0.8418810 0.8418801 0.8418793 0.8418784
#> [1499] 0.8418775 0.8418767 0.8418758 0.8418750 0.8418741 0.8418732 0.8418724
#> [1506] 0.8418715 0.8418707 0.8418698 0.8418690 0.8418681 0.8418672 0.8418664
#> [1513] 0.8418655 0.8418647 0.8418638 0.8418629 0.8418621 0.8418612 0.8418604
#> [1520] 0.8418595 0.8418586 0.8418578 0.8418569 0.8418561 0.8418552 0.8418543
#> [1527] 0.8418535 0.8418526 0.8418518 0.8418509 0.8418501 0.8418492 0.8418483
#> [1534] 0.8418475 0.8418466 0.8418458 0.8418449 0.8418440 0.8418432 0.8418423
#> [1541] 0.8418415 0.8418406 0.8418397 0.8418389 0.8418380 0.8418372 0.8418363
#> [1548] 0.8418354 0.8418346 0.8418337 0.8418329 0.8418320 0.8418312 0.8418303
#> [1555] 0.8418294 0.8418286 0.8418277 0.8418269 0.8418260 0.8418251 0.8418243
#> [1562] 0.8418234 0.8418226 0.8418217 0.8418208 0.8418200 0.8418191 0.8418183
#> [1569] 0.8418174 0.8418165 0.8418157 0.8418148 0.8418140 0.8418131 0.8418123
#> [1576] 0.8418114 0.8418105 0.8418097 0.8418088 0.8418080 0.8418071 0.8418062
#> [1583] 0.8418054 0.8418045 0.8418037 0.8418028 0.8418019 0.8418011 0.8418002
#> [1590] 0.8417994 0.8417985 0.8417976 0.8417968 0.8417959 0.8417951 0.8417942
#> [1597] 0.8417934 0.8417925 0.8417916 0.8417908 0.8417899 0.8417891 0.8417882
#> [1604] 0.8417873 0.8417865 0.8417856 0.8417848 0.8417839 0.8417830 0.8417822
#> [1611] 0.8417813 0.8417805 0.8417796 0.8417787 0.8417779 0.8417770 0.8417762
#> [1618] 0.8417753 0.8417745 0.8417736 0.8417727 0.8417719 0.8417710 0.8417702
#> [1625] 0.8417693 0.8417684 0.8417676 0.8417667 0.8417659 0.8417650 0.8417641
#> [1632] 0.8417633 0.8417624 0.8417616 0.8417607 0.8417598 0.8417590 0.8417581
#> [1639] 0.8417573 0.8417564 0.8417556 0.8417547 0.8417538 0.8417530 0.8417521
#> [1646] 0.8417513 0.8417504 0.8417495 0.8417487 0.8417478 0.8417470 0.8417461
#> [1653] 0.8417452 0.8417444 0.8417435 0.8417427 0.8417418 0.8417409 0.8417401
#> [1660] 0.8417392 0.8417384 0.8417375 0.8417367 0.8417358 0.8417349 0.8417341
#> [1667] 0.8417332 0.8417324 0.8417315 0.8417306 0.8417298 0.8417289 0.8417281
#> [1674] 0.8417272 0.8417263 0.8417255 0.8417246 0.8417238 0.8417229 0.8417220
#> [1681] 0.8417212 0.8417203 0.8417195 0.8417186 0.8417178 0.8417169 0.8417160
#> [1688] 0.8417152 0.8417143 0.8417135 0.8417126 0.8417117 0.8417109 0.8417100
#> [1695] 0.8417092 0.8417083 0.8417074 0.8417066 0.8417057 0.8417049 0.8417040
#> [1702] 0.8417031 0.8417023 0.8417014 0.8417006 0.8416997 0.8416989 0.8416980
#> [1709] 0.8416971 0.8416963 0.8416954 0.8416946 0.8416937 0.8416928 0.8416920
#> [1716] 0.8416911 0.8416903 0.8416894 0.8416885 0.8416877 0.8416868 0.8416860
#> [1723] 0.8416851 0.8416842 0.8416834 0.8416825 0.8416817 0.8416808 0.8416800
#> [1730] 0.8416791 0.8416782 0.8416774 0.8416765 0.8416757 0.8416748 0.8416739
#> [1737] 0.8416731 0.8416722 0.8416714 0.8416705 0.8416696 0.8416688 0.8416679
#> [1744] 0.8416671 0.8416662 0.8416653 0.8416645 0.8416636 0.8416628 0.8416619
#> [1751] 0.8416611 0.8416602 0.8416593 0.8416585 0.8416576 0.8416568 0.8416559
#> [1758] 0.8416550 0.8416542 0.8416533 0.8416525 0.8416516 0.8416507 0.8416499
#> [1765] 0.8416490 0.8416482 0.8416473 0.8416465 0.8416456 0.8416447 0.8416439
#> [1772] 0.8416430 0.8416422 0.8416413 0.8416404 0.8416396 0.8416387 0.8416379
#> [1779] 0.8416370 0.8416361 0.8416353 0.8416344 0.8416336 0.8416327 0.8416318
#> [1786] 0.8416310 0.8416301 0.8416293 0.8416284 0.8416276 0.8416267 0.8416258
#> [1793] 0.8416250 0.8416241 0.8416233 0.8416224 0.8416215 0.8416207 0.8416198
#> [1800] 0.8416190 0.8416181 0.8416172 0.8416164 0.8416155 0.8416147 0.8416138
#> [1807] 0.8416129 0.8416121 0.8416112 0.8416104 0.8416095 0.8416087 0.8416078
#> [1814] 0.8416069 0.8416061 0.8416052 0.8416044 0.8416035 0.8416026 0.8416018
#> [1821] 0.8416009 0.8416001 0.8415992 0.8415983 0.8415975 0.8415966 0.8415958
#> [1828] 0.8415949 0.8415940 0.8415932 0.8415923 0.8415915 0.8415906 0.8415898
#> [1835] 0.8415889 0.8415880 0.8415872 0.8415863 0.8415855 0.8415846 0.8415837
#> [1842] 0.8415829 0.8415820 0.8415812 0.8415803 0.8415794 0.8415786 0.8415777
#> [1849] 0.8415769 0.8415760 0.8415751 0.8415743 0.8415734 0.8415726 0.8415717
#> [1856] 0.8415709 0.8415700 0.8415691 0.8415683 0.8415674 0.8415666 0.8415657
#> [1863] 0.8415648 0.8415640 0.8415631 0.8415623 0.8415614 0.8415605 0.8415597
#> [1870] 0.8415588 0.8415580 0.8415571 0.8415562 0.8415554 0.8415545 0.8415537
#> [1877] 0.8415528 0.8415520 0.8415511 0.8415502 0.8415494 0.8415485 0.8415477
#> [1884] 0.8415468 0.8415459 0.8415451 0.8415442 0.8415434 0.8415425 0.8415416
#> [1891] 0.8415408 0.8415399 0.8415391 0.8415382 0.8415374 0.8415365 0.8415356
#> [1898] 0.8415348 0.8415339 0.8415331 0.8415322 0.8415313 0.8415305 0.8415296
#> [1905] 0.8415288 0.8415279 0.8415270 0.8415262 0.8415253 0.8415245 0.8415236
#> [1912] 0.8415227 0.8415219 0.8415210 0.8415202 0.8415193 0.8415185 0.8415176
#> [1919] 0.8415167 0.8415159 0.8415150 0.8415142 0.8415133 0.8415124 0.8415116
#> [1926] 0.8415107 0.8415099 0.8415090 0.8415081 0.8415073 0.8415064 0.8415056
#> [1933] 0.8415047 0.8415038 0.8415030 0.8415021 0.8415013 0.8415004 0.8414996
#> [1940] 0.8414987 0.8414978 0.8414970 0.8414961 0.8414953 0.8414944 0.8414935
#> [1947] 0.8414927 0.8414918 0.8414910 0.8414901 0.8414892 0.8414884 0.8414875
#> [1954] 0.8414867 0.8414858 0.8414849 0.8414841 0.8414832 0.8414824 0.8414815
#> [1961] 0.8414807 0.8414798 0.8414789 0.8414781 0.8414772 0.8414764 0.8414755
#> [1968] 0.8414746 0.8414738 0.8414729 0.8414721 0.8414712 0.8414703 0.8414695
#> [1975] 0.8414686 0.8414678 0.8414669 0.8414660 0.8414652 0.8414643 0.8414635
#> [1982] 0.8414626 0.8414618 0.8414609 0.8414600 0.8414592 0.8414583 0.8414575
#> [1989] 0.8414566 0.8414557 0.8414549 0.8414540 0.8414532 0.8414523 0.8414514
#> [1996] 0.8414506 0.8414497 0.8414489 0.8414480 0.8414472 0.8414463