Computes Q_3 from modified Corrado-Su model.

Q_3 formula implementation from modified Corrado-Su model.

Usage

csmq3(spot, sigma, time, d1, w)

Arguments

spot

current stock price.

sigma

volatility of the stock price.

time

time to option expiration in years.

d1

d value from modified Corrado-Su model.

w

constant from modified Corrado-Su model.

Value

Vector with Q_3.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

See also

Examples

spot <- 25
sigma <- 0.2
time <- 0.5
strike <- 26
rate <- 0.12
yield <- 0
skewness <- 0
kurtosis <- 3
w <- csmw(sigma, time, skewness, kurtosis)
d <- csmd(spot, strike, time, rate, yield, sigma, w)
csmq3(spot, sigma, time, d, w)
#> [1] 0.01505004
csmq3(200:320 / 10, sigma, time, d, w)
#>   [1] 0.01204003 0.01210023 0.01216043 0.01222063 0.01228083 0.01234103
#>   [7] 0.01240123 0.01246143 0.01252163 0.01258183 0.01264204 0.01270224
#>  [13] 0.01276244 0.01282264 0.01288284 0.01294304 0.01300324 0.01306344
#>  [19] 0.01312364 0.01318384 0.01324404 0.01330424 0.01336444 0.01342464
#>  [25] 0.01348484 0.01354504 0.01360524 0.01366544 0.01372564 0.01378584
#>  [31] 0.01384604 0.01390624 0.01396644 0.01402664 0.01408684 0.01414704
#>  [37] 0.01420724 0.01426744 0.01432764 0.01438784 0.01444804 0.01450824
#>  [43] 0.01456844 0.01462864 0.01468884 0.01474904 0.01480924 0.01486944
#>  [49] 0.01492964 0.01498984 0.01505004 0.01511024 0.01517044 0.01523064
#>  [55] 0.01529084 0.01535104 0.01541124 0.01547144 0.01553164 0.01559184
#>  [61] 0.01565204 0.01571224 0.01577244 0.01583264 0.01589284 0.01595304
#>  [67] 0.01601324 0.01607344 0.01613364 0.01619384 0.01625405 0.01631425
#>  [73] 0.01637445 0.01643465 0.01649485 0.01655505 0.01661525 0.01667545
#>  [79] 0.01673565 0.01679585 0.01685605 0.01691625 0.01697645 0.01703665
#>  [85] 0.01709685 0.01715705 0.01721725 0.01727745 0.01733765 0.01739785
#>  [91] 0.01745805 0.01751825 0.01757845 0.01763865 0.01769885 0.01775905
#>  [97] 0.01781925 0.01787945 0.01793965 0.01799985 0.01806005 0.01812025
#> [103] 0.01818045 0.01824065 0.01830085 0.01836105 0.01842125 0.01848145
#> [109] 0.01854165 0.01860185 0.01866205 0.01872225 0.01878245 0.01884265
#> [115] 0.01890285 0.01896305 0.01902325 0.01908345 0.01914365 0.01920385
#> [121] 0.01926405