Computes Q_4 from modified Corrado-Su model.

Q_4 formula implementation from modified Corrado-Su model.

Usage

csmq4(spot, sigma, time, d1, w)

Arguments

spot

current stock price.

sigma

volatility of the stock price.

time

time to option expiration in years.

d1

d value from modified Corrado-Su model.

w

constant from modified Corrado-Su model.

Value

Vector with Q_4.

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

See also

Examples

spot <- 25
sigma <- 0.2
time <- 0.5
strike <- 26
rate <- 0.12
yield <- 0
skewness <- 0
kurtosis <- 3
w <- csmw(sigma, time, skewness, kurtosis)
d <- csmd(spot, strike, time, rate, yield, sigma, w)
csmq4(spot, sigma, time, d, w)
#> [1] -0.05653327
csmq4(200:320 / 10, sigma, time, d, w)
#>   [1] -0.04522662 -0.04545275 -0.04567888 -0.04590502 -0.04613115 -0.04635728
#>   [7] -0.04658342 -0.04680955 -0.04703568 -0.04726182 -0.04748795 -0.04771408
#>  [13] -0.04794022 -0.04816635 -0.04839248 -0.04861862 -0.04884475 -0.04907088
#>  [19] -0.04929701 -0.04952315 -0.04974928 -0.04997541 -0.05020155 -0.05042768
#>  [25] -0.05065381 -0.05087995 -0.05110608 -0.05133221 -0.05155835 -0.05178448
#>  [31] -0.05201061 -0.05223674 -0.05246288 -0.05268901 -0.05291514 -0.05314128
#>  [37] -0.05336741 -0.05359354 -0.05381968 -0.05404581 -0.05427194 -0.05449808
#>  [43] -0.05472421 -0.05495034 -0.05517647 -0.05540261 -0.05562874 -0.05585487
#>  [49] -0.05608101 -0.05630714 -0.05653327 -0.05675941 -0.05698554 -0.05721167
#>  [55] -0.05743781 -0.05766394 -0.05789007 -0.05811621 -0.05834234 -0.05856847
#>  [61] -0.05879460 -0.05902074 -0.05924687 -0.05947300 -0.05969914 -0.05992527
#>  [67] -0.06015140 -0.06037754 -0.06060367 -0.06082980 -0.06105594 -0.06128207
#>  [73] -0.06150820 -0.06173433 -0.06196047 -0.06218660 -0.06241273 -0.06263887
#>  [79] -0.06286500 -0.06309113 -0.06331727 -0.06354340 -0.06376953 -0.06399567
#>  [85] -0.06422180 -0.06444793 -0.06467406 -0.06490020 -0.06512633 -0.06535246
#>  [91] -0.06557860 -0.06580473 -0.06603086 -0.06625700 -0.06648313 -0.06670926
#>  [97] -0.06693540 -0.06716153 -0.06738766 -0.06761379 -0.06783993 -0.06806606
#> [103] -0.06829219 -0.06851833 -0.06874446 -0.06897059 -0.06919673 -0.06942286
#> [109] -0.06964899 -0.06987513 -0.07010126 -0.07032739 -0.07055353 -0.07077966
#> [115] -0.07100579 -0.07123192 -0.07145806 -0.07168419 -0.07191032 -0.07213646
#> [121] -0.07236259