SVI Variance

Calculates the variance as given by the SVI model for the provided parameters.

Usage

svi_var(a, b, m, rho, x, sigma)

Arguments

a

'a' in the SVI model (instantaneous variance)

b

'b' in the SVI model (speed of variance mean reversion)

m

'm' in the SVI model (long term expected variance)

rho

'rho' in the SVI model (correlation between variance and price)

x

'x' in the SVI model, given by ln(K/F) where K is an options strike price and F its underlying's future

sigma

'sigma' in the SVI model (volatility of variance)

Value

Volatility value as given by the SVI model

Recycle rule

These arguments handle the recycle rule so vectors can be provided and once those vectors differs in length the recycle rule is applied.

See also